### 2025-02-25 #7851 @editorialbot
[PRE REVIEW]: SOAP: A Python package for calculating the properties of galaxies and halos formed in cosmological simulations
**Submitting author:** @robjmcgibbon (<a href="http://orcid.org/0000-0003-0651-0776">Robert McGibbon</a>)
**Repository:** https://github.com/SWIFTSIM/SOAP
**Branch with paper.md** (empty if default branch): soap_runtime
**Version:** v1.0
