### 2025-02-26 8898a71
Minor paper compilation updates
