### 2025-02-26 5967772
Stop ranks hanging when reading binding energy
