### 2025-02-26 ef8cbe8
Read SnapshotIndexOfLastIsolation
