### 2025-02-27 #7851 @warrickball
Hi @robjmcgibbon, and thanks for the submission. All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!

In the meantime, I first wanted to check that I've understood the code's history correctly, it looks like it was originally developed in co-author John C. Helly's project namespace before being in effect "moved" into the SWIFTSIM org namespace. This is fine but I just want to check I've got that right.

For the sake of the review, when the time comes, I want to draw your attention to our review criteria for ["Installation instructions"](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions) (emphasis mine):

> Software dependencies should be clearly documented and their installation handled by an automated procedure. Where possible, software installation should be managed with a package manager. However, this criterion depends upon the maturity and availability of software packaging and distribution in the programming language being used. For example, **Python packages should be pip installable, and have adopted [packaging conventions](https://packaging.python.org/)**, while Fortran submissions with a Makefile may be sufficient.

I recommend separating the code into the two scripts that I think serve as entry points (`group_membership.py` and `compute_halo_properties.py`) and the remaining code, which is mostly the library, though I think there are some tests embedded in the scripts that ought to be separated into a `tests` folder or similar and ideally be run as continuous integration steps (i.e., GitHub actions).

The [official Python packaging guide](https://packaging.python.org/) is a good place to start, supplemented by specific documentation for some of the tools that can be used. Note that it is **not** necessary to register a project on the Python Package Index (PyPI) and upload it there, as long as you can `pip install` from the git repository with e.g.
```
pip install git+https://github.com/SWIFTSIM/SOAP.git
```
This will work as long as `pip` finds a package spec (e.g. a `pyproject.toml` file).

Note also that the documentation, though perhaps passable, will be much more favourably reviewed if you could render it online, rather than requiring the local build step. I presume there's some in-house expertise available to you because the GitHub organisation presents the main SWIFT docs at https://swift.strw.leidenuniv.nl/docs. It would be great if you could render similar Sphinx docs (or other docs system, like MkDocs) perhaps at https://swift.strw.leidenuniv.nl/soap?
