+++ pyproject.toml
@@ -36,7 +36,15 @@ test = [
     "pytest-mpi",
 ]
 
-[tool.setuptools.packages.find]
+[tool.setuptools]
+packages = [
+    "SOAP",
+    "SOAP.catalogue_readers",
+    "SOAP.core",
+    "SOAP.particle_filter",
+    "SOAP.particle_selection",
+    "SOAP.property_calculation",
+]
 
 [project.urls]
 "Homepage" = "https://github.com/SWIFTSIM/SOAP"
