### 2025-03-02 e02d672
Find packages automatically
