### 2025-03-03 476545a
Add small DMO example
