### 2025-03-03 6b70e12
Test subhalo rank with HBT data
