### 2025-03-04 581cc02
Make test helper script MPI-compatible
