### 2025-03-04 35e9f0b
Generate documentation for small test
