### 2025-03-04 6625023
Move Rockstar test and read with numpy instead of pandas
