+++ requirements.txt
@@ -5,7 +5,6 @@ unyt>=3
 astropy
 numba
 scipy
-pandas
 pytest-mpi
 matplotlib
 psutil
