### 2025-03-05 c2a1b97
Address pre-review comments (#133)
* Use pyproject.toml instead of setup.py

* Update packages

* List packages

* Find packages automatically

* Add small DMO example

* Test subhalo rank with HBT data

* Add scripts for running HBT/VR

* Make test helper script MPI-compatible

* Generate documentation for small test

* Move Rockstar test and read with numpy instead of pandas

* Add rockstar test

* Run formatter

* Move subfind test
