### 2025-03-05 #7851 @robjmcgibbon
Hi @warrickball thank you for the feedback! I've made a couple of commits to the branch to address your comments.

You are correct about the code's history. John started development of the code and it was then forked to the SWIFTSIM namespace (where we keep most of our code). John has made commits to the new repo since then, he is still involved in the code. Unfortunately I don't know if github allows you to disconnect a forked repo without losing lots of commit info.

I have verified that it is possible to install the code with `pip install git+https://github.com/SWIFTSIM/SOAP.git@soap_runtime` , and added a note on this installation method to the README. The problem is that we require parallel h5py, and pip tends to default to the serial version.

All tests are now in the `tests` directory. I've also added an example script that will download a small simulation volume and run SOAP on it.

I will have a think about the best way to host the documentation.
