### 2025-03-06 73fd19d
Don't use hydrogen fraction for dust properties
