+++ documentation/footnote_dust.tex
@@ -20,4 +20,4 @@ The mass of dust in molecular gas is defined as
     \sum_i m_d \frac{m_{\rm{H}_2,i}}{m_{\rm{H}, i}}
 \end{equation}
 where $m_d$ is the dust mass of particle $i$, $m_{\rm{H}_2,i}$ is the molecular hydrogen mass of the particle,
-and $m_{\rm{H},i}$ is the total hydrogen mass of the particle. A similar expression is just for the mass of dust in atomic gas.
+and $m_{\rm{H},i}$ is the total hydrogen mass of the particle. A similar expression is used for the mass of dust in atomic gas.
