### 2025-03-13 52deb7d
Read in potential energy instead of binding
