+++ SOAP/particle_selection/SO_properties.py
@@ -2082,9 +2082,9 @@ class SOParticleData:
         """
         if self.Ngas == 0:
             return None
-        ekin_gas = self.gas_masses * ((self.gas_vel - self.vcom_gas[None, :]) ** 2).sum(
-            axis=1
-        )
+        v_gas = self.gas_vel - self.vcom_gas[None, :]
+        v_gas += self.gas_pos * self.cosmology["H"]
+        ekin_gas = self.gas_masses * (v_gas**2).sum(axis=1)
         return 0.5 * ekin_gas.sum()
 
     @lazy_property
@@ -2254,9 +2254,10 @@ class SOParticleData:
         """
         if self.Nstar == 0:
             return None
-        ekin_star = self.star_masses * (
-            (self.star_vel - self.vcom_star[None, :]) ** 2
-        ).sum(axis=1)
+
+        v_star = self.star_vel - self.vcom_star[None, :]
+        v_star += self.star_pos * self.cosmology["H"]
+        ekin_star = self.star_masses * (v_star**2).sum(axis=1)
         return 0.5 * ekin_star.sum()
 
     @lazy_property
