### 2025-03-14 afa5495
Use vCoM from all particles
