### 2025-03-17 a48ff63
Add shorthand to obtain grik luminosities.
