### 2025-03-17 e4d49eb
Add basic functions to measure disc to total luminosity in different bands.
