### 2025-03-17 ec461e8
Make new properties not be computed by default
