+++ SOAP/particle_selection/aperture_properties.py
@@ -3207,6 +3207,10 @@ class ApertureProperties(HaloProperty):
         "HalfMassRadiusBaryon": False,
         "DtoTgas": False,
         "DtoTstar": False,
+        "DtoTstar_gBand": False,
+        "DtoTstar_rBand": False,
+        "DtoTstar_iBand": False,
+        "DtoTstar_kBand": False,
         "starOfrac": False,
         "starFefrac": False,
         "stellar_age_mw": False,
