### 2025-03-17 1f7e223
Add call of unimplemented function that computes light weighted L quantities.
