### 2025-03-17 cced970
Define function for luminosity weighted L quantities. Not implemented for now.
