### 2025-03-17 078232b
Add docs on parameters for get_angular_momentum_and_kappa_corot_luminosity_weighted
