### 2025-03-17 01214a0
Make kappa_corot and counterotating mass give a value per GAMMA band
