### 2025-03-17 ecca593
Add counterrotating luminosity array
