### 2025-03-17 5b8d483
Handle multiple luminosity bands when computing average angular momentum
