### 2025-03-17 bfae566
Compute projected angular momentum if we do_counterrot_luminosity
