### 2025-03-17 3226029
We now compute if stars are counterrotating or not for each luminosity band. Expanded description of unyt array shapes to keep track in each step.
