+++ SOAP/property_calculation/kinematic_properties.py
@@ -281,7 +281,7 @@ def get_angular_momentum_and_kappa_corot_luminosity_weighted(
         )
 
     if do_counterrot_luminosity:
-        M_counterrot = unyt.unyt_array(
+        L_counterrot = unyt.unyt_array(
             np.zeros(luminosities.shape[-1]), dtype=np.float32, units=luminosities.units, registry=luminosities.units.registry
         )
 
