### 2025-03-18 38eb62e
Add correct return and remove NotImplemented error
