### 2025-03-18 9de6b88
Change variables assigned to subhaloes to prevent overwritting mass-weighted calculations
