### 2025-03-18 6594d04
Only return a single property, which includes all GAMMA bands.
