### 2025-03-18 5c54917
Return a mass and light-based D/T.
