### 2025-03-18 5f2ad3a
Add kappa_corot for luminosity bands
