### 2025-03-18 b57ff58
Add property to save all luminosity-weighted angular momentum
