### 2025-03-18 448ac82
Add property definition in the Aperture properties
