### 2025-03-18 8767703
Added new properties to the property table
