### 2025-03-18 e2a8162
Add properties to the subhalo properties parameter file of COLIBRE
