### 2025-03-18 7de6981
Remove lazy properties that were not needed anymore
