### 2025-03-18 16ca2a6
Add luminosity-weighted angular momenta to bound subhalo properties
