### 2025-03-18 1c60e81
Add properties to the ApertureProperties parameter file section
