### 2025-03-18 d808a26
Change condition to execute new properties. Mainly because of an annoying inconsistent between how starless subhaloes are handled. Mstar = 0 but Luminosities = None.
