+++ SOAP/particle_selection/aperture_properties.py
@@ -1317,7 +1317,7 @@ class ApertureParticleData:
         Mcountrot_star_luminosity_weighted and Lcountrot_star_luminosity_weighted
         by compute_Lstar_luminosity_weighted_props().
         """
-        if np.all(self.StellarLuminosity == 0):
+        if self.Nstar == 0:
             return None
         if not hasattr(self, "internal_Lstar_luminosity_weighted"):
             self.compute_Lstar_luminosity_weighted_props()
@@ -1347,7 +1347,7 @@ class ApertureParticleData:
         Mcountrot_star_luminosity_weighted and Lcountrot_star_luminosity_weighted
         by compute_Lstar_luminosity_weighted_props().
         """
-        if np.all(self.StellarLuminosity == 0):
+        if self.Nstar == 0:
             return None
         if not hasattr(self, "internal_kappa_star_luminosity_weighted"):
             self.compute_Lstar_luminosity_weighted_props()
@@ -1379,7 +1379,7 @@ class ApertureParticleData:
         Mcountrot_star_luminosity_weighted and Lcountrot_star_luminosity_weighted
         by compute_Lstar_luminosity_weighted_props().
         """
-        if np.all(self.StellarLuminosity == 0):
+        if self.Nstar == 0:
             return None
         if not hasattr(self, "internal_Lcountrot_star_luminosity_weighted"):
             self.compute_Lstar_luminosity_weighted_props()
@@ -1398,7 +1398,7 @@ class ApertureParticleData:
         Mcountrot_star_luminosity_weighted and Lcountrot_star_luminosity_weighted
         by compute_Lstar_luminosity_weighted_props().
         """
-        if np.all(self.StellarLuminosity == 0) | (self.Mstar == 0):
+        if self.Nstar == 0:
             return None
         if not hasattr(self, "internal_Mcountrot_star_luminosity_weighted"):
             self.compute_Lstar_luminosity_weighted_props()
