### 2025-03-18 25dcc5e
Fix some indicies and if condition
