### 2025-03-18 dcc55f1
Move subhalo_properties check for Ntot
