### 2025-03-19 1602384
Remove Ntot warning
