+++ SOAP/particle_selection/subhalo_properties.py
@@ -2189,7 +2189,9 @@ class SubhaloProperties(HaloProperty):
         Nexpected = input_halo["nr_bound_part"]
         if Ntot < Nexpected:
             # Try again with a larger search radius
-            print(f"Ntot = {Ntot}, Nexpected = {Nexpected}, search_radius = {search_radius}")
+            # print(
+            #     f"Ntot = {Ntot}, Nexpected = {Nexpected}, search_radius = {search_radius}"
+            # )
             raise SearchRadiusTooSmallError(
                 "Search radius does not contain expected number of particles!"
             )
