### 2025-03-20 5d7ea96
Fix bug in kappa_corot definition.
