### 2025-03-20 cf36666
Update parameter file README (#136)
* Update parameter file README

* Mention pdf doco
