+++ parameter_files/README.md
@@ -2,6 +2,8 @@
 
 The parameter files are a YAML dictionary which define the parameters and settings for running SOAP.
 This file describes the structure of a parameter file, including all possible fields which can be specified.
+This file does not detail what the differences are between the various aperture types, for that
+see the main pdf documenation.
 
 ### Parameters
 
@@ -33,7 +35,7 @@ If a dataset is present in both the snapshot and the extra input files, the valu
 
 Settings for the halo finding algorithm and output file locations.
 
-- **type**: The halo finder being used. Possible options are `HBTplus`, `VR`, `Subfind`, and `Rockstar`.
+- **type**: The subhalo finder being used. Possible options are `HBTplus`, `VR`, `Subfind`, and `Rockstar`.
 - **filename**: Template for input halo catalogue files. The format of this depends on the halo finder as they have different output structure. HBTplus example: `"{sim_dir}/{sim_name}/HBT/{snap_nr:03d}/SubSnap_{snap_nr:03d}"`
 - **fof_filename**: Template for FOF catalog files. Used for storing host FOF information for central subhalos. This is currently only supported for HBTplus
 
@@ -50,21 +52,35 @@ Settings for writing the output SOAP catalogues properties, and for handling tem
 - **filename**: Template for the halo properties file paths, e.g. `"{output_dir}/{sim_name}/SOAP_uncompressed/{halo_finder}/halo_properties_{snap_nr:04d}.hdf5"`
 - **chunk_dir**: Directory for temporary chunk output files. e.g. `"{scratch_dir}/{sim_name}/SOAP-tmp/{halo_finder}/"`
 
+### SubhaloProperties
+
+Define which properties to compute for each subhalo.
+
+- **properties**: A list of properties to compute. The key for each property should be either a boolean, a string, or a dictionary. Passing a boolean simply flags whether to enable or disable that calculation (for all subhalos). If the property should only be computed for subhalos that meet a certain criteria then the name of the filter should be passed as a string (see the Filters section below). Alternatively a dictionary can be passed with two entries: `snapshot` and `snipshot`. In this case the behaviour will be determined based on whether the `--snipshot` flag is passed when running SOAP.
+
+An example is as follows
+```
+SubhaloProperties:
+  properties:
+    TotalMass: true
+    StellarMass: general
+    GasMass:
+      snapshot: true
+      snipshot: false
+```
+
 ### ApertureProperties
 
 Define which fixed spherical apertures to compute, and what properties to compute within them.
 
-- **properties**: A list of properties to compute. Each property can be passed a boolean, which simply flags whether to enable or disable calculation. Alternatively a dictionary can be passed with two entries: `snapshot` and `snipshot`. In this case the behaviour will be determined based on whether the `--snipshot` flag is passed when running SOAP.
-- **variations**: A list of which apertures to compute. Each aperture must have a name, a boolean key `inclusive` to indicate whether to include unbound particles, and a key `radius_in_kpc` to indicate the **physical size** of the aperture. Each aperture can optionally be passed the `filter` key. If the filter key is set then the aperture will only be computed for subhalos that fulfill the filter criteria. If no `filter` key is passed then the aperture will be computed for all halos. For inclusive apertures the boolean flag `skip_gt_enclose_radius` can be set (defaults to False). If it is set then properties will not be calculated for any halos where all bound particles are within the aperture radius.
+- **properties**: The same format as used for SubhaloProperties.
[…]
