### 2025-03-20 c19eb8b
Add luminosity-weighted quantities to SO apertures
