### 2025-03-21 d6dd6ba
Individual angular momentum for particles is now correct for kinetic energy calculation
