### 2025-03-21 c6e2846
Add documentation of SpecificPotentialEnergies
