### 2025-03-21 e7b567e
Remove mass term form average angular momentum we keep it in the individual particle angular momentum to obtain the kinetic energy in rotation.
