### 2025-03-24 1a0f4f6
Create copy of inertia tensor to implement luminosity weighting
