### 2025-03-24 193297a
Update description and inputs of inertia tensor
