### 2025-03-24 ac58381
Mask luminosities if we mask other properties
