### 2025-03-24 69b7bff
Create correct shape for eigenval and eigenvec matricies
