### 2025-03-24 d80dcec
Shapes for each luminosity band
