### 2025-03-24 5cd9ea0
Add stopping criteria on a per band basis
