### 2025-03-24 48375a4
Define the luminosity-weighted inertia tensor for BoundSubhalos
