+++ SOAP/property_table.py
@@ -2074,6 +2074,54 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=0,
         ),
+        "StellarInertiaTensor_LuminosityWeighted": Property(
+            name="StellarInertiaTensor_LuminosityWeighted",
+            shape=54, # 6 times each GAMA band
+            dtype=np.float32,
+            unit="snap_length**2",
+            description="3D inertia tensor computed iteratively from the stellar luminosity distribution, relative to the halo centre and for each GAMA band. Diagonal components and one off-diagonal triangle as (1,1), (2,2), (3,3), (1,2), (1,3), (2,3). Only calculated when we have more than 20 particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType4/Coordinates", "PartType4/Masses", "PartType4/Luminosities"],
+            output_physical=True,
+            a_scale_exponent=2,
+        ),
+        "StellarInertiaTensorReduced_LuminosityWeighted": Property(
+            name="StellarInertiaTensorReduced_LuminosityWeighted",
+            shape=54, # 6 times each GAMA band
+            dtype=np.float32,
+            unit="dimensionless",
+            description="Reduced 3D inertia tensor computed iteratively from the stellar luminosity distribution, relative to the halo centre and for each GAMA band. Diagonal components and one off-diagonal triangle as (1,1), (2,2), (3,3), (1,2), (1,3), (2,3). Only calculated when we have more than 20 particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType4/Coordinates", "PartType4/Masses", "PartType4/Luminosities"],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
+        "StellarInertiaTensorNoniterative_LuminosityWeighted": Property(
+            name="StellarInertiaTensorNoniterative_LuminosityWeighted",
+            shape=54, # 6 times each GAMA band
+            dtype=np.float32,
+            unit="snap_length**2",
+            description="3D inertia tensor computed in a single iteration from the stellar luminosity distribution, relative to the halo centre and for each GAMA band. Diagonal components and one off-diagonal triangle as (1,1), (2,2), (3,3), (1,2), (1,3), (2,3). Only calculated when we have more than 20 particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType4/Coordinates", "PartType4/Masses", "PartType4/Luminosities"],
+            output_physical=True,
+            a_scale_exponent=2,
+        ),
+        "StellarInertiaTensorReducedNoniterative_LuminosityWeighted": Property(
+            name="StellarInertiaTensorReducedNoniterative_LuminosityWeighted",
+            shape=54, # 6 times each GAMA band
+            dtype=np.float32,
+            unit="dimensionless",
+            description="Reduced 3D inertia tensor computed in a single iteration from the stellar luminosity distribution, relative to the halo centre and for each GAMA band. Diagonal components and one off-diagonal triangle as (1,1), (2,2), (3,3), (1,2), (1,3), (2,3). Only calculated when we have more than 20 particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType4/Coordinates", "PartType4/Masses", "PartType4/Luminosities"],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
         "StellarLuminosity": Property(
             name="StellarLuminosity",
             shape=9,
