### 2025-03-24 c65d9e9
Compute particle positions for each band. Perhaps unneccessary since they will have the same positions in every band... Check how to improve later on.
