### 2025-03-24 99b5b64
Update break condition to be done on a luminosity band basis
