### 2025-03-24 71c7c2f
Assign luminosity-based weights
