### 2025-03-24 3bdf43e
Check each halo type for valid properties
