### 2025-03-24 4f3e85d
Check each halo type for valid properties (#137)
