### 2025-03-25 8c93fd1
is_converged is now a boolean array
