### 2025-03-25 883d480
Remove comment that was already addressed
