### 2025-03-25 7e7578a
Tensor for each band done. Still need to check if it gives me the correct element ordering relative to the default calculation of the inertia tensor.
