### 2025-03-25 3e35a81
Re-add removed calculation of inertia tensor eigenvalue/vectors
