### 2025-03-25 395f30e
Create variable to hold the number of provided luminosity bands
