### 2025-03-25 cd1a169
Return flattened matricies for each band
