### 2025-03-25 3cd34df
Remove bug in stopping criteria for iterative tensor
