### 2025-03-26 8b4bdb4
Added missing np.abs() call It was causing the new function to stop iterating early
