+++ SOAP/property_calculation/kinematic_properties.py
@@ -549,7 +549,7 @@ def get_inertia_tensor_luminosity_weighted(
         # Identify bands with converged results. The calculations will only be
         # done for bands that are not yet converged. If all are converged, we are 
         # done.
-        is_converged[((old_q - q) / q) < tol] = 1
+        is_converged[np.abs((old_q - q) / q) < tol] = 1
         if (~is_converged).sum() == 0:
           break
 
