### 2025-03-26 7362357
Fix to prevent overflows into negative values for eigenvectors of inertia tensor
