### 2025-03-26 #7851 @robjmcgibbon
Hi @warrickball, we are hoping to submit some more papers in a couple of weeks which make use of the SOAP package. We'd like to put a preprint version of the paper on arxiv if it hasn't been reviewed by then. In that case is it possible to get a version of the paper without the large "DRAFT" text in the background?
