### 2025-03-28 7167008
HostHaloIndex points to itself for centrals
