### 2025-03-28 bf47acd
Have central's HostHaloIndex point to themselves (#138)
