### 2025-03-31 e558ac1
Add more EAGLE snapshot properties
