+++ misc/convert_eagle.py
@@ -244,8 +244,8 @@ if comm_rank == 0:
 #                         should be h-free
 #
 # In the case of EAGLE snapshots most of this information is available
-# in the metadata of the snapshots themselves. We therefore leave those
-# values blank in the properties dictionary, and read them in directly.
+# in the metadata of the snapshots themselves. We therefore set those
+# values as None in the properties dictionary, and read them in directly.
 properties = {
     "PartType0": {
         "Coordinates": {
@@ -283,6 +283,58 @@ properties = {
             "description": None,
             "conversion_factor": None,
         },
+        "Density": {
+            "swift_name": "Densities",
+            "exponents": {"L": -3, "M": 1, "T": 0, "t": 0},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "InternalEnergy": {
+            "swift_name": "InternalEnergies",
+            "exponents": {"L": 2, "M": 0, "T": 0, "t": -2},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "IronMassFracFromSNIa": {
+            "swift_name": "IronMassFractionsFromSNIa",
+            "exponents": {"L": 0, "M": 0, "T": 0, "t": 0},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "Metallicity": {
+            "swift_name": "MetalMassFractions",
+            "exponents": {"L": 0, "M": 0, "T": 0, "t": 0},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "StarFormationRate": {
+            "swift_name": "StarFormationRates",
+            "exponents": {"L": 0, "M": 1, "T": 0, "t": -1},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "Temperature": {
+            "swift_name": "Temperatures",
+            "exponents": {"L": 0, "M": 0, "T": 1, "t": 0},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        # Note this is handled differently to normal properties.
+        # EAGLE stores each element in its own group, SWIFT expects
+        # all elements to be in a 2D array
+        "ElementMassFractions": {
+            "swift_name": "ElementMassFractions",
+            "exponents": None,
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
     },
     "PartType1": {
         "Coordinates": {
@@ -299,7 +351,9 @@ properties = {
             "description": None,
             "conversion_factor": None,
         },
-        # Note this is different to all other properties
+        # Note this is handled differently to normal properties.
+        # All particles have the same mass in EAGLE, so there
+        # is no mass dataset for dark matter.
         "Mass": {
             "swift_name": "Masses",
             "exponents": {"L": 0, "M": 1, "T": 0, "t": 0},
@@ -358,6 +412,51 @@ properties = {
             "description": None,
             "conversion_factor": None,
         },
+        "BirthDensity": {
+            "swift_name": "BirthDensities",
+            "exponents": {"L": -3, "M": 1, "T": 0, "t": 0},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "InitialMass": {
+            "swift_name": "InitialMasses",
+            "exponents": {"L": 0, "M": 1, "T": 0, "t": 0},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
[…]
