+++ parameter_files/EAGLE.yml
@@ -2,9 +2,9 @@
 # The simulation name (box size and resolution) and snapshot will be appended
 # to these to get the full name of the input/output files/directories
 Parameters:
-  sim_dir: /snap7/scratch/dp004/dc-mcgi1
-  output_dir: /snap7/scratch/dp004/dc-mcgi1
-  scratch_dir: /snap7/scratch/dp004/dc-mcgi1
+  sim_dir: /snap7/scratch/dp004/dc-mcgi1/SOAP_EAGLE
+  output_dir: /snap7/scratch/dp004/dc-mcgi1/SOAP_EAGLE
+  scratch_dir: /snap7/scratch/dp004/dc-mcgi1/SOAP_EAGLE
 
 # Location of the Swift snapshots:
 Snapshots:
@@ -24,44 +24,288 @@ HaloProperties:
   # Where to write the halo properties file
   filename: "{output_dir}/{sim_name}/SOAP_uncompressed/halo_properties_{snap_nr:04d}.hdf5"
   # Where to write temporary chunk output
-  chunk_dir: "{scratch_dir}/SOAP-tmp/"
+  chunk_dir: "{scratch_dir}/{sim_name}/SOAP-tmp/"
 
 ApertureProperties:
   properties:
+    AngularMomentumBaryons: general
+    AngularMomentumDarkMatter: general
+    AngularMomentumGas: general
+    AngularMomentumStars: general
+    BlackHolesDynamicalMass: true
+    BlackHolesSubgridMass: true
+    CentreOfMass: true
+    CentreOfMassVelocity: true
+    DarkMatterMass: true
+    DarkMatterVelocityDispersionMatrix: general
+    DiscToTotalGasMassFraction: general
+    DiscToTotalStellarMassFraction: general
+    GasMass: true
+    GasMassFractionInIron: true
+    GasMassFractionInMetals: true
+    GasMassFractionInOxygen: true
+    GasMassInColdDenseGas: true
+    GasTemperature: true
+    GasVelocityDispersionMatrix: general
+    HalfMassRadiusBaryons: general
+    HalfMassRadiusDarkMatter: general
+    HalfMassRadiusGas: general
     HalfMassRadiusStars: true
+    HeliumMass: true
+    HydrogenMass: true
+    KappaCorotBaryons: general
+    KappaCorotGas: general
+    KappaCorotStars: general
+    KineticEnergyGas: general
+    KineticEnergyStars: general
+    LinearMassWeightedIronFromSNIaOverHydrogenOfStars: true
+    LinearMassWeightedIronOverHydrogenOfStars: true
+    LinearMassWeightedMagnesiumOverHydrogenOfStars: true
+    LinearMassWeightedOxygenOverHydrogenOfGas: true
+    LinearMassWeightedNitrogenOverOxygenOfGas: true
+    LinearMassWeightedCarbonOverOxygenOfGas: true
+    LogarithmicMassWeightedIronFromSNIaOverHydrogenOfStarsLowLimit: true
+    LogarithmicMassWeightedIronOverHydrogenOfStarsHighLimit: true
+    LogarithmicMassWeightedMagnesiumOverHydrogenOfStarsHighLimit: true
+    LogarithmicMassWeightedIronOverHydrogenOfStarsLowLimit: true
+    LogarithmicMassWeightedMagnesiumOverHydrogenOfStarsLowLimit: true
+    MassWeightedMeanStellarAge: true
+    MostMassiveBlackHoleAccretionRate: true
+    MostMassiveBlackHoleID: true
+    MostMassiveBlackHoleMass: true
+    MostMassiveBlackHolePosition: true
+    MostMassiveBlackHoleVelocity: true
[…]
