### 2025-04-01 783b56c
Defined luminosity weighted projected inertia tensor
