+++ SOAP/particle_selection/projected_aperture_properties.py
@@ -1033,6 +1033,61 @@ class SingleProjectionProjectedApertureParticleData:
         )
 
     @lazy_property
+    def ProjectedStellarInertiaTensor_LuminosityWeighted(self) -> unyt.unyt_array:
+        """
+        Inertia tensor of the stellar luminosity distribution for each GAMA band in projection.
+        Computed iteratively using an ellipse with area equal to that of a circle with radius
+        equal to the aperture radius. Only considers bound particles within the projected aperture.
+        """
+        if self.Mstar == 0:
+            return None
+        return self.stellar_inertia_tensor()
+
+    @lazy_property
+    def ProjectedStellarInertiaTensorReduced_LuminosityWeighted(self) -> unyt.unyt_array:
+        """
+        Reduced inertia tensor of the stellar luminosity distribution for each GAMA band in projection.
+        Computed iteratively using an ellipse with area equal to that of a circle with radius
+        equal to the aperture radius. Only considers bound particles within the projected aperture.
+        """
+        if self.Mstar == 0:
+            return None
+        return self.stellar_inertia_tensor(reduced=True)
+
+    @lazy_property
+    def ProjectedStellarInertiaTensorNoniterative_LuminosityWeighted(self) -> unyt.unyt_array:
+        """
+        Inertia tensor of the stellar luminosity distribution for each GAMA band in projection.
+        Computed using all bound star particles within the projected aperture.
+        """
+        if self.Mstar == 0:
+            return None
+        return get_projected_inertia_tensor(
+            self.proj_mass_star,
+            self.proj_pos_star,
+            self.iproj,
+            self.aperture_radius,
+            max_iterations=1,
+        )
+
+    @lazy_property
+    def ProjectedStellarInertiaTensorReducedNoniterative_LuminosityWeighted(self) -> unyt.unyt_array:
+        """
+        Reduced inertia tensor of the stellar luminosity distribution for each GAMA band in projection.
+        Computed using all bound star particles within the projected aperture.
+        """
+        if self.Mstar == 0:
+            return None
+        return get_projected_inertia_tensor(
+            self.proj_mass_star,
+            self.proj_pos_star,
+            self.iproj,
+            self.aperture_radius,
+            reduced=True,
+            max_iterations=1,
+        )
+
+    @lazy_property
     def gas_mask_all(self) -> NDArray[bool]:
         """
         Mask for masking out gas particles in raw PartType0 arrays.
@@ -1469,15 +1524,19 @@ class ProjectedApertureProperties(HaloProperty):
         "ProjectedTotalInertiaTensor": True,
         "ProjectedGasInertiaTensor": True,
         "ProjectedStellarInertiaTensor": True,
+        "ProjectedStellarInertiaTensor_LuminosityWeighted": True,
         "ProjectedTotalInertiaTensorReduced": True,
         "ProjectedGasInertiaTensorReduced": True,
         "ProjectedStellarInertiaTensorReduced": True,
+        "ProjectedStellarInertiaTensorReduced_LuminosityWeighted": True,
         "ProjectedTotalInertiaTensorNoniterative": False,
         "ProjectedGasInertiaTensorNoniterative": False,
         "ProjectedStellarInertiaTensorNoniterative": False,
+        "ProjectedStellarInertiaTensorNoniterative_LuminosityWeighted": False,
         "ProjectedTotalInertiaTensorReducedNoniterative": False,
         "ProjectedGasInertiaTensorReducedNoniterative": False,
         "ProjectedStellarInertiaTensorReducedNoniterative": False,
+        "ProjectedStellarInertiaTensorReducedNoniterative_LuminosityWeighted": False,
         "HydrogenMass": False,
         "HeliumMass": False,
         "MolecularHydrogenMass": False,
