+++ SOAP/property_table.py
@@ -2002,6 +2002,54 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=0,
         ),
+        "ProjectedStellarInertiaTensor_LuminosityWeighted": Property(
+            name="ProjectedStellarInertiaTensor_LuminosityWeighted",
+            shape=27,
+            dtype=np.float32,
+            unit="snap_length**2",
+            description="2D inertia tensor computed iteratively from the stellar luminosity distribution in different GAMA bands, relative to the halo centre. Diagonal components and one off-diagonal value as (1,1), (2,2), (1,2). Only calculated when we have more than 20 particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType4/Coordinates", "PartType4/Masses", "PartType4/Luminosities"],
+            output_physical=True,
+            a_scale_exponent=2,
+        ),
+        "ProjectedStellarInertiaTensorReduced_LuminosityWeighted": Property(
+            name="ProjectedStellarInertiaTensorReduced_LuminosityWeighted",
+            shape=27,
+            dtype=np.float32,
+            unit="dimensionless",
+            description="Reduced 2D inertia tensor computed iteratively from the stellar luminosity distribution in different GAMA bands, relative to the halo centre. Diagonal components and one off-diagonal value as (1,1), (2,2), (1,2). Only calculated when we have more than 20 particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType4/Coordinates", "PartType4/Masses", "PartType4/Luminosities"],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
+        "ProjectedStellarInertiaTensorNoniterative_LuminosityWeighted": Property(
+            name="ProjectedStellarInertiaTensorNoniterative_LuminosityWeighted",
+            shape=27,
+            dtype=np.float32,
+            unit="snap_length**2",
+            description="2D inertia tensor computed in a single iteration from the stellar luminosity distribution in different GAMA bands, relative to the halo centre. Diagonal components and one off-diagonal value as (1,1), (2,2), (1,2). Only calculated when we have more than 20 particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType4/Coordinates", "PartType4/Masses", "PartType4/Luminosities"],
+            output_physical=True,
+            a_scale_exponent=2,
+        ),
+        "ProjectedStellarInertiaTensorReducedNoniterative_LuminosityWeighted": Property(
+            name="ProjectedStellarInertiaTensorReducedNoniterative_LuminosityWeighted",
+            shape=27,
+            dtype=np.float32,
+            unit="dimensionless",
+            description="Reduced 2D inertia tensor computed in a single iteration from the stellar luminosity distribution in different GAMA bands, relative to the halo centre. Diagonal components and one off-diagonal value as (1,1), (2,2), (1,2). Only calculated when we have more than 20 particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType4/Coordinates", "PartType4/Masses", "PartType4/Luminosities"],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
         "SFR": Property(
             name="StarFormationRate",
             shape=1,
