### 2025-04-03 5a7c9f6
Allow for different alias for snipshots
