### 2025-04-03 1294548
Allow for different alias for snipshots (#139)
