### 2025-04-03 86d3aaa
Add InclusiveSphere/EncloseRadius definition
