+++ SOAP/property_table.py
@@ -835,6 +835,21 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=0,
         ),
+        "DustMass": Property(
+            name="DustMass",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_mass",
+            description="Total dust mass.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType0/Masses",
+                "PartType0/TotalDustMassFractions",
+            ],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
         "DustGraphiteMass": Property(
             name="DustGraphiteMass",
             shape=1,
@@ -1171,6 +1186,22 @@ class PropertyTable:
             lossy_compression_filter="FMantissa9",
             dmo_property=True,
             particle_properties=["PartType1/Coordinates", "PartType1/Masses"],
+            output_physical=False,
+            a_scale_exponent=1,
+        ),
+        "HalfMassRadiusDust": Property(
+            name="HalfMassRadiusDust",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_length",
+            description="Dust half mass radius.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType0/Coordinates",
+                "PartType0/Masses",
+                "PartType0/TotalDustMassFractions",
+            ],
             output_physical=False,
             a_scale_exponent=1,
         ),

[…]
