### 2025-04-11 69cc6eb
Add stellar CoM
