### 2025-04-11 3da7fed
Small refactor
