### 2025-04-16 3dcb4da
Adding a number of extra properties (#141)
* Add dust properties

* Add stellar CoM

* Update test

* Minor name update
