### 2025-04-16 fd32129
Don't update skipped SOAP properties
