+++ SOAP/core/combine_chunks.py
@@ -91,44 +91,51 @@ def combine_chunks(
     # Determine total number of halos
     total_nr_halos = comm_world.allreduce(len(order))
 
-    # Certain properties need to be kept for calculating the SOAP properties
-    props_to_keep = set()
-    subhalo_rank_props = {
-        "VR": (
+    # Handle derived (SOAP) quantities: these don't exist in the chunk
+    # output but will be computed by combining other halo properties.
+    soap_props = set()      # SOAP properties which will be calculated
+    props_to_keep = set()   # Properties from the chunk files which are
+                            # required in order to calculate the SOAP properties
+    if args.halo_format == 'VR':
+        # Subhalo rank
+        soap_props.add("SOAP/SubhaloRankByBoundMass")
+        props_to_keep.update((
             "InputHalos/VR/ID",
             "BoundSubhalo/TotalMass",
             "InputHalos/VR/HostHaloID",
-        ),
-        "HBTplus": (
+        ))
+        # Host halo index
+        soap_props.add("SOAP/HostHaloIndex")
+        props_to_keep.update((
+            "InputHalos/VR/ID", "InputHalos/VR/HostHaloID"
+        ))
+    elif args.halo_format == 'HBTplus':
+        # Subhalo rank
+        soap_props.add("SOAP/SubhaloRankByBoundMass")
+        props_to_keep.update((
             "InputHalos/HBTplus/HostFOFId",
             "BoundSubhalo/TotalMass",
             "InputHalos/HBTplus/TrackId",
-        ),
-    }.get(args.halo_format, ())
-    props_to_keep.update(subhalo_rank_props)
-    host_halo_index_props = {
-        "VR": ("InputHalos/VR/ID", "InputHalos/VR/HostHaloID"),
-        "HBTplus": ("InputHalos/HBTplus/HostFOFId", "InputHalos/IsCentral"),
-    }.get(args.halo_format, ())
-    props_to_keep.update(host_halo_index_props)
-    fof_props = {
-        "HBTplus": ("InputHalos/HBTplus/HostFOFId", "InputHalos/IsCentral")
-    }.get(args.halo_format, ())
-    props_to_keep.update(fof_props)
-    index_previous_snap_props = {
-        "HBTplus": ("InputHalos/HBTplus/TrackId")
-    }.get(args.halo_format, ())
-    props_to_keep.update(index_previous_snap_props)
+        ))
+        # Host halo index
+        soap_props.add("SOAP/HostHaloIndex")
+        props_to_keep.update((
+            "InputHalos/HBTplus/HostFOFId", "InputHalos/IsCentral"
+        ))
+
     # Also keep M200c for calculating reduced_snapshot flag
     if "reduced_snapshots" in args.calculations:
+        # TODO: Check ref_metadata for SO 200_crit
+        # for metadata in ref_metadata
+        soap_props.add("SOAP/IncludedInReducedSnapshot")
         props_to_keep.add("SO/200_crit/TotalMass")
-    props_kept = {}
 
-    # Get metadata for derived quantities: these don't exist in the chunk
-    # output but will be computed by combining other halo properties.
+    # Get metadata for derived quantities
     soap_metadata = []
     for key, prop in PropertyTable.full_property_list.items():
         if not key.startswith("SOAP/"):
+            continue
+        if not key in soap_props:
             continue
         name = prop.name
         size = prop.shape
@@ -172,6 +179,11 @@ def combine_chunks(
             fof_metadata.append(
                 (name, size, unit, dtype, description, physical, a_exponent)
             )
+
+        # Keep the properties required for matching subhalos to FOFs
+        props_to_keep.update((
+            "InputHalos/HBTplus/HostFOFId", "InputHalos/IsCentral"
+        ))
 
     # First MPI rank sets up the output file
[…]
