### 2025-04-17 9cd103f
Remove references to centre of potential (#143)
