### 2025-04-17 00bcef2
Calculate and store progenitor/descendant indices
