+++ SOAP/core/combine_chunks.py
@@ -1,5 +1,6 @@
 #!/bin/env python
 
+import os
 import socket
 import time
 
@@ -10,6 +11,7 @@ import virgo.mpi.parallel_hdf5 as phdf5
 import virgo.mpi.parallel_sort as psort
 from virgo.util.partial_formatter import PartialFormatter
 
+from catalogue_readers import read_hbtplus
 from property_calculation.subhalo_rank import compute_subhalo_rank
 from . import swift_units
 from .mpi_timer import MPITimer
@@ -26,6 +28,36 @@ def sub_snapnum(filename, snapnum):
     pf = PartialFormatter()
     filename = pf.format(filename, snap_nr=snapnum, file_nr=None)
     return filename
+
+
+def spatial_sort(halo_cofp, halo_index, cellgrid, comm):
+    """
+    Sort the halos based on what cell they are in, then based
+    on their catalogue index
+    """
+    cell_indices = (halo_cofp // cellgrid.cell_size).value.astype("int64")
+    assert cellgrid.dimension[0] >= cellgrid.dimension[1] >= cellgrid.dimension[2]
+    # Assert that all halos are within the box
+    assert np.min(cell_indices) >= 0
+    for i_cell in range(3):
+        assert np.max(cell_indices[:, i_cell]) < cellgrid.dimension[i_cell]
+    # Sort first based on position, then on catalogue index
+    sort_hash_dtype = [("cell_index", np.int64), ("catalogue_index", np.int64)]
+    sort_hash = np.zeros(cell_indices.shape[0], dtype=sort_hash_dtype)
+    sort_hash["cell_index"] += cell_indices[:, 0] * cellgrid.dimension[0] ** 2
+    sort_hash["cell_index"] += cell_indices[:, 1] * cellgrid.dimension[1]
+    sort_hash["cell_index"] += cell_indices[:, 2]
+    sort_hash["catalogue_index"] = halo_index
+    order = psort.parallel_sort(sort_hash, return_index=True, comm=comm)
+
+    # Calculate local count of halos in each cell, and combine on rank 0
+    local_cell_counts = np.bincount(
+        sort_hash["cell_index"], minlength=cellgrid.nr_cells[0]
+    ).astype("int64")
+    assert local_cell_counts.shape[0] == np.prod(cellgrid.dimension)
+    cell_counts = comm.reduce(local_cell_counts)
+
+    return order, cell_counts
 
 
 def combine_chunks(
@@ -60,33 +92,11 @@ def combine_chunks(
     ):
         halo_cofp = scratch_file.read("InputHalos/HaloCentre") * cofp_units
         halo_index = scratch_file.read("InputHalos/HaloCatalogueIndex")
-        cell_indices = (halo_cofp // cellgrid.cell_size).value.astype("int64")
-        assert cellgrid.dimension[0] >= cellgrid.dimension[1] >= cellgrid.dimension[2]
-        # Assert that all halos are within the box
-        assert np.min(cell_indices) >= 0
-        for i_cell in range(3):
-            assert np.max(cell_indices[:, i_cell]) < cellgrid.dimension[i_cell]
-        # Sort first based on position, then on catalogue index
-        sort_hash_dtype = [("cell_index", np.int64), ("catalogue_index", np.int64)]
-        sort_hash = np.zeros(cell_indices.shape[0], dtype=sort_hash_dtype)
-        sort_hash["cell_index"] += cell_indices[:, 0] * cellgrid.dimension[0] ** 2
-        sort_hash["cell_index"] += cell_indices[:, 1] * cellgrid.dimension[1]
-        sort_hash["cell_index"] += cell_indices[:, 2]
-        sort_hash["catalogue_index"] = halo_index
[…]
