+++ SOAP/property_table.py
@@ -4234,6 +4234,30 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=None,
         ),
+        "SOAP/ProgenitorIndex": Property(
+            name="SOAP/ProgenitorIndex",
+            shape=1,
+            dtype=np.int32,
+            unit="dimensionless",
+            description="Index (within the previous snapshot SOAP arrays) of the main progenitor of this subhalo.",
+            lossy_compression_filter="None",
+            dmo_property=True,
+            particle_properties=[],
+            output_physical=True,
+            a_scale_exponent=None,
+        ),
+        "SOAP/DescendantIndex": Property(
+            name="SOAP/DescendantIndex",
+            shape=1,
+            dtype=np.int32,
+            unit="dimensionless",
+            description="Index (within the next snapshot SOAP arrays) of the main descendant of this subhalo.",
+            lossy_compression_filter="None",
+            dmo_property=True,
+            particle_properties=[],
+            output_physical=True,
+            a_scale_exponent=None,
+        ),
     }
 
     def __init__(self, parameters, snipshot_parameters, units_cgs):
