### 2025-04-17 0f81e4d
Update handling of named columns
