+++ SOAP/core/snapshot_datasets.py
@@ -36,8 +36,6 @@ class SnapshotDatasets:
     dataset_map: Dict
     # mapping from dataset + column names to column index
     named_columns: Dict
-    # grain compositions in the dust model (currently not used)
-    dust_grain_composition: NDArray[float]
     # constants defined in the parameter file
     defined_constants: Dict
 
@@ -72,28 +70,24 @@ class SnapshotDatasets:
                     "NamedColumns" not in file_handle["SubgridScheme"]
                 ):
                     continue
+
                 # As the snapshot filename is done first, if one of the extra-input
                 # files has a named column entry in common with the snapshot then
-                # we use then one from the extra-input file.
-                for name in file_handle["SubgridScheme"]["NamedColumns"]:
-                    column_names = file_handle["SubgridScheme"]["NamedColumns"][name][:]
-                    self.named_columns[name] = {}
-                    # turn the list into a dictionary that maps a column name to
-                    # a colum index
-                    for iname, colname in enumerate(column_names):
-                        self.named_columns[name][colname.decode("utf-8")] = iname
-
-                try:
-                    self.dust_grain_composition = file_handle["SubgridScheme"][
-                        "GrainToElementMapping"
-                    ][:]
-                except KeyError:
-                    try:
-                        self.dust_grain_composition = file_handle["SubgridScheme"][
-                            "DustMassFractionsToElementMassFractionsMapping"
-                        ][:]
-                    except KeyError:
-                        pass
+                # we use then one from the extra-input file. We append the ptype
+                # in case there of conflict (e.g. we use PartType0/SpeciesFractions
+                # from the snapshot file, but PartType4/SpeciesFractions from the
+                # extra-input files.
+                for dset in file_handle["SubgridScheme"]["NamedColumns"]:
+                    column_names = file_handle["SubgridScheme"]["NamedColumns"][dset][:]
+                    for group, datasets in self.datasets_in_file.items():
+                        if not dset in datasets:
+                            continue
+                        name = f'{group}/{dset}'
+                        self.named_columns[name] = {}
+                        # turn the list into a dictionary that maps a column name to
+                        # a colum index
+                        for iname, colname in enumerate(column_names):
+                            self.named_columns[name][colname.decode("utf-8")] = iname
 
     def setup_aliases(self, aliases: Dict):
         """
@@ -130,8 +124,8 @@ class SnapshotDatasets:
             SOAP_ptype, SOAP_dset = alias.split("/")
             snap_ptype, snap_dset = aliases[alias].split("/")
             self.dataset_map[alias] = (snap_ptype, snap_dset)
-            if snap_dset in self.named_columns:
-                self.named_columns[SOAP_dset] = dict(self.named_columns[snap_dset])
+            if aliases[alias] in self.named_columns:
+                self.named_columns[alias] = dict(self.named_columns[aliases[alias]])
 
     def setup_defined_constants(self, defined_constants: Dict):
         """
@@ -185,36 +179,13 @@ class SnapshotDatasets:
             raise KeyError(f"Failed to read {name} from input files!")
         return data_dict[ptype][dset]
 
-    def get_dataset_column(
-        self, name: str, column_name: str, data_dict: Dict
-    ) -> unyt.unyt_array:
-        """
-        Get the data for the given named column in the dataset with the given
-        generic name.
-
-        Parameters:
-         - name: str
-           Generic name of a dataset, as used by halo property calculations.
-         - column_name: str
-           Name of a named column, as defined in the snapshot metadata and used
-           by halo property calculations.
-         - data_dict: Dict
-           Dictionary of particle properties, as read from the snapshot.
-
-        Returns the corresponding data, taking into account potential
-        aliases and the named column metadata.
[…]
