+++ SOAP/particle_selection/aperture_properties.py
@@ -598,7 +598,7 @@ class ApertureParticleData:
             self.star_element_fractions[
                 :,
                 self.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Oxygen"
+                    "PartType4/ElementMassFractions", "Oxygen"
                 ),
             ]
             * self.mass_star
@@ -615,7 +615,7 @@ class ApertureParticleData:
             self.star_element_fractions[
                 :,
                 self.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Magnesium"
+                    "PartType4/ElementMassFractions", "Magnesium"
                 ),
             ]
             * self.mass_star
@@ -631,7 +631,7 @@ class ApertureParticleData:
         return (
             self.star_element_fractions[
                 :,
-                self.snapshot_datasets.get_column_index("ElementMassFractions", "Iron"),
+                self.snapshot_datasets.get_column_index("PartType4/ElementMassFractions", "Iron"),
             ]
             * self.mass_star
         )
@@ -714,7 +714,7 @@ class ApertureParticleData:
         if self.Nstar == 0:
             return None
         Lr = self.stellar_luminosities[
-            :, self.snapshot_datasets.get_column_index("Luminosities", "GAMA_r")
+            :, self.snapshot_datasets.get_column_index("PartType4/Luminosities", "GAMA_r")
         ]
         Lrtot = Lr.sum()
         if Lrtot == 0:
@@ -1543,7 +1543,7 @@ class ApertureParticleData:
             ][
                 :,
                 self.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Oxygen"
+                    "PartType0/ElementMassFractions", "Oxygen"
                 ),
             ]
         )
@@ -1579,7 +1579,7 @@ class ApertureParticleData:
                 self.gas_mask_ap
             ][
                 :,
-                self.snapshot_datasets.get_column_index("ElementMassFractions", "Iron"),
+                self.snapshot_datasets.get_column_index("PartType0/ElementMassFractions", "Iron"),
             ]
         )
 
@@ -1687,7 +1687,7 @@ class ApertureParticleData:
             self.gas_element_fractions[
                 :,
                 self.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Hydrogen"
+                    "PartType0/ElementMassFractions", "Hydrogen"
                 ),
             ]
             * self.mass_gas
@@ -1704,7 +1704,7 @@ class ApertureParticleData:
             self.gas_element_fractions[
                 :,
                 self.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Helium"
+                    "PartType0/ElementMassFractions", "Helium"
                 ),
             ]
             * self.mass_gas
@@ -1731,7 +1731,7 @@ class ApertureParticleData:
         return (
             self.gas_mass_H
             * self.gas_species_fractions[
-                :, self.snapshot_datasets.get_column_index("SpeciesFractions", "HI")
+                :, self.snapshot_datasets.get_column_index("PartType0/SpeciesFractions", "HI")
             ]
         )
 
@@ -1745,7 +1745,7 @@ class ApertureParticleData:
         return (
             self.gas_mass_H
             * self.gas_species_fractions[
-                :, self.snapshot_datasets.get_column_index("SpeciesFractions", "H2")
+                :, self.snapshot_datasets.get_column_index("PartType0/SpeciesFractions", "H2")
             ]
             * 2.0
         )
@@ -1807,7 +1807,7 @@ class ApertureParticleData:
[…]
