+++ SOAP/particle_selection/aperture_properties.py
@@ -631,7 +631,9 @@ class ApertureParticleData:
         return (
             self.star_element_fractions[
                 :,
-                self.snapshot_datasets.get_column_index("PartType4/ElementMassFractions", "Iron"),
+                self.snapshot_datasets.get_column_index(
+                    "PartType4/ElementMassFractions", "Iron"
+                ),
             ]
             * self.mass_star
         )
@@ -714,7 +716,8 @@ class ApertureParticleData:
         if self.Nstar == 0:
             return None
         Lr = self.stellar_luminosities[
-            :, self.snapshot_datasets.get_column_index("PartType4/Luminosities", "GAMA_r")
+            :,
+            self.snapshot_datasets.get_column_index("PartType4/Luminosities", "GAMA_r"),
         ]
         Lrtot = Lr.sum()
         if Lrtot == 0:
@@ -1579,7 +1582,9 @@ class ApertureParticleData:
                 self.gas_mask_ap
             ][
                 :,
-                self.snapshot_datasets.get_column_index("PartType0/ElementMassFractions", "Iron"),
+                self.snapshot_datasets.get_column_index(
+                    "PartType0/ElementMassFractions", "Iron"
+                ),
             ]
         )
 
@@ -1731,7 +1736,10 @@ class ApertureParticleData:
         return (
             self.gas_mass_H
             * self.gas_species_fractions[
-                :, self.snapshot_datasets.get_column_index("PartType0/SpeciesFractions", "HI")
+                :,
+                self.snapshot_datasets.get_column_index(
+                    "PartType0/SpeciesFractions", "HI"
+                ),
             ]
         )
 
@@ -1745,7 +1753,10 @@ class ApertureParticleData:
         return (
             self.gas_mass_H
             * self.gas_species_fractions[
-                :, self.snapshot_datasets.get_column_index("PartType0/SpeciesFractions", "H2")
+                :,
+                self.snapshot_datasets.get_column_index(
+                    "PartType0/SpeciesFractions", "H2"
+                ),
             ]
             * 2.0
         )
@@ -1946,7 +1957,10 @@ class ApertureParticleData:
             return None
         atomic_gas_mass = (
             self.gas_species_fractions[
-                :, self.snapshot_datasets.get_column_index("PartType0/SpeciesFractions", "HI")
+                :,
+                self.snapshot_datasets.get_column_index(
+                    "PartType0/SpeciesFractions", "HI"
+                ),
             ]
             * self.mass_gas
         )
@@ -1962,7 +1976,10 @@ class ApertureParticleData:
         molecular_gas_mass = (
             2
             * self.gas_species_fractions[
-                :, self.snapshot_datasets.get_column_index("PartType0/SpeciesFractions", "H2")
+                :,
+                self.snapshot_datasets.get_column_index(
+                    "PartType0/SpeciesFractions", "H2"
+                ),
             ]
             * self.mass_gas
         )
@@ -1998,7 +2015,10 @@ class ApertureParticleData:
             return None
         atomic_gas_mass = (
             self.gas_species_fractions[
-                :, self.snapshot_datasets.get_column_index("PartType0/SpeciesFractions", "HI")
+                :,
+                self.snapshot_datasets.get_column_index(
+                    "PartType0/SpeciesFractions", "HI"
+                ),
             ]
             * self.mass_gas
         )
@@ -2014,7 +2034,10 @@ class ApertureParticleData:
         molecular_gas_mass = (
             2
             * self.gas_species_fractions[
[…]
