### 2025-04-22 b04a60d
Calculate SOAP index of progenitor/descendant (#142)
* Small refactor

* Don't update skipped SOAP properties

* Format

* Calculate and store progenitor/descendant indices

* Update handling of named columns

* Format

* Add footnote
