+++ SOAP/particle_selection/aperture_properties.py
@@ -605,7 +605,7 @@ class ApertureParticleData:
             self.star_element_fractions[
                 :,
                 self.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Oxygen"
+                    "PartType4/ElementMassFractions", "Oxygen"
                 ),
             ]
             * self.mass_star
@@ -622,7 +622,7 @@ class ApertureParticleData:
             self.star_element_fractions[
                 :,
                 self.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Magnesium"
+                    "PartType4/ElementMassFractions", "Magnesium"
                 ),
             ]
             * self.mass_star
@@ -638,7 +638,9 @@ class ApertureParticleData:
         return (
             self.star_element_fractions[
                 :,
-                self.snapshot_datasets.get_column_index("ElementMassFractions", "Iron"),
+                self.snapshot_datasets.get_column_index(
+                    "PartType4/ElementMassFractions", "Iron"
+                ),
             ]
             * self.mass_star
         )
@@ -721,7 +723,8 @@ class ApertureParticleData:
         if self.Nstar == 0:
             return None
         Lr = self.stellar_luminosities[
-            :, self.snapshot_datasets.get_column_index("Luminosities", "GAMA_r")
+            :,
+            self.snapshot_datasets.get_column_index("PartType4/Luminosities", "GAMA_r"),
         ]
         Lrtot = Lr.sum()
         if Lrtot == 0:
@@ -1550,7 +1553,7 @@ class ApertureParticleData:
             ][
                 :,
                 self.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Oxygen"
+                    "PartType0/ElementMassFractions", "Oxygen"
                 ),
             ]
         )
@@ -1586,7 +1589,9 @@ class ApertureParticleData:
                 self.gas_mask_ap
             ][
                 :,
-                self.snapshot_datasets.get_column_index("ElementMassFractions", "Iron"),
+                self.snapshot_datasets.get_column_index(
+                    "PartType0/ElementMassFractions", "Iron"
+                ),
             ]
         )
 
@@ -1694,7 +1699,7 @@ class ApertureParticleData:
             self.gas_element_fractions[
                 :,
                 self.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Hydrogen"
+                    "PartType0/ElementMassFractions", "Hydrogen"
                 ),
             ]
             * self.mass_gas
@@ -1711,7 +1716,7 @@ class ApertureParticleData:
             self.gas_element_fractions[
                 :,
                 self.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Helium"
+                    "PartType0/ElementMassFractions", "Helium"
                 ),
             ]
             * self.mass_gas
@@ -1738,7 +1743,10 @@ class ApertureParticleData:
         return (
             self.gas_mass_H
             * self.gas_species_fractions[
-                :, self.snapshot_datasets.get_column_index("SpeciesFractions", "HI")
+                :,
+                self.snapshot_datasets.get_column_index(
+                    "PartType0/SpeciesFractions", "HI"
+                ),
             ]
         )
 
@@ -1752,7 +1760,10 @@ class ApertureParticleData:
         return (
             self.gas_mass_H
             * self.gas_species_fractions[
-                :, self.snapshot_datasets.get_column_index("SpeciesFractions", "H2")
+                :,
+                self.snapshot_datasets.get_column_index(
+                    "PartType0/SpeciesFractions", "H2"
+                ),
             ]
[…]
