+++ SOAP/particle_selection/projected_aperture_properties.py
@@ -1195,7 +1195,7 @@ class SingleProjectionProjectedApertureParticleData:
             self.gas_element_fractions[
                 :,
                 self.part_props.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Hydrogen"
+                    "PartType0/ElementMassFractions", "Hydrogen"
                 ),
             ]
             * self.proj_mass_gas
@@ -1212,7 +1212,7 @@ class SingleProjectionProjectedApertureParticleData:
             self.gas_element_fractions[
                 :,
                 self.part_props.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Helium"
+                    "PartType0/ElementMassFractions", "Helium"
                 ),
             ]
             * self.proj_mass_gas
@@ -1241,7 +1241,7 @@ class SingleProjectionProjectedApertureParticleData:
             * self.gas_species_fractions[
                 :,
                 self.part_props.snapshot_datasets.get_column_index(
-                    "SpeciesFractions", "HI"
+                    "PartType0/SpeciesFractions", "HI"
                 ),
             ]
         )
@@ -1258,7 +1258,7 @@ class SingleProjectionProjectedApertureParticleData:
             * self.gas_species_fractions[
                 :,
                 self.part_props.snapshot_datasets.get_column_index(
-                    "SpeciesFractions", "H2"
+                    "PartType0/SpeciesFractions", "H2"
                 ),
             ]
             * 2.0
@@ -1322,7 +1322,7 @@ class SingleProjectionProjectedApertureParticleData:
             self.star_element_fractions[
                 :,
                 self.part_props.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Oxygen"
+                    "PartType4/ElementMassFractions", "Oxygen"
                 ),
             ]
             * self.proj_mass_star
@@ -1339,7 +1339,7 @@ class SingleProjectionProjectedApertureParticleData:
             self.star_element_fractions[
                 :,
                 self.part_props.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Magnesium"
+                    "PartType4/ElementMassFractions", "Magnesium"
                 ),
             ]
             * self.proj_mass_star
@@ -1356,7 +1356,7 @@ class SingleProjectionProjectedApertureParticleData:
             self.star_element_fractions[
                 :,
                 self.part_props.snapshot_datasets.get_column_index(
-                    "ElementMassFractions", "Iron"
+                    "PartType4/ElementMassFractions", "Iron"
                 ),
             ]
             * self.proj_mass_star
