### 2025-04-23 39ea24b
NewAperture definition
