+++ SOAP/particle_selection/aperture_properties.py
@@ -190,6 +190,7 @@ class ApertureParticleData:
         data: Dict,
         types_present: List[str],
         inclusive: bool,
+        include_central_particles: bool,
         aperture_radius: unyt.unyt_quantity,
         stellar_age_calculator: StellarAgeCalculator,
         recently_heated_gas_filter: RecentlyHeatedGasFilter,
@@ -212,6 +213,8 @@ class ApertureParticleData:
          - inclusive: bool
            Whether or not to include particles not gravitationally bound to the subhalo
            in the property calculations.
+         - inclusive: bool
+           Whether or not to include particles bound to the host central subhalo
          - aperture_radius: unyt.unyt_quantity
            Aperture radius.
          - stellar_age_calculator: StellarAgeCalculator
@@ -234,6 +237,7 @@ class ApertureParticleData:
         self.data = data
         self.types_present = types_present
         self.inclusive = inclusive
+        self.include_central_particles = include_central_particles
         self.aperture_radius = aperture_radius
         self.stellar_age_calculator = stellar_age_calculator
         self.recently_heated_gas_filter = recently_heated_gas_filter
@@ -256,6 +260,7 @@ class ApertureParticleData:
         """
         self.centre = self.input_halo["cofp"]
         self.index = self.input_halo["index"]
+        self.host_index = self.input_halo["HBTplus/HostHaloCatalogueIndex"]
         mass = []
         position = []
         radius = []
@@ -266,8 +271,13 @@ class ApertureParticleData:
             grnr = self.get_dataset(f"{ptype}/GroupNr_bound")
             if self.inclusive:
                 in_halo = np.ones(grnr.shape, dtype=bool)
+                assert not self.include_central_particles
             else:
                 in_halo = grnr == self.index
+                # TODO: This is a massive waste of time for centrals, we need to skip them
+                if self.include_central_particles:
+                    # print(np.sum(grnr == self.host_index))
+                    in_halo |= grnr == self.host_index
             mass.append(self.get_dataset(f"{ptype}/{mass_dataset(ptype)}")[in_halo])
             pos = (
                 self.get_dataset(f"{ptype}/Coordinates")[in_halo, :]
@@ -289,7 +299,7 @@ class ApertureParticleData:
         self.types = np.concatenate(types)
         self.softening = np.concatenate(softening)
 
-        self.mask = self.radius <= self.aperture_radius
+        self.mask = self.radius <= 1.0001 * self.aperture_radius
 
         self.mass = self.mass[self.mask]
         self.position = self.position[self.mask]
@@ -526,7 +536,10 @@ class ApertureParticleData:
         if self.inclusive:
             return np.ones(groupnr_bound.shape, dtype=bool)
         else:
-            return groupnr_bound == self.index
+            in_halo = groupnr_bound == self.index
+            if self.include_central_particles:
+                in_halo |= groupnr_bound == self.host_index
+            return in_halo
 
     @lazy_property
     def Mstar_init(self) -> unyt.unyt_quantity:
@@ -745,7 +758,10 @@ class ApertureParticleData:
         if self.inclusive:
             return np.ones(groupnr_bound.shape, dtype=bool)
         else:
-            return groupnr_bound == self.index
+            in_halo = groupnr_bound == self.index
+            if self.include_central_particles:
+                in_halo |= groupnr_bound == self.host_index
+            return in_halo
 
     @lazy_property
     def BH_subgrid_masses(self) -> unyt.unyt_array:
@@ -1410,7 +1426,10 @@ class ApertureParticleData:
         if self.inclusive:
             return np.ones(groupnr_bound.shape, dtype=bool)
         else:
-            return groupnr_bound == self.index
+            in_halo = groupnr_bound == self.index
[…]
