### 2025-04-23 #7851 @robjmcgibbon
I put together a list of a few peole who might be suitable to review this, but haven't worked with SOAP at all. Some are people who have worked on SWIFT, others are from looking through the JOSS reviewer list. I haven't had any contact with these people re a review.

Stuart McAlpine		stuart.mcalpine@fysik.su.se
Mark Lovell 		        m.r.lovell@durham.ac.uk
Dylan Nelson		dnelson@uni-heidelberg.de
Thomas Berlok		tberlok@aip.de
Terrence Tricco		https://github.com/ttricco
