### 2025-04-28 d49e450
Use proper index for SOAP EAGLE
