+++ scripts/EAGLE.sh
@@ -51,8 +51,24 @@ set -e
 
 mpirun -- python -u misc/convert_eagle.py \
     --snap-basename "${output_dir}/gadget_snapshots/snapshot_${snap_nr}/snap_${snap_nr}" \
+    --subfind-basename "${output_group_dir}/subfind_tab_${snap_nr}" \
     --output-basename "${output_dir}/swift_snapshots/swift_${snap_nr}/snap_${snap_nr}" \
     --membership-basename "${output_dir}/SOAP_uncompressed/membership_${snap_nr}/membership_${snap_nr}"
+
+######### Estimate SpeciesFraction of hydrogen
+
+mpirun -- python -u misc/approximate_hydrogen_fractions.py \
+    --snap-basename "${output_dir}/swift_snapshots/swift_${snap_nr}/snap_${snap_nr}" \
+    --output-basename "${output_dir}/species_fractions/swift_${snap_nr}/snap_${snap_nr}"
+
+######### Create virtual snapshot
+# Must be run from the snapshot directory itself or there will be issues with paths
+
+cd "${output_dir}/swift_snapshots/swift_${snap_nr}"
+wget https://gitlab.cosma.dur.ac.uk/swift/swiftsim/-/raw/master/tools/create_virtual_snapshot.py
+python create_virtual_snapshot.py "snap_${snap_nr}.0.hdf5"
+rm create_virtual_snapshot.py*
+cd -
 
 ######### Run SOAP
 
