### 2025-04-29 509a47c
Add call to luminosity-weighted function (undefined  yet)
