### 2025-04-29 b333617
Define luminosity weighted function and pass luminosity arrays.
