### 2025-04-29 7523f1d
Finish implementation of get_projected_inertia_tensor_luminosity_weighted. Directly taken from 3D inertia calculations. Will likely need some restructuring.
