### 2025-04-29 e4890c9
Bug: using 3D positions instead of projected...
