+++ misc/fof_prop.py
@@ -1,0 +1,253 @@
+#!/bin/env python
+
+"""
+TODO: WRITE DOCSTRING
+convert_eagle.py
+
+This script converts EAGLE GADGET snapshots into SWIFT snapshots.
+Usage:
+
+  mpirun -- python convert_eagle.py \
+          --snapshot-basename=SNAPSHOT \
+          --fof-basename=FOF \
+          --output-basename=OUTPUT \
+          --membership-basename=MEMBERSHIP
+
+where SNAPSHOT is the EAGLE snapshot (use the particledata_*** files,
+since the normal snapshots don't store SubGroupNumber), and OUTPUT &
+MEMBERSHIP are the names of the output files. You must run with
+the same number of ranks as input files.
+"""
+
+import argparse
+import os
+
+from mpi4py import MPI
+
+comm = MPI.COMM_WORLD
+comm_rank = comm.Get_rank()
+comm_size = comm.Get_size()
+
+import h5py
+import numpy as np
+import unyt
+
+import virgo.mpi.parallel_sort as psort
+import virgo.mpi.parallel_hdf5 as phdf5
+from virgo.mpi.gather_array import gather_array
+
+# Parse arguments
+parser = argparse.ArgumentParser(
+    description=(
+        "Script to calculate extent of FoF groups."
+    )
+)
+parser.add_argument(
+    "--snap-basename",
+    type=str,
+    required=True,
+    help=(
+        "The basename for the snapshot files (the snapshot "
+        "name without the .{file_nr}.hdf5 suffix)"
+    ),
+)
+parser.add_argument(
+    "--fof-basename",
+    type=str,
+    required=True,
+    help="The basename for the output files",
+)
+parser.add_argument(
+    "--output-basename",
+    type=str,
+    required=True,
+    help=(
+        "The basename for the output files"
+    ),
+)
+parser.add_argument(
+    "--null-fof-id",
+    type=int,
+    required=False,
+    default=2147483647,
+    help=(
+        "The FOFGroupIDs of particles not in a FOF group"
+    ),
+)
+args = parser.parse_args()
+snap_filename = args.snap_basename + ".{file_nr}.hdf5"
+fof_filename = args.fof_basename + ".{file_nr}.hdf5"
+output_filename = args.output_basename + ".{file_nr}.hdf5"
+os.makedirs(os.path.dirname(output_filename), exist_ok=True)
+
+if comm_rank == 0:
+    with h5py.File(snap_filename.format(file_nr=0), "r") as file:
+        header = file['Header'].attrs
+        boxsize = header['BoxSize']
+        ptypes = np.where(header['TotalNumberOfParticles'] != 0)[0]
+        coordinate_unit_attrs = dict(file['PartType1/Coordinates'].attrs)
+
+
+    # TODO: Split across ranks
+    for i_file in range(4):
+        src_filename = fof_filename.format(file_nr=i_file)
+        dst_filename = output_filename.format(file_nr=i_file)
+
+        def copy_attrs(src_obj, dst_obj):
+            for key, val in src_obj.attrs.items():
+                dst_obj.attrs[key] = val
+
+        with h5py.File(src_filename, "r") as src_file, h5py.File(dst_filename, "w") as dst_file:
+
+            def create_virtual_datasets(src_group, dst_group, src_filename, prefix=""):
+                copy_attrs(src_group, dst_group)
+
+                for name, item in src_group.items():
+                    if isinstance(item, h5py.Dataset):
+                        shape = item.shape
+                        dtype = item.dtype
[…]
