### 2025-04-30 612ea29
Handle q = 0 values when testing convergence.
