+++ misc/calculate_fof_positions.py
@@ -35,9 +35,7 @@ from virgo.mpi.gather_array import gather_array
 
 # Parse arguments
 parser = argparse.ArgumentParser(
-    description=(
-        "Script to calculate extent of FoF groups."
-    )
+    description=("Script to calculate extent of FoF groups.")
 )
 parser.add_argument(
     "--snap-basename",
@@ -58,18 +56,14 @@ parser.add_argument(
     "--output-basename",
     type=str,
     required=True,
-    help=(
-        "The basename for the output files"
-    ),
+    help=("The basename for the output files"),
 )
 parser.add_argument(
     "--null-fof-id",
     type=int,
     required=False,
     default=2147483647,
-    help=(
-        "The FOFGroupIDs of particles not in a FOF group"
-    ),
+    help=("The FOFGroupIDs of particles not in a FOF group"),
 )
 args = parser.parse_args()
 snap_filename = args.snap_basename + ".{file_nr}.hdf5"
@@ -79,21 +73,23 @@ os.makedirs(os.path.dirname(output_filename), exist_ok=True)
 
 if comm_rank == 0:
     with h5py.File(snap_filename.format(file_nr=0), "r") as file:
-        header = dict(file['Header'].attrs)
-        coordinate_unit_attrs = dict(file['PartType1/Coordinates'].attrs)
+        header = dict(file["Header"].attrs)
+        coordinate_unit_attrs = dict(file["PartType1/Coordinates"].attrs)
 else:
     header = None
     coordinate_unit_attrs = None
 header = comm.bcast(header)
 coordinate_unit_attrs = comm.bcast(coordinate_unit_attrs)
 
-boxsize = header['BoxSize']
-ptypes = np.where(header['TotalNumberOfParticles'] != 0)[0]
-nr_files = header['NumFilesPerSnapshot'][0]
+boxsize = header["BoxSize"]
+ptypes = np.where(header["TotalNumberOfParticles"] != 0)[0]
+nr_files = header["NumFilesPerSnapshot"][0]
+
 
 def copy_attrs(src_obj, dst_obj):
     for key, val in src_obj.attrs.items():
         dst_obj.attrs[key] = val
+
 
 def copy_object(src_obj, dst_obj, src_filename, prefix="", skip_datasets=False):
     copy_attrs(src_obj, dst_obj)
@@ -110,16 +106,23 @@ def copy_object(src_obj, dst_obj, src_filename, prefix="", skip_datasets=False):
             copy_attrs(item, vds)
         elif isinstance(item, h5py.Group):
             new_group = dst_obj.create_group(name)
-            copy_object(item, new_group, src_filename, prefix + name + "/", skip_datasets=skip_datasets)
+            copy_object(
+                item,
+                new_group,
+                src_filename,
+                prefix + name + "/",
+                skip_datasets=skip_datasets,
+            )
+
 
 # Assign files to ranks
 files_per_rank = np.zeros(comm_size, dtype=int)
 files_per_rank[:] = nr_files // comm_size
 remainder = nr_files % comm_size
 if remainder > 0:
-    step = max(nr_files // (remainder+1), 1)
+    step = max(nr_files // (remainder + 1), 1)
     for i in range(remainder):
-        files_per_rank[(i*step) % comm_size] += 1
+        files_per_rank[(i * step) % comm_size] += 1
 first_file = np.cumsum(files_per_rank) - files_per_rank
 assert sum(files_per_rank) == nr_files
 
@@ -129,7 +132,10 @@ for i_file in range(
 ):
     src_filename = fof_filename.format(file_nr=i_file)
     dst_filename = output_filename.format(file_nr=i_file)
-    with h5py.File(src_filename, "r") as src_file, h5py.File(dst_filename, "w") as dst_file:
+    with (
[…]
