### 2025-04-30 dcb9c61
Handle 0 / 0 division.
