### 2025-04-30 0fb06c1
Make q be an array from the get go
