### 2025-04-30 2d6e3bc
Only compute tensors for non converged bands
