### 2025-04-30 6d09416
Do not rely on disabling warnings to handle q = 0 cases.
