### 2025-04-30 7baea83
Make q = 0 tensors to have all matrix entries to be equal to 0.
