+++ parameter_files/COLIBRE_HYBRID.yml
@@ -35,6 +35,7 @@ ApertureProperties:
     AngularMomentumDarkMatter: general
     AngularMomentumGas: general
     AngularMomentumStars: general
+    AngularMomentumStars_LuminosityWeighted: false
     AtomicHydrogenMass:
       snapshot: true
       snipshot: false
@@ -64,6 +65,8 @@ ApertureProperties:
       snipshot: false
     DiscToTotalGasMassFraction: general
     DiscToTotalStellarMassFraction: general
+    DiscToTotalMassRatio_LuminosityWeighted: false
+    DiscToTotalLuminosityRatio_LuminosityWeighted: false
     DustGraphiteMass:
       snapshot: true
       snipshot: false
@@ -131,6 +134,7 @@ ApertureProperties:
     KappaCorotBaryons: general
     KappaCorotGas: general
     KappaCorotStars: general
+    KappaCorotStars_LuminosityWeighted: false
     KineticEnergyGas: general
     KineticEnergyStars: general
     LinearMassWeightedDiffuseOxygenOverHydrogenOfGas:
@@ -223,18 +227,22 @@ ApertureProperties:
     GasInertiaTensor: false
     DarkMatterInertiaTensor: false
     StellarInertiaTensor: false
+    StellarInertiaTensor_LuminosityWeighted: false
     TotalInertiaTensorReduced: false
     GasInertiaTensorReduced: false
     DarkMatterInertiaTensorReduced: false
     StellarInertiaTensorReduced: false
+    StellarInertiaTensorReduced_LuminosityWeighted: false
     TotalInertiaTensorNoniterative: general
     GasInertiaTensorNoniterative: general
     DarkMatterInertiaTensorNoniterative: general
     StellarInertiaTensorNoniterative: general
+    StellarInertiaTensorNoniterative_LuminosityWeighted: false
     TotalInertiaTensorReducedNoniterative: false
     GasInertiaTensorReducedNoniterative: false
     DarkMatterInertiaTensorReducedNoniterative: false
     StellarInertiaTensorReducedNoniterative: false
+    StellarInertiaTensorReducedNoniterative_LuminosityWeighted: false
     StarFormationRate: true
     AveragedStarFormationRate:
       snapshot: true
@@ -371,9 +379,13 @@ ProjectedApertureProperties:
     ProjectedGasInertiaTensorNoniterative: general
     ProjectedGasInertiaTensorReducedNoniterative: false
     ProjectedStellarInertiaTensor: false
+    ProjectedStellarInertiaTensor_LuminosityWeighted: false
     ProjectedStellarInertiaTensorReduced: false
+    ProjectedStellarInertiaTensorReduced_LuminosityWeighted: false
     ProjectedStellarInertiaTensorNoniterative: general
+    ProjectedStellarInertiaTensorNoniterative_LuminosityWeighted: false
     ProjectedStellarInertiaTensorReducedNoniterative: false
+    ProjectedStellarInertiaTensorReducedNoniterative_LuminosityWeighted: false
     StarFormationRate: true
     AveragedStarFormationRate:
       snapshot: true
@@ -427,6 +439,9 @@ SOProperties:
     DarkMatterMass: true
     DiscToTotalGasMassFraction: general
     DiscToTotalStellarMassFraction: general
+    DiscToTotalLuminosityRatio_LuminosityWeighted: false
+    DiscToTotalMassRatio_LuminosityWeighted: false
+    AngularMomentumStars_LuminosityWeighted: false
     DopplerB: false
     GasCentreOfMass: true
     GasCentreOfMassVelocity: true
@@ -587,6 +602,7 @@ SubhaloProperties:
     AngularMomentumDarkMatter: general
[…]
