### 2025-05-02 889bd46
Make get_inertia_tensor_luminosity_weighted only do computations for non-converged bands More similar to get_projected_inertia_tensor_luminosity_weighted
