### 2025-05-02 03efb2d
Update SOAP to read new values
