### 2025-05-02 1dbca3b
Calculate fof radii (#144)
* Script v1

* Create files across ranks

* Update docs

* Create virtual file

* Use relative paths

* Format

* Add test

* Use absolute positions

* Proper no wrap

* FOF radii

* Update SOAP to read new values
