### 2025-05-02 8a1e9a7
Defined new half light radii property
