+++ SOAP/particle_selection/aperture_properties.py
@@ -141,7 +141,7 @@ import unyt
 
 from .halo_properties import HaloProperty, SearchRadiusTooSmallError
 from SOAP.core.dataset_names import mass_dataset
-from SOAP.property_calculation.half_mass_radius import get_half_mass_radius
+from SOAP.property_calculation.half_mass_radius import get_half_mass_radius, get_half_light_radius
 from SOAP.property_calculation.kinematic_properties import (
     get_velocity_dispersion_matrix,
     get_angular_momentum,
@@ -3175,6 +3175,15 @@ class ApertureParticleData:
         """
         return get_half_mass_radius(
             self.radius[self.type == 4], self.mass_star, self.Mstar
+        )
+
+    @lazy_property
+    def HalfLightRadiusStar(self) -> unyt.unyt_array:
+        """
+        Half light radius of stars for the 9 GAMA bands.
+        """
+        return get_half_light_radius(
+            self.radius[self.type == 4], self.stellar_luminosities, self.StellarLuminosity
         )
 
     @lazy_property
