+++ parameter_files/COLIBRE_THERMAL.yml
@@ -122,6 +122,7 @@ ApertureProperties:
     HalfMassRadiusGas: general
     HalfMassRadiusDust: general
     HalfMassRadiusStars: true
+    HalfLightRadiusStars: false
     HeliumMass: true
     HydrogenMass: true
     KappaCorotBaryons: general
@@ -306,6 +307,7 @@ ProjectedApertureProperties:
     HalfMassRadiusGas: general
     HalfMassRadiusDust: general
     HalfMassRadiusStars: true
+    HalfLightRadiusStars: false
     HeliumMass: true
     HydrogenMass: true
     MolecularHydrogenMass: true
@@ -564,6 +566,7 @@ SubhaloProperties:
     HalfMassRadiusGas: general
     HalfMassRadiusDust: general
     HalfMassRadiusStars: true
+    HalfLightRadiusStars: false
     HalfMassRadiusTotal: true
     EncloseRadius: true
     KappaCorotBaryons: general
