### 2025-05-02 3f5c08c
Copied get_half_mass_radius to start implementing luminosity weighting.
