### 2025-05-02 b4cde48
Create general function to find radius enclosing half of a given quantity
