### 2025-05-02 fb0f663
Defined half light radius.
