+++ SOAP/property_calculation/half_mass_radius.py
@@ -120,83 +120,31 @@ def get_half_mass_radius(
     return get_half_weight_radius(radius, mass, total_mass) 
 
 def get_half_light_radius(
-    radius: unyt.unyt_array, mass: unyt.unyt_array, total_mass: unyt.unyt_quantity
+    radius: unyt.unyt_array, band_luminosity: unyt.unyt_array, total_band_luminosites: unyt.unyt_array
 ) -> unyt.unyt_quantity:
     """
-    Get the half mass radius of the given particle distribution.
+    Get the half light radius of the given particle distribution for the 9 GAMA 
+    bands.
 
-    We obtain the half mass radius by sorting the particles on radius and then computing
-    the cumulative mass profile from this. We then determine in which "bin" the cumulative
-    mass profile intersects the target half mass value and obtain the corresponding
+    We obtain the half light radius by sorting the particles on radius and then computing
+    the cumulative light profile from this. We then determine in which "bin" the cumulative
+    light profile intersects the target half light value and obtain the corresponding
     radius from linear interpolation.
 
     Parameters:
      - radius: unyt.unyt_array
        Radii of the particles.
-     - mass: unyt.unyt_array
-       Mass of the particles.
-     - total_mass: unyt.unyt_quantity
-       Total mass of the particles. Should be mass.sum(). We pass this on as an argument
-       because this value might already have been computed before. If it was not, then
-       computing it in the function call is still an efficient way to do this.
+     - band_luminosity: unyt.unyt_array
+       Luminosity of the particles in each GAMA band.
+     - total_band_luminosites: unyt.unyt_array
+       Total luminosisty of the particles in each GAMA band. Should be luminosity.sum(axis=0). 
+       We pass this on as an argument because this value might already have been computed before. 
+       If it was not, then computing it in the function call is still an efficient way to do this.
 
-    Returns the half mass radius, defined as the radius at which the cumulative mass profile
-    reaches 0.5*total_mass.
+    Returns the half light radius, defined as the radius at which the cumulative mass profile
+    reaches 0.5 * total_luminosity, for each GAMA band.
     """
-    if total_mass == 0.0 * total_mass.units or len(mass) < 1:
-        return 0.0 * radius.units
-
-    target_mass = 0.5 * total_mass
-
-    isort = np.argsort(radius)
-    sorted_radius = radius[isort]
-    # compute sum in double precision to avoid numerical overflow due to
-    # weird unit conversions in unyt
-    cumulative_mass = mass[isort].cumsum(dtype=np.float64)
-
-    # consistency check
-    # np.sum() and np.cumsum() use different orders, so we have to allow for
-    # some small difference
-    if cumulative_mass[-1] < 0.999 * total_mass:
-        raise RuntimeError(
-            "Masses sum up to less than the given total mass:"
-            f" cumulative_mass[-1] = {cumulative_mass[-1]},"
-            f" total_mass = {total_mass}!"
-        )
-
-    # find the intersection point
-    # if that is the first bin, set the lower limits to 0
-    ihalf = np.argmax(cumulative_mass >= target_mass)
-    if ihalf == 0:
-        rmin = 0.0 * radius.units
-        Mmin = 0.0 * mass.units
-    else:
-        rmin = sorted_radius[ihalf - 1]
-        Mmin = cumulative_mass[ihalf - 1]
-    rmax = sorted_radius[ihalf]
-    Mmax = cumulative_mass[ihalf]
-
-    # now get the radius by linearly interpolating
-    # if the bin edges coincide (two particles at exactly the same radius)
[…]
