### 2025-05-06 a046c96
Add half-light radii for each GAMA band. (#145)
* Defined new half light radii property

* Copied get_half_mass_radius to start implementing luminosity weighting.

* Create general function to find radius enclosing half of a given quantity

* Defined half light radius.

* Fix a couple of bugs.
