### 2025-05-07 77cdf39
Make the script load the TrackIds and properties from a text file.
