### 2025-05-07 0677de7
Script to obtain the SOAP property evolution for specified HBT-HERONS subhaloes  (#146)
* Add script

* Fix redshift saving in HDF5

* Update misc README.md

* Make the script load the TrackIds and properties from a text file.

* Fixing typos
