### 2025-05-08 0f7d560
Make 3D inertia tensors general
