+++ SOAP/property_calculation/kinematic_properties.py
@@ -374,9 +374,9 @@ def get_vmax(
     return ordered_radius[imax], np.sqrt(v_over_G[imax] * G)
 
 
-def get_inertia_tensor(
-    mass,
-    position,
+def get_weighted_inertia_tensor(
+    particle_weights,
+    particle_positions,
     sphere_radius,
     search_radius=None,
     reduced=False,
@@ -384,13 +384,14 @@ def get_inertia_tensor(
     min_particles=20,
 ):
     """
-    Get the inertia tensor of the given particle distribution, computed as
-    I_{ij} = m*x_i*x_j / Mtot.
+    Get the inertia tensor of the given particle distribution weighted by a given
+    quantity. Computed as:
+    I_{ij} = w*x_i*x_j / Wtot.
 
     Parameters:
-     - mass: unyt.unyt_array
-       Masses of the particles.
-     - position: unyt.unyt_array
+     - particle_weights: unyt.unyt_array
+       Weight given to each particle.
+     - particle_positions: unyt.unyt_array
        Positions of the particles.
      - sphere_radius: unyt.unyt_quantity
        Use all particles within a sphere of this size for the calculation
@@ -406,20 +407,23 @@ def get_inertia_tensor(
        The number of particles required within the initial sphere. The inertia tensor
        is not computed if this threshold is not met.
 
-    Returns the inertia tensor.
+    Returns a flattened representation of the weighted inertia tensor, with the 
+    first 3 entries corresponding to the diagonal terms and the rest to the
+    off-diagonal terms.
     """
 
     # Check we have at least "min_particles" particles
-    if mass.shape[0] < min_particles:
+    if particle_weights.shape[0] < min_particles:
         return None
 
     # Remove particles at centre if calculating reduced tensor
     if reduced:
-        norm = np.linalg.norm(position, axis=1) ** 2
+        norm = np.linalg.norm(particle_positions, axis=1) ** 2
         mask = np.logical_not(np.isclose(norm, 0))
-        position = position[mask]
-        mass = mass[mask]
+
         norm = norm[mask]
+        particle_weights = particle_weights[mask]
+        particle_positions = particle_positions[mask]
 
     # Set stopping criteria
     tol = 0.0001
@@ -427,9 +431,9 @@ def get_inertia_tensor(
 
     # Ensure we have consistent units
     R = sphere_radius.to("kpc")
-    position = position.to("kpc")
+    particle_positions = particle_positions.to("kpc")
 
-    # Start with a sphere
+    # Start with a sphere of size equal to the initial aperture
     eig_val = [1, 1, 1]
     eig_vec = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 1]])
 
@@ -448,14 +452,15 @@ def get_inertia_tensor(
         axis = R * np.array(
             [1 * np.cbrt(s * p), 1 * np.cbrt(q / p), 1 / np.cbrt(q * s)]
         )
-        p = np.dot(position, eig_vec) / axis
+        p = np.dot(particle_positions, eig_vec) / axis
         r = np.linalg.norm(p, axis=1)
+
         # We want to skip the calculation if we have less than "min_particles"
         # inside the initial sphere. We do the check here since this is the first
         # time we calculate how many particles are within the sphere.
         if (i_iter == 0) and (np.sum(r <= 1) < min_particles):
             return None
-        weight = mass / np.sum(mass[r <= 1])
+        weight = particle_weights / np.sum(particle_weights[r <= 1])
         weight[r > 1] = 0
 
         # Check if we have exceeded the search radius. For subhalo_properties we
@@ -464,18 +469,22 @@ def get_inertia_tensor(
[…]
