+++ SOAP/particle_selection/SO_properties.py
@@ -3673,7 +3673,7 @@ class SOProperties(HaloProperty):
                                     registry=registry,
                                 )
                             else:
-                                err = f'Overflow for halo {input_halo["index"]} when'
+                                err = f'Overflow for halo {input_halo["index"].value} when '
                                 err += f"calculating {name} in SO_properties"
                                 assert np.max(np.abs(val.to(unit).value)) < float(
                                     "inf"
