+++ SOAP/particle_selection/projected_aperture_properties.py
@@ -1848,7 +1848,7 @@ class ProjectedApertureProperties(HaloProperty):
                                     registry=registry,
                                 )
                             else:
-                                err = f'Overflow for halo {input_halo["index"]} when'
+                                err = f'Overflow for halo {input_halo["index"].value} when '
                                 err += f"calculating {name} in projected_properties"
                                 assert np.max(np.abs(val.to(unit).value)) < float(
                                     "inf"
