+++ SOAP/particle_selection/subhalo_properties.py
@@ -2301,7 +2301,7 @@ class SubhaloProperties(HaloProperty):
                             registry=registry,
                         )
                     else:
-                        err = f'Overflow for halo {input_halo["index"]} when'
+                        err = f'Overflow for halo {input_halo["index"].value} when '
                         err += f"calculating {name} in subhalo_properties"
                         assert np.max(np.abs(val.to(unit).value)) < float("inf"), err
                         subhalo[name] += val
