### 2025-05-08 #7851 @warrickball
Hi @robjmcgibbon, I do apologise for the wait. The AASS track currently has a heavy backlog that means editors are only around getting to new articles after a 3-4 month wait. Your suggestions of referees are truly helpful for when an editor comes around, but right now it looks like it'll probably be over a month before someone potentially does.

Rest assured that SOAP is in the stack and not forgotten!
