### 2025-05-09 0f4c880
Make mass-weighted inertia tensor more self-evident
