+++ SOAP/particle_selection/SO_properties.py
@@ -35,7 +35,7 @@ from SOAP.property_calculation.kinematic_properties import (
     get_angular_momentum_and_kappa_corot,
     get_angular_momentum_and_kappa_corot_luminosity_weighted,
     get_vmax,
-    get_inertia_tensor,
+    get_inertia_tensor_mass_weighted,
 )
 from SOAP.particle_filter.recently_heated_gas_filter import RecentlyHeatedGasFilter
 from SOAP.property_table import PropertyTable
@@ -628,7 +628,7 @@ class SOParticleData:
             return None
         mass = np.concatenate([self.mass, self.surrounding_mass], axis=0)
         position = np.concatenate([self.position, self.surrounding_position], axis=0)
-        return get_inertia_tensor(
+        return get_inertia_tensor_mass_weighted(
             mass, position, self.SO_r, search_radius=self.search_radius
         )
 
@@ -643,7 +643,7 @@ class SOParticleData:
             return None
         mass = np.concatenate([self.mass, self.surrounding_mass], axis=0)
         position = np.concatenate([self.position, self.surrounding_position], axis=0)
-        return get_inertia_tensor(
+        return get_inertia_tensor_mass_weighted(
             mass, position, self.SO_r, search_radius=self.search_radius, reduced=True
         )
 
@@ -655,7 +655,7 @@ class SOParticleData:
         """
         if self.Mtotpart == 0:
             return None
-        return get_inertia_tensor(self.mass, self.position, self.SO_r, max_iterations=1)
+        return get_inertia_tensor_mass_weighted(self.mass, self.position, self.SO_r, max_iterations=1)
 
     @lazy_property
     def TotalInertiaTensorReducedNoniterative(self) -> unyt.unyt_array:
@@ -665,7 +665,7 @@ class SOParticleData:
         """
         if self.Mtotpart == 0:
             return None
-        return get_inertia_tensor(
+        return get_inertia_tensor_mass_weighted(
             self.mass, self.position, self.SO_r, reduced=True, max_iterations=1
         )
 
@@ -801,7 +801,7 @@ class SOParticleData:
         surrounding_position = self.surrounding_position[self.surrounding_types == 0]
         mass = np.concatenate([self.gas_masses, surrounding_mass], axis=0)
         position = np.concatenate([self.gas_pos, surrounding_position], axis=0)
-        return get_inertia_tensor(
+        return get_inertia_tensor_mass_weighted(
             mass, position, self.SO_r, search_radius=self.search_radius, **kwargs
         )
 
@@ -835,7 +835,7 @@ class SOParticleData:
         """
         if self.Mgas == 0:
             return None
-        return get_inertia_tensor(
+        return get_inertia_tensor_mass_weighted(
             self.gas_masses, self.gas_pos, self.SO_r, max_iterations=1
         )
 
@@ -847,7 +847,7 @@ class SOParticleData:
         """
         if self.Mgas == 0:
             return None
-        return get_inertia_tensor(
+        return get_inertia_tensor_mass_weighted(
             self.gas_masses, self.gas_pos, self.SO_r, reduced=True, max_iterations=1
         )
 
@@ -918,7 +918,7 @@ class SOParticleData:
         surrounding_position = self.surrounding_position[self.surrounding_types == 1]
         mass = np.concatenate([self.dm_masses, surrounding_mass], axis=0)
         position = np.concatenate([self.dm_pos, surrounding_position], axis=0)
-        return get_inertia_tensor(
+        return get_inertia_tensor_mass_weighted(
             mass, position, self.SO_r, search_radius=self.search_radius, **kwargs
         )
 
@@ -952,7 +952,7 @@ class SOParticleData:
         """
         if self.Mdm == 0:
[…]
