### 2025-05-09 713252f
Make projected inertia tensors general
