+++ SOAP/property_calculation/kinematic_properties.py
@@ -587,8 +587,8 @@ def get_inertia_tensor_luminosity_weighted(
 
     return flattened_inertia_tensors
 
-def get_projected_inertia_tensor(
-    mass, position, axis, radius, reduced=False, max_iterations=20, min_particles=20
+def get_weighted_projected_inertia_tensor(
+    particle_weights, particle_positions, axis, radius, reduced=False, max_iterations=20, min_particles=20
 ):
     """
     Takes in the particle distribution projected along a given axis, and calculates the inertia
@@ -598,9 +598,9 @@ def get_projected_inertia_tensor(
     is because all the bound particles are passed to this function.
 
     Parameters:
-     - mass: unyt.unyt_array
-       Masses of the particles.
-     - position: unyt.unyt_array
+     - particle_weights: unyt.unyt_array
+       Weight given to each particle.
+     - particle_positions: unyt.unyt_array
        Positions of the particles.
      - axis: 0, 1, 2
        Projection axis. Only the coordinates perpendicular to this axis are
@@ -619,21 +619,21 @@ def get_projected_inertia_tensor(
     """
 
     # Check we have at least "min_particles" particles
-    if mass.shape[0] < min_particles:
+    if particle_weights.shape[0] < min_particles:
         return None
 
     projected_position = unyt.unyt_array(
-        np.zeros((position.shape[0], 2)), units=position.units, dtype=position.dtype
+        np.zeros((particle_positions.shape[0], 2)), units=particle_positions.units, dtype=particle_positions.dtype
     )
     if axis == 0:
-        projected_position[:, 0] = position[:, 1]
-        projected_position[:, 1] = position[:, 2]
+        projected_position[:, 0] = particle_positions[:, 1]
+        projected_position[:, 1] = particle_positions[:, 2]
     elif axis == 1:
-        projected_position[:, 0] = position[:, 2]
-        projected_position[:, 1] = position[:, 0]
+        projected_position[:, 0] = particle_positions[:, 2]
+        projected_position[:, 1] = particle_positions[:, 0]
     elif axis == 2:
-        projected_position[:, 0] = position[:, 0]
-        projected_position[:, 1] = position[:, 1]
+        projected_position[:, 0] = particle_positions[:, 0]
+        projected_position[:, 1] = particle_positions[:, 1]
     else:
         raise AttributeError(f"Invalid axis: {axis}!")
 
@@ -641,9 +641,10 @@ def get_projected_inertia_tensor(
     if reduced:
         norm = np.linalg.norm(projected_position, axis=1) ** 2
         mask = np.logical_not(np.isclose(norm, 0))
-        projected_position = projected_position[mask]
-        mass = mass[mask]
+
         norm = norm[mask]
+        particle_weights = particle_weights[mask]
+        projected_position = projected_position[mask]
 
     # Set stopping criteria
     tol = 0.0001
@@ -653,7 +654,7 @@ def get_projected_inertia_tensor(
     R = radius.to("kpc")
     projected_position = projected_position.to("kpc")
 
-    # Start with a circle
+    # Start with a circle of size equal to the initial aperture
     eig_val = [1, 1]
     eig_vec = np.array([[1, 0], [0, 1]])
 
@@ -670,12 +671,13 @@ def get_projected_inertia_tensor(
         axis = R * np.array([1 * np.sqrt(q), 1 / np.sqrt(q)])
         p = np.dot(projected_position, eig_vec) / axis
         r = np.linalg.norm(p, axis=1)
+
         # We want to skip the calculation if we have less than "min_particles"
         # inside the initial circle. We do the check here since this is the first
         # time we calculate how many particles are within the circle.
[…]
