### 2025-05-13 3d0fdf0
Allow for property defined apertures
