+++ SOAP/particle_selection/aperture_properties.py
@@ -3351,7 +3351,8 @@ class ApertureProperties(HaloProperty):
         self,
         cellgrid: SWIFTCellGrid,
         parameters: ParameterFile,
-        physical_radius_kpc: float,
+        aperture_physical_radius_kpc: float | None,
+        aperture_property: tuple[str, float] | None,
         recently_heated_gas_filter: RecentlyHeatedGasFilter,
         stellar_age_calculator: StellarAgeCalculator,
         cold_dense_gas_filter: ColdDenseGasFilter,
@@ -3372,9 +3373,17 @@ class ApertureProperties(HaloProperty):
          - parameters: ParameterFile
            Parameter file object containing the parameters from the parameter
            file.
-         - physical_radius_kpc: float
+         - aperture_physical_radius_kpc: float | None
            Physical radius of the aperture. Unitless and assumed to be expressed
-           in units of kpc.
+           in units of kpc. If not None then aperture_property must be None.
+           If None then aperture_property must be passed.
+         - aperture_property: tuple[str, float] | None,
+           Tuple to indicate the radius of this aperture based on a previous property
+           calculated by SOAP. The first element should be the full name of the
+           property to use (e.g. BoundSubhalo/HalfMassRadius). The second element
+           is a multipler (e.g. if you want the aperture radius to be twice the
+           value of the property, then pass 2). If not None then aperture_physical_radius_kpc
+           must be None. If None then aperture_physical_radius_kpc must be passed.
          - recently_heated_gas_filter: RecentlyHeatedGasFilter
            Filter used to mask out gas particles that were recently heated by
            AGN feedback.
@@ -3395,7 +3404,7 @@ class ApertureProperties(HaloProperty):
            to the subhalo?
          - all_radii_kpc: list
            A list of all the radii for which we are computing an ApertureProperties.
-           This can allow us to skip property calculation for larger apertures
+           This can allow us to skip property calculation for larger apertures.
         """
 
         super().__init__(cellgrid)
@@ -3414,18 +3423,30 @@ class ApertureProperties(HaloProperty):
         self.all_radii_kpc = all_radii_kpc
         self.strict_halo_copy = parameters.strict_halo_copy()
         self.boxsize = cellgrid.boxsize
-
-        # Minimum physical radius to read in (pMpc)
-        self.physical_radius_mpc = 0.001 * physical_radius_kpc
-
+        self.aperture_physical_radius_kpc = aperture_physical_radius_kpc
+        self.aperture_property = aperture_property
         self.inclusive = inclusive
 
-        if self.inclusive:
-            self.name = f"inclusive_sphere_{physical_radius_kpc:.0f}kpc"
-            self.group_name = f"InclusiveSphere/{self.physical_radius_mpc*1000.:.0f}kpc"
+        if self.aperture_physical_radius_kpc is not None:
+            aperture_name = f'{self.aperture_physical_radius_kpc:.0f}kpc'
+            self.physical_radius_mpc = 0.001 * self.aperture_physical_radius_kpc
         else:
-            self.name = f"exclusive_sphere_{physical_radius_kpc:.0f}kpc"
-            self.group_name = f"ExclusiveSphere/{self.physical_radius_mpc*1000.:.0f}kpc"
+            prop = self.aperture_property[0].split('/')[-1]
+            multiplier = self.aperture_property[1]
+            if multiplier == 1:
+                aperture_name = prop
+            else:
+                aperture_name = f'{multiplier}x{prop}'
+            # This value needs to be set since it's used to guess the initial
+            # load region for each particle
+            self.physical_radius_mpc = 0
+
+        if self.inclusive:
+            self.name = f"inclusive_sphere_{aperture_name}"
+            self.group_name = f"InclusiveSphere/{aperture_name}"
+        else:
[…]
