### 2025-05-13 fdfa327
Two bugs: incorrect function call and units in reduced tensors
