### 2025-05-13 1fcc5e0
Move inertia tensor functions into their own file
