+++ SOAP/property_calculation/kinematic_properties.py
@@ -13,9 +13,6 @@ import numpy as np
 from typing import Union, Tuple
 import unyt
 
-from SOAP.particle_selection.halo_properties import SearchRadiusTooSmallError
-
-
 def get_velocity_dispersion_matrix(
     mass_fraction: unyt.unyt_array,
     velocity: unyt.unyt_array,
@@ -53,7 +50,6 @@ def get_velocity_dispersion_matrix(
     result[5] += (mass_fraction * vrel[:, 1] * vrel[:, 2]).sum()
 
     return result
-
 
 def get_angular_momentum(
     mass: unyt.unyt_array,
@@ -97,7 +93,6 @@ def get_angular_momentum(
     else:
         vrel = velocity - ref_velocity[None, :]
     return (mass[:, None] * np.cross(prel, vrel)).sum(axis=0)
-
 
 def get_angular_momentum_and_kappa_corot(
     mass: unyt.unyt_array,
@@ -193,7 +188,6 @@ def get_angular_momentum_and_kappa_corot(
         return Ltot, kappa_corot, M_counterrot
     else:
         return Ltot, kappa_corot
-
 
 def get_angular_momentum_and_kappa_corot_luminosity_weighted(
     mass: unyt.unyt_array,
@@ -373,438 +367,8 @@ def get_vmax(
     imax = np.argmax(v_over_G)
     return ordered_radius[imax], np.sqrt(v_over_G[imax] * G)
 
-
-def get_weighted_inertia_tensor(
-    particle_weights,
-    particle_positions,
-    sphere_radius,
-    search_radius=None,
-    reduced=False,
-    max_iterations=20,
-    min_particles=20,
-):
-    """
-    Get the inertia tensor of the given particle distribution weighted by a given
-    quantity. Computed as:
-    I_{ij} = w*x_i*x_j / Wtot.
-
-    Parameters:
-     - particle_weights: unyt.unyt_array
-       Weight given to each particle.
-     - particle_positions: unyt.unyt_array
-       Positions of the particles.
-     - sphere_radius: unyt.unyt_quantity
-       Use all particles within a sphere of this size for the calculation
-     - search_radius: unyt.unyt_quantity
-       Radius of the region of the simulation for which we have particle data
-       This function throws a SearchRadiusTooSmallError if we need particles outside
-       of this region.
-     - reduced: bool
-       Whether to calculate the reduced inertia tensor
-     - max_iterations: int
-       The maximum number of iterations to repeat the inertia tensor calculation
-     - min_particles: int
-       The number of particles required within the initial sphere. The inertia tensor
-       is not computed if this threshold is not met.
-
-    Returns a flattened representation of the weighted inertia tensor, with the 
-    first 3 entries corresponding to the diagonal terms and the rest to the
-    off-diagonal terms.
-    """
-
-    # Check we have at least "min_particles" particles
-    if particle_weights.shape[0] < min_particles:
-        return None
-
-    # Remove particles at centre if calculating reduced tensor
-    if reduced:
-        norm = np.linalg.norm(particle_positions, axis=1) ** 2
-        mask = np.logical_not(np.isclose(norm, 0))
-
-        norm = norm[mask]
-        particle_weights = particle_weights[mask]
-        particle_positions = particle_positions[mask]
-
-    # Set stopping criteria
-    tol = 0.0001
-    q = 1000
-
-    # Ensure we have consistent units
-    R = sphere_radius.to("kpc")
-    particle_positions = particle_positions.to("kpc")
-
-    # Start with a sphere of size equal to the initial aperture
-    eig_val = [1, 1, 1]
[…]
