+++ SOAP/property_calculation/inertia_tensors.py
@@ -1,0 +1,451 @@
+#! /usr/bin/env python
+
+"""
+inertia_tensors.py
+
+Some utility functions to compute inertia tensors for particle spatial 
+distributions.
+
+We put them in a separate file to facilitate unit testing.
+"""
+
+import numpy as np
+from typing import Union, Tuple
+import unyt
+
+from SOAP.particle_selection.halo_properties import SearchRadiusTooSmallError
+
+def get_weighted_inertia_tensor(
+    particle_weights,
+    particle_positions,
+    sphere_radius,
+    search_radius=None,
+    reduced=False,
+    max_iterations=20,
+    min_particles=20,
+):
+    """
+    Get the inertia tensor of the given particle distribution weighted by a given
+    quantity. Computed as:
+    I_{ij} = w*x_i*x_j / Wtot.
+
+    Parameters:
+     - particle_weights: unyt.unyt_array
+       Weight given to each particle.
+     - particle_positions: unyt.unyt_array
+       Positions of the particles.
+     - sphere_radius: unyt.unyt_quantity
+       Use all particles within a sphere of this size for the calculation
+     - search_radius: unyt.unyt_quantity
+       Radius of the region of the simulation for which we have particle data
+       This function throws a SearchRadiusTooSmallError if we need particles outside
+       of this region.
+     - reduced: bool
+       Whether to calculate the reduced inertia tensor
+     - max_iterations: int
+       The maximum number of iterations to repeat the inertia tensor calculation
+     - min_particles: int
+       The number of particles required within the initial sphere. The inertia tensor
+       is not computed if this threshold is not met.
+
+    Returns a flattened representation of the weighted inertia tensor, with the 
+    first 3 entries corresponding to the diagonal terms and the rest to the
+    off-diagonal terms.
+    """
+
+    # Check we have at least "min_particles" particles
+    if particle_weights.shape[0] < min_particles:
+        return None
+
+    # Remove particles at centre if calculating reduced tensor
+    if reduced:
+        norm = np.linalg.norm(particle_positions, axis=1) ** 2
+        mask = np.logical_not(np.isclose(norm, 0))
+
+        norm = norm[mask]
+        particle_weights = particle_weights[mask]
+        particle_positions = particle_positions[mask]
+
+    # Set stopping criteria
+    tol = 0.0001
+    q = 1000
+
+    # Ensure we have consistent units
+    R = sphere_radius.to("kpc")
+    particle_positions = particle_positions.to("kpc")
+
+    # Start with a sphere of size equal to the initial aperture
+    eig_val = [1, 1, 1]
+    eig_vec = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 1]])
+
+    for i_iter in range(max_iterations):
+        # Calculate shape
+        old_q = q
+        q = np.sqrt(eig_val[1] / eig_val[2])
+        s = np.sqrt(eig_val[0] / eig_val[2])
+        p = np.sqrt(eig_val[0] / eig_val[1])
+
+        # Break if converged
+        if abs((old_q - q) / q) < tol:
+            break
+
+        # Calculate ellipsoid, determine which particles are inside
+        axis = R * np.array(
+            [1 * np.cbrt(s * p), 1 * np.cbrt(q / p), 1 / np.cbrt(q * s)]
+        )
+        p = np.dot(particle_positions, eig_vec) / axis
+        r = np.linalg.norm(p, axis=1)
+
+        # We want to skip the calculation if we have less than "min_particles"
+        # inside the initial sphere. We do the check here since this is the first
+        # time we calculate how many particles are within the sphere.
+        if (i_iter == 0) and (np.sum(r <= 1) < min_particles):
+            return None
[…]
