### 2025-05-13 022cac4
Add error messages if required radii are disabled
