### 2025-05-13 5bfd9c9
Fix unyt array initialisation to specify unit registry
