### 2025-05-13 21dff94
Generalise kappa_corot
