+++ SOAP/property_calculation/kinematic_properties.py
@@ -94,108 +94,191 @@ def get_angular_momentum(
         vrel = velocity - ref_velocity[None, :]
     return (mass[:, None] * np.cross(prel, vrel)).sum(axis=0)
 
-def get_angular_momentum_and_kappa_corot(
-    mass: unyt.unyt_array,
-    position: unyt.unyt_array,
-    velocity: unyt.unyt_array,
-    ref_position: Union[None, unyt.unyt_array] = None,
-    ref_velocity: Union[None, unyt.unyt_array] = None,
+def get_angular_momentum_and_kappa_corot_weighted(
+    particle_masses: unyt.unyt_array,
+    particle_positions: unyt.unyt_array,
+    particle_velocities: unyt.unyt_array,
+    particle_weights: Union[None, unyt.unyt_array] = None,
+    reference_position: Union[None, unyt.unyt_array] = None,
+    reference_velocity: Union[None, unyt.unyt_array] = None,
+    do_counterrot_mass: bool = False,
+    do_counterrot_weight: bool = False,
+) -> Union[
+    Tuple[unyt.unyt_array, unyt.unyt_array],
+    Tuple[unyt.unyt_array, unyt.unyt_array, unyt.unyt_array],
+    Tuple[unyt.unyt_array, unyt.unyt_array, unyt.unyt_array, unyt.unyt_array],
+]:
+    """
+    Get the total angular momentum vector and kappa_corot (Correa et al., 2017) 
+    for the particles with the given masses, positions, velocities and weights.
+    It uses the given reference position and velocity as the spatial and velocity 
+    centres. It optionally returns the total mass and total weight value of
+    counterrotating particles. 
+
+    We use this function if both kappa_corot and the angular momentum vector are 
+    requested, as it is more efficient than get_angular_momentum() and 
+    get_kappa_corot() (if that would even exist).
+
+    Parameters:
+     - particle_masses: unyt.unyt_array
+       Masses of the particles.
+     - particle_positions: unyt.unyt_array
+       Position of the particles.
+     - particle_velocities: unyt.unyt_array
+       Velocities of the particles.
+     - particle_weights: unyt.unyt_array or None
+       Weights given to each particle when computing the total angular momentum. 
+       If not provided, this function returns the mass-weighted inertia tensor.
+     - reference_position: unyt.unyt_array or None
+       Reference position used as centre for the angular momentum calculation.
+       particle_positions and reference_position are assumed to use the same reference point upon
+       entry into this function. If None, particle_positions is assumed to be already using
+       the desired referece point.
+     - reference_velocity: unyt.unyt_array or None
+       Reference point in velocity space for the angular momentum calculation.
+       particle_velocities and reference_velocity are assumed to use the same reference point upon
+       entry into this function. If None, particle_velocities is assumed to be already using
+       the desired reference point.
+     - do_counterrot_mass: bool
+       Also compute the counterrotating mass?
+     - do_counterrot_weight: bool
+       Also compute the counterrotating weight quantity?
+
+    Returns:
+     - The weighted total angular momentum vector, or if no weights are provided,
+       the usual definition of angular momentum.
+     - The ratio of the kinetic energy in counterrotating movement and the total
+       kinetic energy, kappa_corot.
+     - The total mass of counterrotating particles if do_counterrot_mass == True.
+     - The total weight of counterrotating particles if do_counterrot_weight == True.
+    """
+
+    if reference_position is None:
+        prel = particle_positions
+    else:
+        prel = particle_positions - reference_position[None, :]
+    if reference_velocity is None:
+        vrel = particle_velocities
[…]
