### 2025-05-13 1859e9e
Shorten the description of helper functions
