### 2025-05-13 62dd976
Skip assert check for ranks with no subhalos (#148)
