### 2025-05-15 96a2992
(Untested) implementation of Vrot calculation
