+++ SOAP/property_calculation/kinematic_properties.py
@@ -13,6 +13,7 @@ import numpy as np
 from typing import Union, Tuple
 import unyt
 
+
 def get_velocity_dispersion_matrix(
     mass_fraction: unyt.unyt_array,
     velocity: unyt.unyt_array,
@@ -50,6 +51,7 @@ def get_velocity_dispersion_matrix(
     result[5] += (mass_fraction * vrel[:, 1] * vrel[:, 2]).sum()
 
     return result
+
 
 def get_angular_momentum(
     mass: unyt.unyt_array,
@@ -94,6 +96,7 @@ def get_angular_momentum(
         vrel = velocity - ref_velocity[None, :]
     return (mass[:, None] * np.cross(prel, vrel)).sum(axis=0)
 
+
 def get_angular_momentum_and_kappa_corot_weighted(
     particle_masses: unyt.unyt_array,
     particle_positions: unyt.unyt_array,
@@ -109,14 +112,14 @@ def get_angular_momentum_and_kappa_corot_weighted(
     Tuple[unyt.unyt_array, unyt.unyt_array, unyt.unyt_array, unyt.unyt_array],
 ]:
     """
-    Get the total angular momentum vector and kappa_corot (Correa et al., 2017) 
+    Get the total angular momentum vector and kappa_corot (Correa et al., 2017)
     for the particles with the given masses, positions, velocities and weights.
-    It uses the given reference position and velocity as the spatial and velocity 
+    It uses the given reference position and velocity as the spatial and velocity
     centres. It optionally returns the total mass and total weight value of
-    counterrotating particles. 
+    counterrotating particles.
 
-    We use this function if both kappa_corot and the angular momentum vector are 
-    requested, as it is more efficient than get_angular_momentum() and 
+    We use this function if both kappa_corot and the angular momentum vector are
+    requested, as it is more efficient than get_angular_momentum() and
     get_kappa_corot() (if that would even exist).
 
     Parameters:
@@ -127,7 +130,7 @@ def get_angular_momentum_and_kappa_corot_weighted(
      - particle_velocities: unyt.unyt_array
        Velocities of the particles.
      - particle_weights: unyt.unyt_array or None
-       Weights given to each particle when computing the total angular momentum. 
+       Weights given to each particle when computing the total angular momentum.
        If not provided, this function returns the mass-weighted inertia tensor.
      - reference_position: unyt.unyt_array or None
        Reference position used as centre for the angular momentum calculation.
@@ -162,23 +165,25 @@ def get_angular_momentum_and_kappa_corot_weighted(
     else:
         vrel = particle_velocities - reference_velocity[None, :]
 
-    # We compute the normal angular momentum because we require it for the 
+    # We compute the normal angular momentum because we require it for the
     # kinetic energy calculation in kappa_corot.
     Lpart = particle_masses[:, None] * np.cross(prel, vrel)
 
-    # If we weight the angular momentum, divide by the particle mass so that 
+    # If we weight the angular momentum, divide by the particle mass so that
     # it L_i = w_i * (r_i x v_i). We also estimate a total weighted mass in order
     # to return an angular momentum with the correct units.
-    if (particle_weights is not None):
+    if particle_weights is not None:
         weights = particle_weights / particle_weights.sum() / particle_masses
-        weighted_total_mass = (particle_weights / particle_weights.sum() * particle_masses).sum() * len(particle_masses)
+        weighted_total_mass = (
+            particle_weights / particle_weights.sum() * particle_masses
+        ).sum() * len(particle_masses)
 
     # Normal (mass-weighted) definition of the angular momentum.
     else:
         weights = np.ones(1)
         weighted_total_mass = 1
 
-    # Weighted average of the (specific) angular momentum. We multiply by the 
+    # Weighted average of the (specific) angular momentum. We multiply by the
[…]
