### 2025-05-15 7ddb5f6
Luminosity-weighted angular momenta and inertia tensors (#135)
* Add shorthand to obtain grik luminosities.

* Add basic functions to measure disc to total luminosity in different bands.

* Make new properties not be computed by default

* Add call of unimplemented function that computes light weighted L quantities.

* Define function for luminosity weighted L quantities. Not implemented for now.

* Add docs on parameters for get_angular_momentum_and_kappa_corot_luminosity_weighted

* Make kappa_corot and counterotating mass give a value per GAMMA band

* Add counterrotating luminosity array


[…]
