+++ SOAP/particle_selection/projected_aperture_properties.py
@@ -33,9 +33,15 @@ from SOAP.core.category_filter import CategoryFilter
 from SOAP.core.parameter_file import ParameterFile
 from SOAP.core.snapshot_datasets import SnapshotDatasets
 from SOAP.core.dataset_names import mass_dataset
-from SOAP.property_calculation.half_mass_radius import get_half_mass_radius, get_half_light_radius
+from SOAP.property_calculation.half_mass_radius import (
+    get_half_mass_radius,
+    get_half_light_radius,
+)
 from SOAP.property_table import PropertyTable
-from SOAP.property_calculation.kinematic_properties import get_projected_inertia_tensor
+from SOAP.property_calculation.inertia_tensors import (
+    get_projected_inertia_tensor_mass_weighted,
+    get_projected_inertia_tensor_luminosity_weighted,
+)
 
 
 class ProjectedApertureParticleData:
@@ -788,7 +794,7 @@ class SingleProjectionProjectedApertureParticleData:
         """
         if self.Mtot == 0:
             return None
-        return get_projected_inertia_tensor(
+        return get_projected_inertia_tensor_mass_weighted(
             self.part_props.mass,
             self.part_props.position,
             self.iproj,
@@ -804,7 +810,7 @@ class SingleProjectionProjectedApertureParticleData:
         """
         if self.Mtot == 0:
             return None
-        return get_projected_inertia_tensor(
+        return get_projected_inertia_tensor_mass_weighted(
             self.part_props.mass,
             self.part_props.position,
             self.iproj,
@@ -820,7 +826,7 @@ class SingleProjectionProjectedApertureParticleData:
         """
         if self.Mtot == 0:
             return None
-        return get_projected_inertia_tensor(
+        return get_projected_inertia_tensor_mass_weighted(
             self.proj_mass,
             self.proj_position,
             self.iproj,
@@ -836,7 +842,7 @@ class SingleProjectionProjectedApertureParticleData:
         """
         if self.Mtot == 0:
             return None
-        return get_projected_inertia_tensor(
+        return get_projected_inertia_tensor_mass_weighted(
             self.proj_mass,
             self.proj_position,
             self.iproj,
@@ -874,7 +880,7 @@ class SingleProjectionProjectedApertureParticleData:
         """
         mass = self.part_props.mass[self.part_props.types == 0]
         position = self.part_props.position[self.part_props.types == 0]
-        return get_projected_inertia_tensor(
+        return get_projected_inertia_tensor_mass_weighted(
             mass, position, self.iproj, self.aperture_radius, **kwargs
         )
 
@@ -908,7 +914,7 @@ class SingleProjectionProjectedApertureParticleData:
         """
         if self.Mgas == 0:
             return None
-        return get_projected_inertia_tensor(
+        return get_projected_inertia_tensor_mass_weighted(
             self.proj_mass_gas,
             self.proj_pos_gas,
             self.iproj,
@@ -924,7 +930,7 @@ class SingleProjectionProjectedApertureParticleData:
         """
         if self.Mgas == 0:
             return None
-        return get_projected_inertia_tensor(
+        return get_projected_inertia_tensor_mass_weighted(
             self.proj_mass_gas,
             self.proj_pos_gas,
             self.iproj,
@@ -985,8 +991,26 @@ class SingleProjectionProjectedApertureParticleData:
         """
         mass = self.part_props.mass[self.part_props.types == 4]
         position = self.part_props.position[self.part_props.types == 4]
-        return get_projected_inertia_tensor(
+        return get_projected_inertia_tensor_mass_weighted(
             mass, position, self.iproj, self.aperture_radius, **kwargs
+        )
+
+    def stellar_inertia_tensor_luminosity_weighted(self, **kwargs) -> unyt.unyt_array:
+        """
[…]
