+++ SOAP/property_table.py
@@ -839,6 +839,23 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=0,
         ),
+        "DtoTstar_luminosity_weighted_luminosity_ratio": Property(
+            name="DiscToTotalLuminosityRatio_LuminosityWeighted",
+            shape=9,
+            dtype=np.float32,
+            unit="dimensionless",
+            description="Fraction of the total stellar luminosity for each band that is in the disc. The band uses its own self-defined luminosity-weighted angular momentum",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType4/Coordinates",
+                "PartType4/Velocities",
+                "PartType4/Masses",
+                "PartType4/Luminosities",
+            ],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
         "DustMass": Property(
             name="DustMass",
             shape=1,
@@ -853,6 +870,57 @@ class PropertyTable:
             ],
             output_physical=True,
             a_scale_exponent=0,
+        ),
+        "DtoTstar_luminosity_weighted_mass_ratio": Property(
+            name="DiscToTotalMassRatio_LuminosityWeighted",
+            shape=9,
+            dtype=np.float32,
+            unit="dimensionless",
+            description="Fraction of the total stellar mass that is in the disc. Each band uses its own self-defined luminosity-weighted angular momentum",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType4/Coordinates",
+                "PartType4/Velocities",
+                "PartType4/Masses",
+                "PartType4/Luminosities",
+            ],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
+        "kappa_corot_star_luminosity_weighted": Property(
+            name="KappaCorotStars_LuminosityWeighted",
+            shape=9,
+            dtype=np.float32,
+            unit="dimensionless",
+            description="Kappa-corot for stars, relative to the centre of potential and the centre of mass velocity of the stars. A value is output for each of the saved luminosity bands, as each is used to define the luminosity-weighted angular momentum.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType4/Coordinates",
+                "PartType4/Masses",
+                "PartType4/Velocities",
+                "PartType4/Luminosities",
+            ],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
+        "Lstar_luminosity_weighted": Property(
+            name="AngularMomentumStars_LuminosityWeighted",
+            shape=27,  # 3D vector for each of the 9 GAMA bands
+            dtype=np.float32,
+            unit="snap_mass*snap_length**2/snap_time",
+            description="Luminosity-weighted total angular momentum of the stars, relative to the centre of potential and stellar centre of mass velocity. A different vector is computed for each of the saved luminosity bands.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType4/Coordinates",
+                "PartType4/Masses",
+                "PartType4/Velocities",
+                "PartType4/Luminosities",
+            ],
+            output_physical=True,
+            a_scale_exponent=1,
         ),
         "DustGraphiteMass": Property(
             name="DustGraphiteMass",
@@ -1241,7 +1309,11 @@ class PropertyTable:
             description="Stellar half light radius in the 9 GAMA bands.",
             lossy_compression_filter="FMantissa9",
             dmo_property=False,
[…]
