+++ documentation/footnote_tensor.tex
@@ -4,7 +4,7 @@
     I_{ij} = \frac{1}{\sum_k m_k} \sum_k m_k \; r_{k,i} \; r_{k, j} 
 \end{equation}
 
-where the index $k$ loops over all particles, $m_k$ is the mass of particle $k$, and $r_{k, i}$ is the $i$-component of the position vector of particle $k$ relative to the halo centre. We first compute the inertia tensor using all particles within a sphere (with radius equal to the aperture size, except for subhalos where we use the half mass radius of the particles). This is the tensor we output in the non-iterative case. In the iterative case we construct an ellipsoid with a volume equal to the initial sphere, but whose shape is given by the inertia tensor. We then recalculate the inertia tensor using only the particles within the ellipsoid. This process is repeated until the value of the $q$ parameter converges, or we reach 20 iterations. For projected apertures the process is similar, except we use circles and ellipses in the projected plane to determine which particles to include.
+where the index $k$ loops over all particles, $m_k$ is the mass of particle $k$, and $r_{k, i}$ is the $i$-component of the position vector of particle $k$ relative to the halo centre. We first compute the inertia tensor using all particles within a sphere (with radius equal to the aperture size, except for subhalos where we use the half mass radius of the particles). This is the tensor we output in the non-iterative case. In the iterative case we construct an ellipsoid with a volume equal to the initial sphere, but whose shape is given by the inertia tensor. We then recalculate the inertia tensor using only the particles within the ellipsoid. This process is repeated until the value of the $q$ parameter converges, or we reach 20 iterations. If at any point during the iterations there is only a single particle within the ellipsoid, we return zero. For projected apertures the process is similar, except we use circles and ellipses in the projected plane to determine which particles to include.
 
 The reduced inertia tensor is calculated as
 

[…]
