### 2025-05-17 4e8d14f
Full SOAP EAGLE catalogues (#150)
* Add more EAGLE snapshot properties

* Add script to approximate H fractions

* Use proper index for SOAP EAGLE

* Update scripts

* Squash bugs
