+++ misc/convert_eagle.py
@@ -83,6 +83,14 @@ parser.add_argument(
     ),
 )
 parser.add_argument(
+    "--subfind-basename",
+    type=str,
+    required=True,
+    help=(
+        "The basename for the subfind files"
+    ),
+)
+parser.add_argument(
     "--output-basename",
     type=str,
     required=True,
@@ -96,6 +104,7 @@ parser.add_argument(
 )
 args = parser.parse_args()
 snap_filename = args.snap_basename + ".{file_nr}.hdf5"
+subfind_filename = args.subfind_basename + ".{file_nr}.hdf5"
 output_filename = args.output_basename + ".{file_nr}.hdf5"
 membership_filename = args.membership_basename + ".{file_nr}.hdf5"
 os.makedirs(os.path.dirname(output_filename), exist_ok=True)
@@ -182,6 +191,10 @@ if comm_rank == 0:
             "Gravity:max_physical_baryon_softening": runtime.attrs[
                 "SofteningGasMaxPhys"
             ],
+            "EAGLE:InitAbundance_Hydrogen": runtime.attrs['InitAbundance_Hydrogen'],
+            "EAGLE:InitAbundance_Helium": runtime.attrs['InitAbundance_Helium'],
+            "EAGLE:EOS_Jeans_GammaEffective": runtime.attrs['EOS_Jeans_GammaEffective'],
+            "EAGLE:EOS_Jeans_TempNorm_K": runtime.attrs['EOS_Jeans_TempNorm_K'],
         }
 
         # Check units are indeed what we are assuming below, since HO,
@@ -244,8 +257,8 @@ if comm_rank == 0:
 #                         should be h-free
 #
 # In the case of EAGLE snapshots most of this information is available
-# in the metadata of the snapshots themselves. We therefore leave those
-# values blank in the properties dictionary, and read them in directly.
+# in the metadata of the snapshots themselves. We therefore set those
+# values as None in the properties dictionary, and read them in directly.
 properties = {
     "PartType0": {
         "Coordinates": {
@@ -283,6 +296,58 @@ properties = {
             "description": None,
             "conversion_factor": None,
         },
+        "Density": {
+            "swift_name": "Densities",
+            "exponents": {"L": -3, "M": 1, "T": 0, "t": 0},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "InternalEnergy": {
+            "swift_name": "InternalEnergies",
+            "exponents": {"L": 2, "M": 0, "T": 0, "t": -2},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "IronMassFracFromSNIa": {
+            "swift_name": "IronMassFractionsFromSNIa",
+            "exponents": {"L": 0, "M": 0, "T": 0, "t": 0},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "Metallicity": {
+            "swift_name": "MetalMassFractions",
+            "exponents": {"L": 0, "M": 0, "T": 0, "t": 0},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "StarFormationRate": {
+            "swift_name": "StarFormationRates",
+            "exponents": {"L": 0, "M": 1, "T": 0, "t": -1},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        "Temperature": {
+            "swift_name": "Temperatures",
+            "exponents": {"L": 0, "M": 0, "T": 1, "t": 0},
+            "a_exponent": None,
+            "description": None,
+            "conversion_factor": None,
+        },
+        # Note this is handled differently to normal properties.
+        # EAGLE stores each element in its own group, SWIFT expects
[…]

[…]
