+++ SOAP/compute_halo_properties.py
@@ -342,7 +342,7 @@ def compute_halo_properties():
     exclusive_radii_kpc = []
     for variation in aperture_variations:
         # We don't consider this for apertures defined based on properties
-        if 'radius_in_kpc' not in aperture_variations[variation]:
+        if "radius_in_kpc" not in aperture_variations[variation]:
             continue
         if aperture_variations[variation]["inclusive"]:
             inclusive_radii_kpc.append(aperture_variations[variation]["radius_in_kpc"])
@@ -353,10 +353,10 @@ def compute_halo_properties():
 
     # Add the apertures defined with fixed physical radii
     for variation in aperture_variations:
-        if 'radius_in_kpc' not in aperture_variations[variation]:
+        if "radius_in_kpc" not in aperture_variations[variation]:
             continue
-        assert 'property' not in aperture_variations[variation]
-        assert 'radius_multiple' not in aperture_variations[variation]
+        assert "property" not in aperture_variations[variation]
+        assert "radius_multiple" not in aperture_variations[variation]
         if aperture_variations[variation]["inclusive"]:
             # If skip_gt_enclose_radius is False (which is the default) then
             # we always want to calculate its properties, regardless of the
@@ -397,10 +397,10 @@ def compute_halo_properties():
 
     # Add the apertures based on SOAP properties
     for variation in aperture_variations:
-        if 'radius_in_kpc' in aperture_variations[variation]:
+        if "radius_in_kpc" in aperture_variations[variation]:
             continue
-        assert 'property' in aperture_variations[variation]
-        radius_multiple = aperture_variations[variation].get('radius_multiple', 1)
+        assert "property" in aperture_variations[variation]
+        radius_multiple = aperture_variations[variation].get("radius_multiple", 1)
         # Only allow integer radius mutiples, otherwise swiftsimio will
         # struggle to handle the group names
         assert int(radius_multiple) == radius_multiple
@@ -410,7 +410,7 @@ def compute_halo_properties():
                     cellgrid,
                     parameter_file,
                     None,
-                    (aperture_variations[variation]['property'], radius_multiple),
+                    (aperture_variations[variation]["property"], radius_multiple),
                     recently_heated_gas_filter,
                     stellar_age_calculator,
                     cold_dense_gas_filter,
@@ -425,7 +425,7 @@ def compute_halo_properties():
                     cellgrid,
                     parameter_file,
                     None,
-                    (aperture_variations[variation]['property'], radius_multiple),
+                    (aperture_variations[variation]["property"], radius_multiple),
                     recently_heated_gas_filter,
                     stellar_age_calculator,
                     cold_dense_gas_filter,
@@ -434,7 +434,6 @@ def compute_halo_properties():
                     [],
                 )
             )
-
 
     projected_aperture_variations = parameter_file.get_halo_type_variations(
         "ProjectedApertureProperties",
@@ -450,13 +449,14 @@ def compute_halo_properties():
     # apertures if all the particles were already included in the previous aperture
     projected_aperture_variations = dict(
         sorted(
-            projected_aperture_variations.items(), key=lambda x: x[1].get("radius_in_kpc", 0)
+            projected_aperture_variations.items(),
+            key=lambda x: x[1].get("radius_in_kpc", 0),
         )
     )
     projected_radii_kpc = []
     for variation in projected_aperture_variations:
         # We don't consider this for apertures defined based on properties
-        if 'radius_in_kpc' not in projected_aperture_variations[variation]:
+        if "radius_in_kpc" not in projected_aperture_variations[variation]:
             continue
         projected_radii_kpc.append(
             projected_aperture_variations[variation]["radius_in_kpc"]
@@ -464,11 +464,10 @@ def compute_halo_properties():
[…]
