+++ misc/hdecompose_hydrogen_fractions.py
@@ -42,9 +42,7 @@ from Hdecompose.BlitzRosolowsky2006 import molecular_frac as calculate_molecular
 
 # Parse arguments
 parser = argparse.ArgumentParser(
-    description=(
-        "Script to estimate HI and H2 species fractions."
-    )
+    description=("Script to estimate HI and H2 species fractions.")
 )
 parser.add_argument(
     "--snap-basename",
@@ -68,31 +66,36 @@ os.makedirs(os.path.dirname(output_filename), exist_ok=True)
 
 # List of gas properties we require from the input snapshots
 property_list = [
-    'Densities', 
-    'StarFormationRates', 
-    'ElementMassFractions', 
-    'Temperatures',
+    "Densities",
+    "StarFormationRates",
+    "ElementMassFractions",
+    "Temperatures",
 ]
 
 if comm_rank == 0:
-    print('Reading in run parameters and units')
+    print("Reading in run parameters and units")
     params = {}
     units = {}
     with h5py.File(snap_filename.format(file_nr=0), "r") as file:
-        params['z'] = file['Cosmology'].attrs['Redshift']
-        params['fH'] = file['Parameters'].attrs['EAGLE:InitAbundance_Hydrogen']
-        params['fHe'] = file['Parameters'].attrs['EAGLE:InitAbundance_Helium']
-        params['gamma'] = file['Parameters'].attrs['EAGLE:EOS_Jeans_GammaEffective']
-        params['T0'] = file['Parameters'].attrs['EAGLE:EOS_Jeans_TempNorm_K'] * astropy.units.K
+        params["z"] = file["Cosmology"].attrs["Redshift"]
+        params["fH"] = file["Parameters"].attrs["EAGLE:InitAbundance_Hydrogen"]
+        params["fHe"] = file["Parameters"].attrs["EAGLE:InitAbundance_Helium"]
+        params["gamma"] = file["Parameters"].attrs["EAGLE:EOS_Jeans_GammaEffective"]
+        params["T0"] = (
+            file["Parameters"].attrs["EAGLE:EOS_Jeans_TempNorm_K"] * astropy.units.K
+        )
 
         reg = swift_units.unit_registry_from_snapshot(file)
         for prop in property_list:
-            attrs = file[f'PartType0/{prop}'].attrs
+            attrs = file[f"PartType0/{prop}"].attrs
             units[prop] = swift_units.units_from_attributes(attrs, reg)
-        elements = [e.decode() for e in file['SubgridScheme/NamedColumns/ElementMassFractions'][:]]
-        i_H = elements.index('Hydrogen')
+        elements = [
+            e.decode()
+            for e in file["SubgridScheme/NamedColumns/ElementMassFractions"][:]
+        ]
+        i_H = elements.index("Hydrogen")
 
-        n_file = file['Header'].attrs['NumFilesPerSnapshot'][0]
+        n_file = file["Header"].attrs["NumFilesPerSnapshot"][0]
 else:
     params = None
     units = None
@@ -102,29 +105,33 @@ units = comm.bcast(units)
 i_H = comm.bcast(i_H)
 
 if comm_rank == 0:
-    print('Loading data')
+    print("Loading data")
 
 # Load raw arrays from file
 snap_file = phdf5.MultiFile(
     snap_filename, file_nr_attr=("Header", "NumFilesPerSnapshot")
 )
-raw_properties = snap_file.read(property_list, 'PartType0')
+raw_properties = snap_file.read(property_list, "PartType0")
 
 # Convert to astropy arrays (physical not comoving)
-densities = (raw_properties['Densities'] * units['Densities']).to('g/cm**3').to_astropy()
-temperatures = (raw_properties['Temperatures'] * units['Temperatures']).to('K').to_astropy()
[…]

[…]
