### 2025-05-19 3bb2c8c
Define apertures based on BoundSubhalo properties (#147)
* Add note on timing flags

* Allow for property defined apertures

* Add error messages if required radii are disabled

* Missing method

* Update pdf documentation

* Add massive halos for testing

* Format
