### 2025-05-19 #8252 @ivastar
@dnelson86, @stuartmcalpine, @nvanalfen welcome to the review of SOAP!

Instructions for generating the review checklist are provided above. For recommendations, please open issues on the SOAP repo and cross-list them here.

I would appreciate the first round of reviews by June 15th (or sooner if you have the bandwidth).

In you have questions about the process or anything else, please don't hesitate to tag me here or reach out via email.
