### 2025-05-20 #8252 @robjmcgibbon
Thanks for agreeing to review folks, much appreciated! Just a note to please use the `soap_runtime` branch. It has better documentation, testing, and the code is packaged differently than it is in `master`.

We are still in the process of adding new features to SOAP (although nothing that substantially changes the core functionality). Am I ok to keep merging things @ivastar, or is it better to have a frozen version for the review?

@dnelson86 this repo contains a script that converts EAGLE snapshots to swift format. If I created a similar conversion script for TNG snapshots would that be useful for when you're testing the package?
