### 2025-05-21 #8252 @dnelson86
Nice idea but not needed just for this review. I should be able to test it on some low-res FLAMINGO snapshots.
