### 2025-05-22 #8252 @ivastar
> We are still in the process of adding new features to SOAP (although nothing that substantially changes the core functionality). Am I ok to keep merging things [@ivastar](https://github.com/ivastar), or is it better to have a frozen version for the review?

@robjmcgibbon it is best if the reviewers are looking at the version that you are planning to release and publish, even if the functionality you are adding does not seem to change the core functionality.

Are these features you would like to be included in the release with the paper? If so, should we pause the review? Or, alternatively, keep making the contributions to a different branch assuming that these would be in a later release?
