### 2025-05-22 #8252 @robjmcgibbon
> Are these features you would like to be included in the release with the paper? If so, should we pause the review? Or, alternatively, keep making the contributions to a different branch assuming that these would be in a later release?

@ivastar thanks, no they don't need to be included with the paper release. They are mostly just small changes where we add additional properties (e.g.  we already had stellar half mass radius, but we've just added half light radius too). I will stop adding (non review related) commits to the `soap_runtime` branch
