### 2025-05-22 #8252 @nvanalfen
Very excited to read and review this paper. It looks interesting and quite useful.

@robjmcgibbon I'm trying to run `./tests/run_small_volume.sh` as described in the documentation on the soap_runtime branch, but I get the error:
> bash: ./run_small_volume.sh: cannot execute: required file not found

Looking through the `run_small_volume.sh` file at the files expected, I can find all except `test_data/snap_0018.hdf5` (I am on the soap_runtime branch of your fork). Is this simply me not doing things properly? Or should I open an issue on your branch to include this?
