### 2025-05-22 #8252 @robjmcgibbon
@nvanalfen Could you please double check you are running it from the main directory (e.g. `./tests/run_small_volume.sh` and not `./run_small_volume.sh`). If that's the problem then I can add a note about it to the documentation, if not then I'll have a look at  what's going on
