### 2025-05-22 #8252 @nvanalfen
@robjmcgibbon Yes, I made sure to run from the main directory (after accidentally running from within tests the first attempt) and got that error. I also see no `test_data` folder, though I do see a `generate_test_data` folder. Would something in there create what I need?
