### 2025-05-22 SWIFTSIM/SOAP#152 @robjmcgibbon
Test data not found for small volume
Following on from https://github.com/openjournals/joss-reviews/issues/8252#issuecomment-2901447521
