### 2025-05-22 SWIFTSIM/SOAP#152 @robjmcgibbon
@nvanalfen the `./tests/run_small_volume.sh` should download the example data into `./test_data`. Does this happen for you?
