### 2025-05-22 SWIFTSIM/SOAP#152 @robjmcgibbon
For reference the `tests/generate_test_data` just contains scripts for generating the small simulation, in case I ever need to produce it again, but it shouldn't be needed in general.
