### 2025-05-22 SWIFTSIM/SOAP#152 @nvanalfen
@robjmcgibbon This doesn't happen for me. I run `./tests/run_small_volume.sh` and I get the error message `bash: ./tests/run_small_volume.sh: cannot execute: required file not found`. I'm making sure to run from the main directory. I installed the package following the documentation using:
`pip3 install git+https://github.com/SWIFTSIM/SOAP.git@soap_runtime`,
Then I uninstalled `h5py` and ran `pip3 install -r requirements.txt` (there was a note that doing the `pip install` route may install a serial h5py while a parallel is needed, so I did this extra step). Then I tried running that test file.

I'm looking in the `tests/helpers.py` file now. This seems related to that missing snap file, is this true?
