### 2025-05-22 SWIFTSIM/SOAP#152 @nvanalfen
I should also note that I'm running on Ubuntu. A quick Google suggests that one possible issue is some DOS-Unix incompatibility with CRLF line endings. Not sure if that's the case here, but I thought it worth mentioning. I'm not even sure if `run_small_volume.sh` is running. At this point, I suspect it's not.
