### 2025-05-22 SWIFTSIM/SOAP#152 @robjmcgibbon
Does `bash ./run_small_volume.sh` give the same error?
