### 2025-05-22 SWIFTSIM/SOAP#152 @nvanalfen
This time I get the error message:

> : invalid optionl_volume.sh: line 5: set: -
set: usage: set [-abefhkmnptuvxBCEHPT] [-o option-name] [--] [-] [arg ...]
./tests/run_small_volume.sh: line 6: $'\r': command not found
./tests/run_small_volume.sh: line 37: syntax error: unexpected end of file

So I suspect it does have something to do with the carriage return `\r` character.
