### 2025-05-22 SWIFTSIM/SOAP#152 @nvanalfen
@robjmcgibbon This did in fact change things. I used VSCode to change CLRF to LF and was able to run the file. I'm getting the error `ValueError: h5py was built without MPI support, can't use mpio driver`, but that has to do with my environment setup, not your code. test_data has been created and the right files downlaoded into it.
