### 2025-05-22 SWIFTSIM/SOAP#152 @robjmcgibbon
Strange, I've never come across this before. If you do a git diff is the file actually changed, or is this something to do with how VScode reads in the file?

I've not been able to reproduce it, so not sure how to fix it just yet. I'm able to run the file on ubuntu, but that's just in the terminal rather than in VScode (or some other IDE)
