### 2025-05-22 SWIFTSIM/SOAP#152 @nvanalfen
It does indeed show up as modified in git status. Was this made using a Windows machine? I know Windows uses CRLF while Linux and Unix use LF. I also know that git often converts one to the other when committing, though it can be turned off and I haven't looked too deeply into it. It seems that the extra `\r` character was being read as part of commands and file names by Ubuntu (which only expects `\n` for newlines instead of `\r\n`) when I ran the `.sh` file .

I think Windows will also respect LF, but I don't know. I'm not sure this needs to be "fixed" as it's not exactly anything wrong with the script. It's odd though. Since it's a `.sh` file, I assume you're using some Unix or Linux system. I wonder why my computer had issues with it but not yours. Can you check the line endings on your computer to see if they're CRLF or LF? Maybe git is converting.
