### 2025-05-22 SWIFTSIM/SOAP#152 @robjmcgibbon
The check from https://unix.stackexchange.com/a/389616 indicates they are LF when I run it on the cluster where I do SOAP development (which runs Rocky Linux).

I wonder if you also would have issues with the other bash scripts in the scripts directory.

I'll close this for now, I can reopen and try and find a fix if someone else has the same issue.
