### 2025-05-28 SWIFTSIM/SOAP#153 @stuartmcalpine
JOSS review notes [McAlpine]
In relation to https://github.com/openjournals/joss-reviews/issues/8252#issuecomment-2901447521

Note I performed these tests on Ubuntu locally using the `soap_runtime` branch.

## Readme

When mentioning the output can be read with swiftsimio, can link to the repo

### Installation

`h5py` for parallel support can sometimes be tricky to get going. You mention it in the installation briefly "but this may install a serial version of h5py.". But as far as i'm aware by default `h5py` will always be installed in serial mode.

Could mention something like `HDF5_MPI="ON" CC=mpicc pip install --no-binary=h5py h5py` or that they will have to (uninstall then) install h5py manually after installing the dependencies.

### Running SOAP

Mention the generated PDF goes into `documentation/SOAP.pdf`

### Tests

Running `pytest -m "not mpi" -W error` I get 5 errors

```bash
FAILED test_SO_properties.py::test_SO_properties_random_halo - AttributeError: 'int' object has no attribute 'value'
FAILED test_SO_properties.py::test_concentration_nfw_halo - AttributeError: 'int' object has no attribute 'value'
FAILED test_aperture_properties.py::test_aperture_properties - TypeError: ExclusiveSphereProperties.__init__() missing 1 required positional argument: 'all_radii_kpc'
FAILED test_projected_aperture_properties.py::test_projected_aperture_properties - TypeError: ProjectedApertureProperties.__init__() missing 1 required positional argument: 'all_radii_kpc'
FAILED test_subhalo_properties.py::test_subhalo_properties - AttributeError: 'int' object has no attribute 'value'
```

Also, why `pytest -m mpi --with-mpi -W error` and not `mpirun -np X pytest -m mpi --with-mpi -W error`? This will test on multiple ranks. Note the MPI tests do pass for me (both with and without `mpirun`).

### Parameter files

All the example parameter files are hydro-sounding-esk. You mention the `--dmo` flag in the `Computing halo properties` section. Can all these example parameter files be run on DMO versions? If so mention that in the parameter file section of the readme, if not, would be good to create a DMO example parameter file.

## General thoughts

Could consider having the two run scripts as console entry points so that people don't have to run the path-dependent invocation.

eg something like

```toml
[project.scripts]
soap-group-membership = "SOAP.group_membership:main"
soap-compute-halo-properties = "SOAP.compute_halo_properties:main"
```

in `pyproject.toml`.

---

In the autogenerated PDF after running the small example, the figure reference doesn't seem to render (Fig. ??).

---

Could mention test coverage or validation strategy in the JOSS paper.
