### 2025-05-28 dc94ed7
Add project scripts & fix tests
