+++ SOAP/particle_selection/aperture_properties.py
@@ -3287,6 +3287,8 @@ class ApertureParticleData:
         """
         Half light radius of stars for the 9 GAMA bands.
         """
+        if self.Nstar == 0:
+            return None
         return get_half_light_radius(
             self.radius[self.type == 4],
             self.stellar_luminosities,
@@ -3759,7 +3761,7 @@ class ApertureProperties(HaloProperty):
                                 registry=registry,
                             )
                         else:
-                            err = f'Overflow for halo {input_halo["index"].value} when '
+                            err = f'Overflow for halo {input_halo["index"]} when '
                             err += f"calculating {name} in aperture_properties"
                             assert np.max(np.abs(val.to(unit).value)) < float(
                                 "inf"
