+++ SOAP/compute_halo_properties.py
@@ -17,12 +17,25 @@ import traceback
 import numpy as np
 import unyt
 
-from SOAP.core import chunk_tasks, halo_centres, lustre, result_set, soap_args, swift_cells, task_queue
+from SOAP.core import (
+    chunk_tasks,
+    halo_centres,
+    lustre,
+    result_set,
+    soap_args,
+    swift_cells,
+    task_queue,
+)
 from SOAP.core.parameter_file import ParameterFile
 from SOAP.core.mpi_timer import MPITimer
 from SOAP.core.combine_chunks import combine_chunks, sub_snapnum
 from SOAP.core.category_filter import CategoryFilter
-from SOAP.particle_selection import aperture_properties, SO_properties, projected_aperture_properties, subhalo_properties
+from SOAP.particle_selection import (
+    aperture_properties,
+    SO_properties,
+    projected_aperture_properties,
+    subhalo_properties,
+)
 from SOAP.property_calculation.stellar_age_calculator import StellarAgeCalculator
 from SOAP.particle_filter.cold_dense_gas_filter import ColdDenseGasFilter
 from SOAP.particle_filter.recently_heated_gas_filter import RecentlyHeatedGasFilter
