+++ SOAP/group_membership.py
@@ -316,5 +316,6 @@ def main():
     if comm_rank == 0:
         print("Done.")
 
+
 if __name__ == "__main__":
     main()
