### 2025-05-28 SWIFTSIM/SOAP#153 @robjmcgibbon
Thanks Stuart, that's all very helpful! I've addressed most of your comments in https://github.com/SWIFTSIM/SOAP/pull/154

### Readme
I've linked to swiftsimio

### Installation
I've added a list of steps which should install h5py in parallel mode, and then install SOAP.

### Running SOAP
I've added a note on the location of the generated test run pdf

### Tests
Tests have been fixed

The script I use to run the tests on cosma does run on multiple ranks, I just didn't copy that over to the README. Fixed now.

### Parameter files
Yep, all the parameters files can be run on the DMO equivalent of those hydro runs. I added a note in the parameter file README about DMO runs.

### General thoughts
I've added those nicer entry points to `pyproject.toml` and the README

I'm too reliant on overleaf, so don't have a good understanding of exactly how `pdflatex` works. However, running it twice fixes the problem, so I've just updated `tests/run_small_volume.sh` to do that.

Indeed, it would be good to have a few lines about test coverage in the paper. I'm currently over the recommended word count though, so I'll wait until I hear from the other reviewers about their suggestions before I make any paper changes.
