### 2025-05-28 3bd6955
JOSS - Stuart comments 1 (#154)
* Merge fof position test

* Update docs

* Add project scripts & fix tests

* Format

* Add venv script for strw
