+++ SOAP/core/parameter_file.py
@@ -341,4 +341,6 @@ class ParameterFile:
         """
         Returns a bool indicating if BoundSubhalo/EncloseRadius is enabled
         """
-        return self.parameters["SubhaloProperties"]["properties"].get("EncloseRadius", False)
+        return self.parameters["SubhaloProperties"]["properties"].get(
+            "EncloseRadius", False
+        )
