+++ tests/test_aperture_properties.py
@@ -59,6 +59,7 @@ def test_aperture_properties():
         dummy_halos.get_cell_grid(),
         parameters,
         50.0,
+        None,
         recently_heated_filter,
         stellar_age_calculator,
         cold_dense_gas_filter,
@@ -70,6 +71,7 @@ def test_aperture_properties():
         dummy_halos.get_cell_grid(),
         parameters,
         50.0,
+        None,
         recently_heated_filter,
         stellar_age_calculator,
         cold_dense_gas_filter,
@@ -84,6 +86,7 @@ def test_aperture_properties():
         dummy_halos.get_cell_grid(),
         parameters,
         50.0,
+        None,
         recently_heated_filter,
         stellar_age_calculator,
         cold_dense_gas_filter,
@@ -175,6 +178,7 @@ def test_aperture_properties():
             dummy_halos.get_cell_grid(),
             single_parameters,
             50.0,
+            None,
             recently_heated_filter,
             stellar_age_calculator,
             cold_dense_gas_filter,
@@ -186,6 +190,7 @@ def test_aperture_properties():
             dummy_halos.get_cell_grid(),
             single_parameters,
             50.0,
+            None,
             recently_heated_filter,
             stellar_age_calculator,
             cold_dense_gas_filter,

[…]
