### 2025-05-28 SWIFTSIM/SOAP#155 @nvanalfen
JOSS Review Questions
@robjmcgibbon Opening an issue for my more in-depth JOSS review notes. Most of these will be me struggling with the documentation. In relation to https://github.com/openjournals/joss-reviews/issues/8252.

First off, I'm overall impressed and excited by this package. I simply have a knack for running into issues, so the lengthy list speaks to that.

# h5py Installation
I saw in the other similar issue some of the same notes I had, so I will omit those here. But as an added note to the difficulty installing parallel h5py, this is what I found to work for me:
<ul>
    <li>Create a conda environment (venv is my preferred, but did not work here as I had clashing h5py and openmpi versions)</li>
    <li>conda install -c conda-forge mpi4py openmpi</li>
    <li>conda install -c conda-forge "h5py=*=mpi_*" (this is the main place where my installation couldn't follow the direction listed in the README)</li>
    <li>pip3+git for your repo and pip3 install -r requirements.txt (as outlined in the README)</li>
</ul>

I had a very hard time getting this working, and the README didn't quite work for me. Once I got this set up, however, the tests ran for me just fine using the pytest calls given in the documentation.

# Package versions
- numpy < 2
I had to explicitly downgrade numpy to 1.26.* to be able to run tests/run_small_volume.sh. I believe this is quite common still as many packages are lagging behind in supporting numpy 2, but I thought it worth mentioning as many fresh environments will install the latest package versions by default.

# Struggled with "Running SOAP" Documentation.
In the README, there is an example under "Computing halo membership for particles in the snapshot" that shows:
snapnum=0077

```
sim=L1000N0900/DMO_FIDUCIAL
mpirun python soap-group-membership \
    --sim-name=${sim} --snap-nr${snapnum} parameter_files/FLAMINGO.yml
```

### Small Typo
- A small typo: `--snap-nr${snapnum}` is missing an `=` ans should be `--snap-nr=${snapnum}`.

### File not found error
When I run it (with the = put in), I get the following error:

> python3: can't open file '/home/nvanalfen/Documents/JOSSReviewer/SOAP/soap-group-membership': [Errno 2] No such file or directory

 If I change `soap-group-membership` to `SOAP/group_membership.py` (my best guess at what was supposed to be called), things started running but then I got the error:

> FileNotFoundError: [Errno 2] Unable to synchronously open file (unable to open file: name = '/cosma8/data/dp004/flamingo/Runs/L1000N0900/DMO_FIDUCIAL/HBT/077/SubSnap_077.0.hdf5', errno = 2, error message = 'No such file or directory', flags = 0, o_flags = 0)

I assume this is mostly due to my not understanding what's going on. Is this code chunk simply a "for example", or is it supposed to be something we can run to test things?

I imagine the `soap-group-membership` is not a typo and my issue stems from me just doing something wrong. Can you clarify this for me? And is it an essential test that I should be able to run?

# Closing
I will continue to use this issue for questions and struggles. I'll try to keep things organized (as best I can with my scattered thought process).
