### 2025-05-28 SWIFTSIM/SOAP#155 @robjmcgibbon
Hi Nick, thanks for the comments, apologies it's been a bit of an issue getting everything running.

### h5py Installation
Yeah I understand installing the parallel version can be a real pain. I haven't put any instructions for installing parallel HDF5 itself since I don't picture people running SOAP on their own PCs much, more on HPC, in which case the modules are likley to be managed by an adminisrator. I could add another section on installing h5py, including the instructiosn you have kindly provided, but I don't think those will work in all cases.

### Package versions
Do you know which package had an issue with numpy < 2? I think older mpi4py/h5py versions didn't like it, but there could be others too.

###Struggled with "Running SOAP" Documentation
You're correct that's a typo in `snap_nr`, I will fix it

The `soap-group-membership` is something that I just added today as a response to Stuart's comments. For that I edited the `pyproject.toml`, so hopefully if you rerun `pip install -e .` it will fix the problem. It just point to `SOAP/group_membership.py` anyway.
