+++ misc/match_group_membership.py
@@ -1,0 +1,319 @@
+#!/bin/env python
+
+"""
+# TODO: Rewrite
+calculate_fof_positions.py
+
+This script calculates the maximum and minimum particles positions for each FOF.
+Usage:
+
+  mpirun -- python misc/calculate_fof_positions.py \
+          --snap-basename=SNAPSHOT \
+          --fof-basename=FOF \
+          --output-basename=OUTPUT
+
+where SNAPSHOT is the basename of the snapshot files (the snapshot
+name without the .{file_nr}.hdf5 suffix), FOF is the basename of the
+fof catalogues, and OUTPUT is the basename of the output fof catalogues.
+"""
+
+import argparse
+import os
+
+from mpi4py import MPI
+
+comm = MPI.COMM_WORLD
+comm_rank = comm.Get_rank()
+comm_size = comm.Get_size()
+
+import h5py
+import numpy as np
+
+import virgo.mpi.parallel_sort as psort
+import virgo.mpi.parallel_hdf5 as phdf5
+from virgo.mpi.gather_array import gather_array
+
+# Parse arguments
+parser = argparse.ArgumentParser(
+    description=("Script to calculate extent of FoF groups.")
+)
+parser.add_argument(
+    "--snap-basename1",
+    type=str,
+    required=True,
+    help=(
+        "The basename for the snapshot files (the snapshot "
+        "name without the .{file_nr}.hdf5 suffix)"
+    ),
+)
+parser.add_argument(
+    "--snap-basename2",
+    type=str,
+    required=True,
+    help=(
+        "The basename for the snapshot files (the snapshot "
+        "name without the .{file_nr}.hdf5 suffix)"
+    ),
+)
+parser.add_argument(
+    "--membership-basename1",
+    type=str,
+    required=True,
+    help=(
+        "The basename for the snapshot files (the snapshot "
+        "name without the .{file_nr}.hdf5 suffix)"
+    ),
+)
+parser.add_argument(
+    "--membership-basename2",
+    type=str,
+    required=True,
+    help=(
+        "The basename for the snapshot files (the snapshot "
+        "name without the .{file_nr}.hdf5 suffix)"
+    ),
+)
+parser.add_argument(
+    "--soap-filename1",
+    type=str,
+    required=True,
+    help=(
+        "The basename for the snapshot files (the snapshot "
+        "name without the .{file_nr}.hdf5 suffix)"
+    ),
+)
+# TODO:
+parser.add_argument(
+    "--soap-filename",
+    type=str,
+    required=True,
+    help=(
+        "The basename for the snapshot files (the snapshot "
+        "name without the .{file_nr}.hdf5 suffix)"
+    ),
+)
+# TODO:
+parser.add_argument(
+    "--output-filename",
+    type=str,
+    required=True,
+    help=(
+        "The basename for the snapshot files (the snapshot "
+        "name without the .{file_nr}.hdf5 suffix)"
+    ),
+)
+# TODO:
+parser.add_argument(
+    "--ptypes",
+    type=bool,
+    required=False,
+    help="Only match central halos",
+)
+parser.add_argument(
+    "--nr_particles",
+    type=int,
+    required=False,
+    default=50,
+    help="Number of particles to use when matching. -1 to use all particles",
+)
+parser.add_argument(
+    "--centrals-only",
+    type=bool,
+    required=False,
+    help="Only match central halos",
+)
+args = parser.parse_args()
+snap_filename1 = args.snap_basename1 + ".{file_nr}.hdf5"
+snap_filename2 = args.snap_basename2 + ".{file_nr}.hdf5"
+membership_filename1 = args.membership_basename1 + ".{file_nr}.hdf5"
[…]
