### 2025-05-30 e18c63e
Fix test in calculate_fof_pos
