+++ misc/match_group_membership.py
@@ -1,7 +1,7 @@
 #!/bin/env python
 
-"""
 # TODO: Rewrite
+"""
 calculate_fof_positions.py
 
 This script calculates the maximum and minimum particles positions for each FOF.
@@ -73,6 +73,7 @@ parser.add_argument(
         "name without the .{file_nr}.hdf5 suffix)"
     ),
 )
+# TODO:
 parser.add_argument(
     "--soap-filename1",
     type=str,
@@ -84,7 +85,7 @@ parser.add_argument(
 )
 # TODO:
 parser.add_argument(
-    "--soap-filename",
+    "--soap-filename2",
     type=str,
     required=True,
     help=(
@@ -123,10 +124,6 @@ parser.add_argument(
     help="Only match central halos",
 )
 args = parser.parse_args()
-snap_filename1 = args.snap_basename1 + ".{file_nr}.hdf5"
-snap_filename2 = args.snap_basename2 + ".{file_nr}.hdf5"
-membership_filename1 = args.membership_basename1 + ".{file_nr}.hdf5"
-membership_filename2 = args.membership_basename2 + ".{file_nr}.hdf5"
 
 # TODO: 
 ptypes = [1]
@@ -139,23 +136,25 @@ def load_particle_data(snap_basename, membership_basename, ptypes, comm):
 
     # Load particle IDs
     snap_filename = snap_basename + ".{file_nr}.hdf5"
-    snap_file = phdf5.MultiFile(
+    file = phdf5.MultiFile(
         snap_filename, file_nr_attr=("Header", "NumFilesPerSnapshot"), comm=comm
     )
     particle_ids = []
     for ptype in ptypes:
-        particle_ids.append(file.read(f"PartType{particle_type}/ParticleIDs"))
+        particle_ids.append(file.read(f"PartType{ptype}/ParticleIDs"))
     particle_ids = np.concatenate(particle_ids)
 
+    # Membership files don't have a header, so create a list of filenames
+    n_file = len(file.filenames)
+    membership_filenames = [f'{membership_basename}.{i}.hdf5' for i in range(n_file)]
     # Load membership information
-    membership_filename = membership_basename + ".{file_nr}.hdf5"
-    membership_file = phdf5.MultiFile(
-        membership_filename, file_nr_attr=("Header", "NumFilesPerSnapshot"), comm=comm
+    file = phdf5.MultiFile(
+        membership_filenames, file_nr_attr=("Header", "NumFilesPerSnapshot"), comm=comm
     )
     halo_catalogue_idx, rank_bound = [], []
     for ptype in ptypes:
-        halo_catalogue_idx.append(file.read(f"PartType{particle_type}/GroupNr_bound"))
-        rank_bound.append(file.read(f"PartType{particle_type}/Rank_bound"))
+        halo_catalogue_idx.append(file.read(f"PartType{ptype}/GroupNr_bound"))
+        rank_bound.append(file.read(f"PartType{ptype}/Rank_bound"))
     halo_catalogue_idx, rank_bound = np.concatenate(halo_catalogue_idx), np.concatenate(rank_bound)
 
     # Check the two files are partitioned the same way
@@ -173,22 +172,21 @@ def load_particle_data(snap_basename, membership_basename, ptypes, comm):
         'rank_bound': rank_bound,
     }
 
-data_1 = load_particle_data(args.snap_basename_1, args.membership_basename_1, ptypes, comm)
-data_2 = load_particle_data(args.snap_basename_2, args.membership_basename_2, ptypes, comm)
-datasets = list(data_1.keys())
 
-# Remove particles which are not bound in both snapshots
-match_1 = psort.parallel_match(data_1['particle_ids'], data_2['particle_ids'], comm=comm)
-for dset in datasets:
[…]
