### 2025-05-30 ad22dca
Match to SOAP idx
