+++ misc/match_group_membership.py
@@ -42,8 +42,8 @@ parser.add_argument(
     type=str,
     required=True,
     help=(
-        "The basename for the snapshot files (the snapshot "
-        "name without the .{file_nr}.hdf5 suffix)"
+        "The basename of the snapshot files (the snapshot "
+        "name without the .{file_nr}.hdf5 suffix) for simulation 1"
     ),
 )
 parser.add_argument(
@@ -51,8 +51,7 @@ parser.add_argument(
     type=str,
     required=True,
     help=(
-        "The basename for the snapshot files (the snapshot "
-        "name without the .{file_nr}.hdf5 suffix)"
+        "The basename of the snapshot files for simulation 2"
     ),
 )
 parser.add_argument(
@@ -60,8 +59,7 @@ parser.add_argument(
     type=str,
     required=True,
     help=(
-        "The basename for the snapshot files (the snapshot "
-        "name without the .{file_nr}.hdf5 suffix)"
+        "The basename of the membership files for simulation 1"
     ),
 )
 parser.add_argument(
@@ -69,8 +67,7 @@ parser.add_argument(
     type=str,
     required=True,
     help=(
-        "The basename for the snapshot files (the snapshot "
-        "name without the .{file_nr}.hdf5 suffix)"
+        "The basename of the membership files for simulation 2"
     ),
 )
 # TODO:
@@ -79,7 +76,7 @@ parser.add_argument(
     type=str,
     required=True,
     help=(
-        "The basename for the snapshot files (the snapshot "
+        "The  for the snapshot files (the snapshot "
         "name without the .{file_nr}.hdf5 suffix)"
     ),
 )
@@ -188,10 +185,13 @@ def load_soap(soap_filename, comm):
     }
 
 
-def match_sim(particle_ids, particle_halo_ids, rank_bound, particle_ids_to_match, particle_halo_ids_to_match):
+def match_sim(particle_ids, particle_halo_ids, rank_bound, particle_ids_to_match, particle_halo_ids_to_match, soap, soap_to_match):
     '''
-    Takes in particle_ids, particle_halo_ids, and rank_bound from simulation 1
-    and particle_ids_to_match, and particle_halo_ids_to_match from simulation 2.
+    Input:
+        - particle_ids, particle_halo_ids, and rank_bound from simulation 1
+        - particle_ids_to_match, and particle_halo_ids_to_match from simulation 2
+        - soap catalogue from simulation 1
+        - soap_to_match catalogue from simulation 2
 
     Returns halo_ids, matched_halo_ids, n_match
     '''
@@ -235,7 +235,6 @@ def match_sim(particle_ids, particle_halo_ids, rank_bound, particle_ids_to_match
             ranks, weights=gathered_counts, minlength=comm_size
         ).astype(np.int64)
         assert np.sum(n_part_per_rank) == np.sum(gathered_counts)
-        print(n_part_per_rank)
     else:
         n_part_per_rank = None
     n_part_per_rank = comm.bcast(n_part_per_rank)
@@ -298,18 +297,33 @@ def match_sim(particle_ids, particle_halo_ids, rank_bound, particle_ids_to_match
     matched_halo_ids = matched_halo_ids[idx]
     combined_counts = combined_counts[idx]
 
-    return halo_ids, matched_halo_ids, combined_counts
+    # Get the SOAP index of each object in matched_halo_ids
+    matched_soap_idx = psort.parallel_match(matched_halo_ids, soap_to_match['halo_catalogue_idx'], comm=comm)
 
-if comm_rank == 0:
-    print('Loading data from simulation 1')
[…]
