### 2025-05-30 169d4a3
Check matches are consistent
