### 2025-05-30 ba2ed01
Add __main__
