+++ misc/match_group_membership.py
@@ -38,90 +38,6 @@ import virgo.mpi.parallel_sort as psort
 import virgo.mpi.parallel_hdf5 as phdf5
 from virgo.mpi.gather_array import gather_array
 
-# Parse arguments
-parser = argparse.ArgumentParser(
-    description=("Script to calculate extent of FoF groups.")
-)
-parser.add_argument(
-    "--snap-basename1",
-    type=str,
-    required=True,
-    help=(
-        "The basename of the snapshot files (the snapshot "
-        "name without the .{file_nr}.hdf5 suffix) for simulation 1"
-    ),
-)
-parser.add_argument(
-    "--snap-basename2",
-    type=str,
-    required=True,
-    help=(
-        "The basename of the snapshot files for simulation 2"
-    ),
-)
-parser.add_argument(
-    "--membership-basename1",
-    type=str,
-    required=True,
-    help=(
-        "The basename of the membership files for simulation 1"
-    ),
-)
-parser.add_argument(
-    "--membership-basename2",
-    type=str,
-    required=True,
-    help=(
-        "The basename of the membership files for simulation 2"
-    ),
-)
-parser.add_argument(
-    "--soap-filename1",
-    type=str,
-    required=True,
-    help=(
-        "The filename of the SOAP catalogue for simulation 1"
-    ),
-)
-parser.add_argument(
-    "--soap-filename2",
-    type=str,
-    required=True,
-    help=(
-        "The filename of the SOAP catalogue for simulation 2"
-    ),
-)
-parser.add_argument(
-    "--output-filename",
-    type=str,
-    required=True,
-    help=(
-        "The filename of the output file"
-    ),
-)
-parser.add_argument(
-    "--ptypes",
-    type=int,
-    required=False,
-    nargs='+',
-    default=[1],
-    help="Particle types to use for the matching. Defaults to [1]",
-)
-parser.add_argument(
-    "--nr_particles",
-    type=int,
-    required=False,
-    default=50,
-    help="Number of particles to use when matching. Defaults to 50. -1 to use all particles",
-)
-parser.add_argument(
-    "--centrals-only",
-    type=bool,
-    required=False,
-    default=True,
-    help="Only match central halos. Defaults to True",
-)
-args = parser.parse_args()
 
 def load_particle_data(snap_basename, membership_basename, ptypes, comm):
     '''
@@ -359,76 +275,162 @@ def mpi_print(string, comm_rank):
     if comm_rank == 0:
         print(string)
 
-mpi_print('Loading data from simulation 1', comm_rank)
-data_1 = load_particle_data(args.snap_basename1, args.membership_basename1, args.ptypes, comm)
-soap_1 = load_soap(args.soap_filename1, comm)
+if __name__ == '__main__':
 
-mpi_print('Loading data from simulation 2', comm_rank)
-data_2 = load_particle_data(args.snap_basename2, args.membership_basename2, args.ptypes, comm)
-soap_2 = load_soap(args.soap_filename2, comm)
+    parser = argparse.ArgumentParser(
+        description=("Script to calculate extent of FoF groups.")
+    )
+    parser.add_argument(
+        "--snap-basename1",
+        type=str,
+        required=True,
+        help=(
+            "The basename of the snapshot files (the snapshot "
+            "name without the .{file_nr}.hdf5 suffix) for simulation 1"
+        ),
+    )
+    parser.add_argument(
+        "--snap-basename2",
+        type=str,
+        required=True,
+        help=(
+            "The basename of the snapshot files for simulation 2"
+        ),
+    )
+    parser.add_argument(
+        "--membership-basename1",
+        type=str,
+        required=True,
+        help=(
[…]
