+++ misc/match_group_membership.py
@@ -40,10 +40,10 @@ from virgo.mpi.gather_array import gather_array
 
 
 def load_particle_data(snap_basename, membership_basename, ptypes, comm):
-    '''
+    """
     Load the particle IDs and halo membership for the particle types
     we will use to match. Removes unbound particles.
-    '''
+    """
 
     # Load particle IDs
     snap_filename = snap_basename + ".{file_nr}.hdf5"
@@ -57,7 +57,7 @@ def load_particle_data(snap_basename, membership_basename, ptypes, comm):
 
     # Membership files don't have a header, so create a list of filenames
     n_file = len(file.filenames)
-    membership_filenames = [f'{membership_basename}.{i}.hdf5' for i in range(n_file)]
+    membership_filenames = [f"{membership_basename}.{i}.hdf5" for i in range(n_file)]
     # Load membership information
     file = phdf5.MultiFile(
         membership_filenames, file_nr_attr=("Header", "NumFilesPerSnapshot"), comm=comm
@@ -66,7 +66,8 @@ def load_particle_data(snap_basename, membership_basename, ptypes, comm):
     for ptype in ptypes:
         halo_catalogue_idx.append(file.read(f"PartType{ptype}/GroupNr_bound"))
         rank_bound.append(file.read(f"PartType{ptype}/Rank_bound"))
-    halo_catalogue_idx, rank_bound = np.concatenate(halo_catalogue_idx), np.concatenate(rank_bound)
+    halo_catalogue_idx = np.concatenate(halo_catalogue_idx)
+    rank_bound = np.concatenate(rank_bound)
 
     # Check the two files are partitioned the same way
     assert particle_ids.shape == halo_catalogue_idx.shape
@@ -78,29 +79,37 @@ def load_particle_data(snap_basename, membership_basename, ptypes, comm):
     rank_bound = rank_bound[mask]
 
     return {
-        'particle_ids': particle_ids,
-        'halo_catalogue_idx': halo_catalogue_idx,
-        'rank_bound': rank_bound,
+        "particle_ids": particle_ids,
+        "halo_catalogue_idx": halo_catalogue_idx,
+        "rank_bound": rank_bound,
     }
 
 
 def load_soap(soap_filename, comm):
-    '''
+    """
     Loads the required fields from a SOAP catalogue
-    '''
+    """
     # Load particle IDs
     file = phdf5.MultiFile(
         soap_filename, file_nr_attr=("Header", "NumFilesPerSnapshot"), comm=comm
     )
     return {
-        'halo_catalogue_idx': file.read(f"InputHalos/HaloCatalogueIndex"),
-        'host_halo_idx': file.read(f"SOAP/HostHaloIndex"),
-        'is_central': file.read(f"InputHalos/IsCentral") == 1,
+        "halo_catalogue_idx": file.read(f"InputHalos/HaloCatalogueIndex"),
+        "host_halo_idx": file.read(f"SOAP/HostHaloIndex"),
+        "is_central": file.read(f"InputHalos/IsCentral") == 1,
     }
 
 
-def match_sim(particle_ids, particle_halo_ids, rank_bound, particle_ids_to_match, particle_halo_ids_to_match, soap, soap_to_match):
-    '''
+def match_sim(
+    particle_ids,
+    particle_halo_ids,
+    rank_bound,
+    particle_ids_to_match,
+    particle_halo_ids_to_match,
+    soap,
+    soap_to_match,
+):
+    """
     Input:
         - particle_ids, particle_halo_ids, and rank_bound from simulation 1
         - particle_ids_to_match, and particle_halo_ids_to_match from simulation 2
[…]
