### 2025-05-30 ea35b57
Match subhalos using membership files (#156)
* Backbone broken

* Fix test in calculate_fof_pos

* Run matching

* Match to SOAP idx

* Check matches are consistent

* Output to file

* Squash bug when saving

* Update docs

* Centrals only

* Update README

* Add __main__

* Run formatter
