+++ misc/match_group_membership.py
@@ -1,0 +1,465 @@
+#!/bin/env python
+
+"""
+match_group_membership.py
+
+This script matches halos between different simulations run from the
+same initial conditions.
+
+Usage:
+
+  mpirun -- python -u misc/match_group_membership \
+            --snap-basename1 SNAP_BASENAME1 \
+            --snap-basename2 SNAP_BASENAME2 \
+            --membership-basename1 MEMBERSHIP_BASENAME1 \
+            --membership-basename2 MEMBERSHIP_BASENAME2 \
+            --soap-filename1 SOAP_FILENAME1 \
+            --soap-filename2 SOAP_FILENAME2 \
+            --output-filename OUTPUT_FILENAME
+
+Run "python misc/match_group_membership.py -h" for a discription
+of the arguments.
+
+"""
+
+import argparse
+import os
+
+from mpi4py import MPI
+
+comm = MPI.COMM_WORLD
+comm_rank = comm.Get_rank()
+comm_size = comm.Get_size()
+
+import h5py
+import numpy as np
+
+import virgo.mpi.parallel_sort as psort
+import virgo.mpi.parallel_hdf5 as phdf5
+from virgo.mpi.gather_array import gather_array
+
+
+def load_particle_data(snap_basename, membership_basename, ptypes, comm):
+    """
+    Load the particle IDs and halo membership for the particle types
+    we will use to match. Removes unbound particles.
+    """
+
+    # Load particle IDs
+    snap_filename = snap_basename + ".{file_nr}.hdf5"
+    file = phdf5.MultiFile(
+        snap_filename, file_nr_attr=("Header", "NumFilesPerSnapshot"), comm=comm
+    )
+    particle_ids = []
+    for ptype in ptypes:
+        particle_ids.append(file.read(f"PartType{ptype}/ParticleIDs"))
+    particle_ids = np.concatenate(particle_ids)
+
+    # Membership files don't have a header, so create a list of filenames
+    n_file = len(file.filenames)
+    membership_filenames = [f"{membership_basename}.{i}.hdf5" for i in range(n_file)]
+    # Load membership information
+    file = phdf5.MultiFile(
+        membership_filenames, file_nr_attr=("Header", "NumFilesPerSnapshot"), comm=comm
+    )
+    halo_catalogue_idx, rank_bound = [], []
+    for ptype in ptypes:
+        halo_catalogue_idx.append(file.read(f"PartType{ptype}/GroupNr_bound"))
+        rank_bound.append(file.read(f"PartType{ptype}/Rank_bound"))
+    halo_catalogue_idx = np.concatenate(halo_catalogue_idx)
+    rank_bound = np.concatenate(rank_bound)
+
+    # Check the two files are partitioned the same way
+    assert particle_ids.shape == halo_catalogue_idx.shape
+
+    # Remove any particles which are not bound to a subhalo
+    mask = halo_catalogue_idx != -1
+    particle_ids = particle_ids[mask]
+    halo_catalogue_idx = halo_catalogue_idx[mask]
+    rank_bound = rank_bound[mask]
+
+    return {
+        "particle_ids": particle_ids,
+        "halo_catalogue_idx": halo_catalogue_idx,
+        "rank_bound": rank_bound,
+    }
+
+
+def load_soap(soap_filename, comm):
+    """
+    Loads the required fields from a SOAP catalogue
+    """
+    # Load particle IDs
+    file = phdf5.MultiFile(
+        soap_filename, file_nr_attr=("Header", "NumFilesPerSnapshot"), comm=comm
+    )
+    return {
+        "halo_catalogue_idx": file.read(f"InputHalos/HaloCatalogueIndex"),
[…]
