### 2025-06-02 SWIFTSIM/SOAP#155 @nvanalfen
@robjmcgibbon If I use numpy version >= 2, I just get errors and things break when I run the tests. I'm still getting some failed tests when I run the tests after downgrading numpy to 1.26.4, but everything runs.

Using numpy < 2, the following tests fail:
- test_SO_properties_random_halo
- test_concentration_nfw_halo
- test_aperture_properties
- test_subhalo_properties

However, ALL the tests seem to fail due to a call to `minimize_scalar` which gives the error:

> ValueError: Use of `bounds` is incompatible with 'method=Brent'
