### 2025-06-02 SWIFTSIM/SOAP#155 @nvanalfen
Ah wait. I didn't have pytest-mpi installed. Installed that and rerunning now
