### 2025-06-02 SWIFTSIM/SOAP#155 @robjmcgibbon
Thanks for working through this, very useful to have people trying it on different setups. I'd prefer not to require python >= 3.11 unless required, but I will add `astropy>6`
