### 2025-06-02 1c07ee0
Require astropy >= 6
