### 2025-06-03 SWIFTSIM/SOAP#155 @nvanalfen
@robjmcgibbon Perhaps this is my setup being difficult again, but I found when I run the compute halo membership and compute halo properties following the README, it doesn't work. I have run them, so they work for me just fine, but here's what I encountered.

# Precursor

First off, I just want to emphasize that I can get these things to run and work. My only issue is the inability to use the README code as written to do so.

# Computing halo membership for particles in the snapshot

Where you wrote (Replacing the FLAMINGO example with the test data small volume):

```
snapnum=0018
sim=test_data
mpirun python soap-group-membership \
    --sim-name=${sim} --snap-nr=${snapnum} tests/small_volume.yml
```

I got the error:

> soap-group-membership': [Errno 2] No such file or directory

It worked perfectly fine if I either
- replaced `soap-group-membership` with `SOAP/group_membership.py`, OR
- removed `python` from the command, instead using just `mpirun soap-group-membership ...`

# Computing halo properties

The same issues here. Using the suggested method gives me:
> soap-compute-halo-properties': [Errno 2] No such file or directory

Which is fixed by either using the python file name instead, or by eliminating the `python` call in the command.
