### 2025-06-04 9c06817
Generate pdf with all properties
