+++ SOAP/property_table.py
@@ -840,7 +840,7 @@ class PropertyTable:
             a_scale_exponent=0,
         ),
         "DtoTstar_luminosity_weighted_luminosity_ratio": Property(
-            name="DiscToTotalLuminosityRatio_LuminosityWeighted",
+            name="DiscToTotalLuminosityRatioLuminosityWeighted",
             shape=9,
             dtype=np.float32,
             unit="dimensionless",
@@ -872,7 +872,7 @@ class PropertyTable:
             a_scale_exponent=0,
         ),
         "DtoTstar_luminosity_weighted_mass_ratio": Property(
-            name="DiscToTotalMassRatio_LuminosityWeighted",
+            name="DiscToTotalMassRatioLuminosityWeighted",
             shape=9,
             dtype=np.float32,
             unit="dimensionless",
@@ -889,7 +889,7 @@ class PropertyTable:
             a_scale_exponent=0,
         ),
         "kappa_corot_star_luminosity_weighted": Property(
-            name="KappaCorotStars_LuminosityWeighted",
+            name="KappaCorotStarsLuminosityWeighted",
             shape=9,
             dtype=np.float32,
             unit="dimensionless",
@@ -906,7 +906,7 @@ class PropertyTable:
             a_scale_exponent=0,
         ),
         "Lstar_luminosity_weighted": Property(
-            name="AngularMomentumStars_LuminosityWeighted",
+            name="AngularMomentumStarsLuminosityWeighted",
             shape=27,  # 3D vector for each of the 9 GAMA bands
             dtype=np.float32,
             unit="snap_mass*snap_length**2/snap_time",
@@ -2053,8 +2053,8 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=0,
         ),
-        "ProjectedStellarInertiaTensor_LuminosityWeighted": Property(
-            name="ProjectedStellarInertiaTensor_LuminosityWeighted",
+        "ProjectedStellarInertiaTensorLuminosityWeighted": Property(
+            name="ProjectedStellarInertiaTensorLuminosityWeighted",
             shape=27,
             dtype=np.float32,
             unit="snap_length**2",
@@ -2069,8 +2069,8 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=2,
         ),
-        "ProjectedStellarInertiaTensorReduced_LuminosityWeighted": Property(
-            name="ProjectedStellarInertiaTensorReduced_LuminosityWeighted",
+        "ProjectedStellarInertiaTensorReducedLuminosityWeighted": Property(
+            name="ProjectedStellarInertiaTensorReducedLuminosityWeighted",
             shape=27,
             dtype=np.float32,
             unit="dimensionless",
@@ -2085,8 +2085,8 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=0,
         ),
-        "ProjectedStellarInertiaTensorNoniterative_LuminosityWeighted": Property(
-            name="ProjectedStellarInertiaTensorNoniterative_LuminosityWeighted",
+        "ProjectedStellarInertiaTensorNoniterativeLuminosityWeighted": Property(
+            name="ProjectedStellarInertiaTensorNoniterativeLuminosityWeighted",
             shape=27,
             dtype=np.float32,
             unit="snap_length**2",
@@ -2101,8 +2101,8 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=2,
         ),
-        "ProjectedStellarInertiaTensorReducedNoniterative_LuminosityWeighted": Property(
-            name="ProjectedStellarInertiaTensorReducedNoniterative_LuminosityWeighted",
+        "ProjectedStellarInertiaTensorReducedNoniterativeLuminosityWeighted": Property(
[…]
