### 2025-06-04 SWIFTSIM/SOAP#155 @robjmcgibbon
Thanks @nvanalfen, I've pushed a couple of changes to address your latest comments.
- I've updated the README to replace `soap-group-membership` with `SOAP/group_membership.py`.
- I've added a line to the README that makes clear the `tests/small_volume.yml` is the parameter file used for the small volume test
- I've added the option to generate the docs outputting all the possible properties. This can be done by running `python SOAP/property_table.py` (e.g. not passing a parameter file or snapshot), and then generating the pdf with pdflatex.
