### 2025-06-10 79fa265
Set lower limit
