+++ SOAP/particle_selection/subhalo_properties.py
@@ -341,7 +341,11 @@ class SubhaloParticleData:
         """
         if self.Ngas == 0:
             return None
-        return self.get_dataset("PartType0/TotalDustMassFractions")[self.gas_mask_all]
+        mass_frac = self.get_dataset("PartType0/TotalDustMassFractions")[
+            self.gas_mask_all
+        ]
+        mass_frac[mass_frac < 10 ** (-10)] = 0 * unyt.dimensionless
+        return mass_frac
 
     @lazy_property
     def Mdm(self) -> unyt.unyt_quantity:
