+++ parameter_files/COLIBRE_HYBRID.yml
@@ -35,7 +35,7 @@ ApertureProperties:
     AngularMomentumDarkMatter: general
     AngularMomentumGas: general
     AngularMomentumStars: general
-    AngularMomentumStars_LuminosityWeighted: false
+    AngularMomentumStarsLuminosityWeighted: false
     AtomicHydrogenMass: true
     BlackHolesDynamicalMass: true
     BlackHolesLastEventScalefactor: true
@@ -45,6 +45,7 @@ ApertureProperties:
     CentreOfMass: true
     CentreOfMassVelocity: true
     DarkMatterMass: true
+    DarkMatterCentreOfMass: true
     DarkMatterVelocityDispersionMatrix: general
     DiffuseCarbonMass:
       snapshot: true
@@ -63,8 +64,8 @@ ApertureProperties:
       snipshot: false
     DiscToTotalGasMassFraction: general
     DiscToTotalStellarMassFraction: general
-    DiscToTotalMassRatio_LuminosityWeighted: false
-    DiscToTotalLuminosityRatio_LuminosityWeighted: false
+    DiscToTotalMassRatioLuminosityWeighted: false
+    DiscToTotalLuminosityRatioLuminosityWeighted: false
     DustMass: true
     DustGraphiteMass:
       snapshot: true
@@ -131,7 +132,7 @@ ApertureProperties:
     KappaCorotBaryons: general
     KappaCorotGas: general
     KappaCorotStars: general
-    KappaCorotStars_LuminosityWeighted: false
+    KappaCorotStarsLuminosityWeighted: false
     KineticEnergyGas: general
     KineticEnergyStars: general
     LinearMassWeightedDiffuseOxygenOverHydrogenOfGas:
@@ -222,22 +223,22 @@ ApertureProperties:
     GasInertiaTensor: false
     DarkMatterInertiaTensor: false
     StellarInertiaTensor: false
-    StellarInertiaTensor_LuminosityWeighted: false
+    StellarInertiaTensorLuminosityWeighted: false
     TotalInertiaTensorReduced: false
     GasInertiaTensorReduced: false
     DarkMatterInertiaTensorReduced: false
     StellarInertiaTensorReduced: false
-    StellarInertiaTensorReduced_LuminosityWeighted: false
+    StellarInertiaTensorReducedLuminosityWeighted: false
     TotalInertiaTensorNoniterative: general
     GasInertiaTensorNoniterative: general
     DarkMatterInertiaTensorNoniterative: general
     StellarInertiaTensorNoniterative: general
-    StellarInertiaTensorNoniterative_LuminosityWeighted: false
+    StellarInertiaTensorNoniterativeLuminosityWeighted: false
     TotalInertiaTensorReducedNoniterative: false
     GasInertiaTensorReducedNoniterative: false
     DarkMatterInertiaTensorReducedNoniterative: false
     StellarInertiaTensorReducedNoniterative: false
-    StellarInertiaTensorReducedNoniterative_LuminosityWeighted: false
+    StellarInertiaTensorReducedNoniterativeLuminosityWeighted: false
     StarFormationRate: true
     AveragedStarFormationRate:
       snapshot: true
@@ -268,6 +269,12 @@ ApertureProperties:
       snipshot: false
     StellarCentreOfMassVelocity: true
   variations:
+    exclusive_100_pc:
+      inclusive: false
+      radius_in_kpc: 0.1
+    exclusive_300_pc:
+      inclusive: false
+      radius_in_kpc: 0.3
     exclusive_1_kpc:
       inclusive: false
       radius_in_kpc: 1.0
@@ -290,6 +297,12 @@ ApertureProperties:
[…]
