### 2025-06-10 f648a07
Skip assert for ranks with no halos
