### 2025-06-10 c9136ec
Add sbatch script for calculating FoF radii (#162)
* Add sbatch script for FOF radii

* Skip assert for ranks with no halos

* Nicer check
