### 2025-06-12 dd8476e
Add --check
