### 2025-06-18 69db816
More logging
