### 2025-06-23 19d0ad5
Exclude COLIBRE core
