+++ SOAP/property_table.py
@@ -2721,6 +2721,30 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=None,
         ),
+        "XRayLuminosityNoSat": Property(
+            name="XRayLuminosityNoSat",
+            shape=3,
+            dtype=np.float64,
+            unit="snap_mass*snap_length**2/snap_time**3",
+            description="Total observer-frame Xray luminosity in three bands. Excludes particles bound to satellites",
+            lossy_compression_filter="DMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType0/XrayLuminosities"],
+            output_physical=True,
+            a_scale_exponent=None,
+        ),
+        "XRayLuminosityCoreExcisionNoSat": Property(
+            name="XRayLuminosityCoreExcisionNoSat",
+            shape=3,
+            dtype=np.float64,
+            unit="snap_mass*snap_length**2/snap_time**3",
+            description="Total observer-frame Xray luminosity in three bands. Excludes gas in the inner {core_excision}, and excludes particles bound to satellites",
+            lossy_compression_filter="DMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType0/XrayLuminosities"],
+            output_physical=True,
+            a_scale_exponent=None,
+        ),
         "Xraylum_restframe_core_excision": Property(
             name="XRayLuminosityInRestframeCoreExcision",
             shape=3,
