+++ parameter_files/COLIBRE_HYBRID.yml
@@ -35,7 +35,6 @@ ApertureProperties:
     AngularMomentumDarkMatter: general
     AngularMomentumGas: general
     AngularMomentumStars: general
-    AngularMomentumStarsLuminosityWeighted: false
     AtomicHydrogenMass: true
     BlackHolesDynamicalMass: true
     BlackHolesLastEventScalefactor: true
@@ -64,8 +63,10 @@ ApertureProperties:
       snipshot: false
     DiscToTotalGasMassFraction: general
     DiscToTotalStellarMassFraction: general
-    DiscToTotalMassRatioLuminosityWeighted: false
     DiscToTotalLuminosityRatioLuminosityWeighted: false
+    DiscToTotalMassRatioLuminosityWeighted: false
+    KappaCorotStarsLuminosityWeighted: false
+    AngularMomentumStarsLuminosityWeighted: false
     DustMass: true
     DustGraphiteMass:
       snapshot: true
@@ -119,7 +120,7 @@ ApertureProperties:
       snipshot: false
     GasMassInColdDenseGas: true
     GasTemperature: true
-    GasTemperatureWithoutRecentAGNHeating: true
+    GasTemperatureWithoutRecentAGNHeating: false
     GasVelocityDispersionMatrix: general
     HalfMassRadiusBaryons: general
     HalfMassRadiusDarkMatter: general
@@ -132,7 +133,6 @@ ApertureProperties:
     KappaCorotBaryons: general
     KappaCorotGas: general
     KappaCorotStars: general
-    KappaCorotStarsLuminosityWeighted: false
     KineticEnergyGas: general
     KineticEnergyStars: general
     LinearMassWeightedDiffuseOxygenOverHydrogenOfGas:
@@ -219,26 +219,6 @@ ApertureProperties:
     NumberOfDarkMatterParticles: true
     NumberOfGasParticles: true
     NumberOfStarParticles: true
-    TotalInertiaTensor: false
-    GasInertiaTensor: false
-    DarkMatterInertiaTensor: false
-    StellarInertiaTensor: false
-    StellarInertiaTensorLuminosityWeighted: false
-    TotalInertiaTensorReduced: false
-    GasInertiaTensorReduced: false
-    DarkMatterInertiaTensorReduced: false
-    StellarInertiaTensorReduced: false
-    StellarInertiaTensorReducedLuminosityWeighted: false
-    TotalInertiaTensorNoniterative: general
-    GasInertiaTensorNoniterative: general
-    DarkMatterInertiaTensorNoniterative: general
-    StellarInertiaTensorNoniterative: general
-    StellarInertiaTensorNoniterativeLuminosityWeighted: false
-    TotalInertiaTensorReducedNoniterative: false
-    GasInertiaTensorReducedNoniterative: false
-    DarkMatterInertiaTensorReducedNoniterative: false
-    StellarInertiaTensorReducedNoniterative: false
-    StellarInertiaTensorReducedNoniterativeLuminosityWeighted: false
     StarFormationRate: true
     AveragedStarFormationRate:
       snapshot: true
@@ -390,12 +370,12 @@ ProjectedApertureProperties:
     ProjectedGasInertiaTensorNoniterative: general
     ProjectedGasInertiaTensorReducedNoniterative: false
     ProjectedStellarInertiaTensor: false
-    ProjectedStellarInertiaTensorLuminosityWeighted: false
     ProjectedStellarInertiaTensorReduced: false
-    ProjectedStellarInertiaTensorReducedLuminosityWeighted: false
     ProjectedStellarInertiaTensorNoniterative: general
-    ProjectedStellarInertiaTensorNoniterativeLuminosityWeighted: false
[…]

[…]
