+++ SOAP/property_table.py
@@ -1277,6 +1277,40 @@ class PropertyTable:
             output_physical=False,
             a_scale_exponent=1,
         ),
+        "HalfMassRadiusAtomicHydrogen": Property(
+            name="HalfMassRadiusAtomicHydrogen",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_length",
+            description="Atomic hydrogen half mass radius.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType0/Coordinates",
+                "PartType0/Masses",
+                "PartType0/SpeciesFractions",
+                "PartType0/ElementMassFractions",
+            ],
+            output_physical=False,
+            a_scale_exponent=1,
+        ),
+        "HalfMassRadiusMolecularHydrogen": Property(
+            name="HalfMassRadiusMolecularHydrogen",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_length",
+            description="Molecular hydrogen half mass radius.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType0/Coordinates",
+                "PartType0/Masses",
+                "PartType0/SpeciesFractions",
+                "PartType0/ElementMassFractions",
+            ],
+            output_physical=False,
+            a_scale_exponent=1,
+        ),
         "HalfMassRadiusGas": Property(
             name="HalfMassRadiusGas",
             shape=1,
