### 2025-06-23 898d22a
COLIBRE core excision (#167)
* Exclude COLIBRE core

* HalfMassRadii H
