+++ SOAP/particle_selection/projected_aperture_properties.py
@@ -1509,6 +1509,27 @@ class SingleProjectionProjectedApertureParticleData:
         )
 
     @lazy_property
+    def HalfMassRadiusAtomicHydrogen(self) -> unyt.unyt_quantity:
+        """
+        Half-mass radius of the atomic hydrogen.
+        """
+        if self.Ngas == 0:
+            return None
+        return get_half_mass_radius(
+            self.proj_radius[self.proj_type == 0], self.gas_mass_HI, self.AtomicHydrogenMass
+        )
+    @lazy_property
+    def HalfMassRadiusMolecularHydrogen(self) -> unyt.unyt_quantity:
+        """
+        Half-mass radius of the molecular hydrogen.
+        """
+        if self.Ngas == 0:
+            return None
+        return get_half_mass_radius(
+            self.proj_radius[self.proj_type == 0], self.gas_mass_H2, self.MolecularHydrogenMass
+        )
+
+    @lazy_property
     def HalfMassRadiusDM(self) -> unyt.unyt_quantity:
         """
         Half mass radius of dark matter.
@@ -1588,6 +1609,8 @@ class ProjectedApertureProperties(HaloProperty):
         "StellarLuminosity": False,
         "HalfMassRadiusGas": False,
         "HalfMassRadiusDust": False,
+        "HalfMassRadiusAtomicHydrogen": False,
+        "HalfMassRadiusMolecularHydrogen": False,
         "HalfMassRadiusDM": False,
         "HalfMassRadiusStar": False,
         "HalfLightRadiusStar": False,
