+++ SOAP/property_table.py
@@ -1277,6 +1277,40 @@ class PropertyTable:
             output_physical=False,
             a_scale_exponent=1,
         ),
+        "HalfMassRadiusAtomicHydrogen": Property(
+            name="HalfMassRadiusAtomicHydrogen",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_length",
+            description="Atomic hydrogen half mass radius.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType0/Coordinates",
+                "PartType0/Masses",
+                "PartType0/SpeciesFractions",
+                "PartType0/ElementMassFractions",
+            ],
+            output_physical=False,
+            a_scale_exponent=1,
+        ),
+        "HalfMassRadiusMolecularHydrogen": Property(
+            name="HalfMassRadiusMolecularHydrogen",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_length",
+            description="Molecular hydrogen half mass radius.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType0/Coordinates",
+                "PartType0/Masses",
+                "PartType0/SpeciesFractions",
+                "PartType0/ElementMassFractions",
+            ],
+            output_physical=False,
+            a_scale_exponent=1,
+        ),
         "HalfMassRadiusGas": Property(
             name="HalfMassRadiusGas",
             shape=1,
@@ -2718,6 +2752,30 @@ class PropertyTable:
             lossy_compression_filter="DMantissa9",
             dmo_property=False,
             particle_properties=["PartType0/XrayLuminosities", "PartType0/Coordinates"],
+            output_physical=True,
+            a_scale_exponent=None,
+        ),
+        "XRayLuminosityNoSat": Property(
+            name="XRayLuminosityNoSat",
+            shape=3,
+            dtype=np.float64,
+            unit="snap_mass*snap_length**2/snap_time**3",
+            description="Total observer-frame Xray luminosity in three bands. Excludes particles bound to satellites",
+            lossy_compression_filter="DMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType0/XrayLuminosities"],
+            output_physical=True,
+            a_scale_exponent=None,
+        ),
+        "XRayLuminosityCoreExcisionNoSat": Property(
+            name="XRayLuminosityCoreExcisionNoSat",
+            shape=3,
+            dtype=np.float64,
+            unit="snap_mass*snap_length**2/snap_time**3",
+            description="Total observer-frame Xray luminosity in three bands. Excludes gas in the inner {core_excision}, and excludes particles bound to satellites",
+            lossy_compression_filter="DMantissa9",
+            dmo_property=False,
+            particle_properties=["PartType0/XrayLuminosities"],
             output_physical=True,
             a_scale_exponent=None,
         ),
