+++ parameter_files/COLIBRE_THERMAL.yml
@@ -120,12 +120,14 @@ ApertureProperties:
       snipshot: false
     GasMassInColdDenseGas: true
     GasTemperature: true
-    GasTemperatureWithoutRecentAGNHeating: true
+    GasTemperatureWithoutRecentAGNHeating: false
     GasVelocityDispersionMatrix: general
     HalfMassRadiusBaryons: general
     HalfMassRadiusDarkMatter: general
     HalfMassRadiusGas: general
     HalfMassRadiusDust: general
+    HalfMassRadiusAtomicHydrogen: true
+    HalfMassRadiusMolecularHydrogen: true
     HalfMassRadiusStars: true
     HalfLightRadiusStars: false
     HeliumMass: true
@@ -331,6 +333,8 @@ ProjectedApertureProperties:
     HalfMassRadiusDarkMatter: general
     HalfMassRadiusGas: general
     HalfMassRadiusDust: general
+    HalfMassRadiusAtomicHydrogen: true
+    HalfMassRadiusMolecularHydrogen: true
     HalfMassRadiusStars: true
     HalfLightRadiusStars: false
     HeliumMass: true
@@ -429,9 +433,7 @@ SOProperties:
     ComptonY:
       snapshot: true
       snipshot: false
-    ComptonYWithoutRecentAGNHeating:
-      snapshot: true
-      snipshot: false
+    ComptonYWithoutRecentAGNHeating: false
     Concentration: true
     ConcentrationUnsoftened: true
     DarkMatterConcentration: general
@@ -454,10 +456,16 @@ SOProperties:
       snipshot: false
     GasComptonYTemperature: false
     GasComptonYTemperatureWithoutRecentAGNHeating: false
+    GasComptonYTemperatureCoreExcision: false
+    GasComptonYTemperatureWithoutRecentAGNHeatingCoreExcision: false
     GasTemperature: true
-    GasTemperatureWithoutCoolGas: true
-    GasTemperatureWithoutCoolGasAndRecentAGNHeating: true
-    GasTemperatureWithoutRecentAGNHeating: true
+    GasTemperatureWithoutCoolGas: false
+    GasTemperatureWithoutCoolGasAndRecentAGNHeating: false
+    GasTemperatureWithoutRecentAGNHeating: false
+    GasTemperatureWithoutCoolGasCoreExcision: false
+    GasTemperatureCoreExcision: true
+    GasTemperatureWithoutCoolGasAndRecentAGNHeatingCoreExcision: false
+    GasTemperatureWithoutRecentAGNHeatingCoreExcision: false
     HotGasMass: true
     KineticEnergyGas: general
     KineticEnergyStars: general
@@ -529,21 +537,37 @@ SOProperties:
     XRayLuminosity:
       snapshot: true
       snipshot: false
-    XRayLuminosityWithoutRecentAGNHeating:
+    XRayLuminosityWithoutRecentAGNHeating: false
+    XRayLuminosityCoreExcision:
       snapshot: true
       snipshot: false
+    XRayLuminosityNoSat:
+      snapshot: true
+      snipshot: false
+    XRayLuminosityCoreExcisionNoSat:
+      snapshot: true
+      snipshot: false
+    XRayLuminosityWithoutRecentAGNHeatingCoreExcision: false
     XRayLuminosityInRestframe: false
     XRayLuminosityInRestframeWithoutRecentAGNHeating: false
+    XRayLuminosityInRestframeCoreExcision: false
+    XRayLuminosityInRestframeWithoutRecentAGNHeatingCoreExcision: false
     XRayPhotonLuminosity:
       snapshot: true
       snipshot: false
-    XRayPhotonLuminosityWithoutRecentAGNHeating:
[…]

[…]
