### 2025-06-23 ad1ceed
Add parameter file option to read energy
