### 2025-06-23 b9fa644
Binding energies and Vmax calculations (#132)
* Read binding energy from HBT

* Output TotalBindingEnergy

* Vmax for aperture properties

* Squash bug and format

* Stop ranks hanging when reading binding energy

* Read SnapshotIndexOfLastIsolation

* Read in potential energy instead of binding

* Remove duplicate line

* Use vCoM from all particles

* Move subhalo_properties check for Ntot

* Update COLIBRE test

* Remove Ntot warning

* Add documentation of SpecificPotentialEnergies

* Add parameter file option to read energy

* Review changes
