+++ documentation/footnote_Ekin.tex
@@ -1,3 +1,1 @@
-\paragraph{$^{$FOOTNOTE_NUMBER$}$The kinetic energy}\label{footnote:$FOOTNOTE_NUMBER$} of the gas and stars is computed using the same relative 
-velocities as used for other properties, i.e. relative to the centre of mass velocity of the gas and stars 
-respectively.
+\paragraph{$^{$FOOTNOTE_NUMBER$}$The kinetic energy}\label{footnote:$FOOTNOTE_NUMBER$} is computed using the velocities relative to the centre of mass velocity of all the particles in the aperture. The Hubble flow is included when computing the velocities.

[…]
