### 2025-06-27 SWIFTSIM/SOAP#168 @dnelson86
small_volume.yml
Just a note, trying `run_small_volume.sh` I saw

```
KeyError: 'read_potential_energies'
  File "/vera/u/dnelson/SOAP/SOAP/group_membership.py", line 180, in main
```

that I had to fix by adding `read_potential_energies:False` to the parameter file.

(Setting it to true leads eventually to a h5py key error:)

```
WARNING: failed to set lustre striping on ./output/membership_0018
Traceback (most recent call last):
  File "/vera/u/dnelson/SOAP/SOAP/group_membership.py", line 369, in <module>
    main()
  File "/vera/u/dnelson/SOAP/SOAP/group_membership.py", line 228, in main
    read_hbtplus.read_hbtplus_groupnr(
  File "/vera/u/dnelson/SOAP/SOAP/catalogue_readers/read_hbtplus.py", line 70, in read_hbtplus_groupnr
    potential_energies.append(infile["PotentialEnergies"][...])
                              ~~~~~~^^^^^^^^^^^^^^^^^^^^^
  File "h5py/_objects.pyx", line 56, in h5py._objects.with_phil.wrapper
  File "h5py/_objects.pyx", line 57, in h5py._objects.with_phil.wrapper
  File "/u/dnelson/.local/envs/soap/lib/python3.12/site-packages/h5py/_hl/group.py", line 360, in __getitem__
    oid = h5o.open(self.id, self._e(name), lapl=self._lapl)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "h5py/_objects.pyx", line 56, in h5py._objects.with_phil.wrapper
  File "h5py/_objects.pyx", line 57, in h5py._objects.with_phil.wrapper
  File "h5py/h5o.pyx", line 257, in h5py.h5o.open
KeyError: "Unable to open object (object 'PotentialEnergies' doesn't exist)"
```
