### 2025-06-27 #8252 @dnelson86
@robjmcgibbon Apologies for the delay, I'm now taking a look at SOAP, and it looks fantastic.

In terms of the actual paper, it seems that it might be appropriate to mention other (public) packages that exist and have similar functionality? This would better place SOAP in context.

In particular, [CAESAR](https://caesar.readthedocs.io/en/latest/index.html) comes to mind (from the SIMBA world). In addition, there is [tangos](https://github.com/pynbody/tangos), with some similarities. This isn't at all exhaustive. In particular, I'm not sure if something similar exists in the world of Ramses/HorizonAGN/Sphinx.

I leave it up to you to add such a sentence to the draft, just a thought.
