### 2025-06-27 SWIFTSIM/SOAP#169 @dnelson
inertia tensors FloatingPointError
Trying to run `parameter_files/FLAMINGO.yml` (with minor modifications) on FLAMINGO/L1000N0900/HYDRO_FIDUCIAL snapshot 77, I run into:

```
Object with HaloCatalogueIndex=1071312 dimensionless encountered an error
Traceback (most recent call last):
  File "/vera/u/dnelson/SOAP/SOAP/compute_halo_properties.py", line 617, in compute_halo_properties
    metadata = task_queue.execute_tasks(
               ^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/vera/u/dnelson/SOAP/SOAP/core/task_queue.py", line 177, in execute_tasks
    result.append(task(*args))
                  ^^^^^^^^^^^
  File "/vera/u/dnelson/SOAP/SOAP/core/chunk_tasks.py", line 316, in __call__
    total_time, task_time, nr_left, nr_done, min_free_mem_gb = process_halos(
                                                               ^^^^^^^^^^^^^^
  File "/vera/u/dnelson/SOAP/SOAP/core/halo_tasks.py", line 373, in process_halos
    halo_result, timings = process_single_halo(
                           ^^^^^^^^^^^^^^^^^^^^
  File "/vera/u/dnelson/SOAP/SOAP/core/halo_tasks.py", line 127, in process_single_halo
    halo_prop.calculate(
  File "/vera/u/dnelson/SOAP/SOAP/particle_selection/projected_aperture_properties.py", line 1926, in calculate
    val = getattr(proj_part_props, name)
          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/vera/u/dnelson/SOAP/SOAP/core/lazy_properties.py", line 53, in _lazy_property
    setattr(self, attr_name, fn(self))
                             ^^^^^^^^
  File "/vera/u/dnelson/SOAP/SOAP/particle_selection/projected_aperture_properties.py", line 1093, in ProjectedStellarInertiaTensorReducedLuminosityWeighted
    return self.stellar_inertia_tensor_luminosity_weighted(reduced=True)
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/vera/u/dnelson/SOAP/SOAP/particle_selection/projected_aperture_properties.py", line 1012, in stellar_inertia_tensor_luminosity_weighted
    return get_projected_inertia_tensor_luminosity_weighted(
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/vera/u/dnelson/SOAP/SOAP/property_calculation/inertia_tensors.py", line 400, in get_projected_inertia_tensor_luminosity_weighted
    flattened_inertia_tensor_i_band = get_weighted_projected_inertia_tensor(
                                      ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/vera/u/dnelson/SOAP/SOAP/property_calculation/inertia_tensors.py", line 305, in get_weighted_projected_inertia_tensor
    q = np.sqrt(eig_val[0] / eig_val[1])
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
FloatingPointError: invalid value encountered in sqrt
```

Not sure if more information is required to reproduce, or if this is some mistake on my end?
