### 2025-06-27 #8252 @dnelson86
I have run a pipeline on the `./tests/run_small_volume.sh` and everything works well (see https://github.com/SWIFTSIM/SOAP/issues/168)

I am also slowly attempting a more manual run on `FLAMINGO/L1000N0900/HYDRO_FIDUCIAL`. During so I'm adding some minor issues, related to documentation, parameter files, and so on: (https://github.com/SWIFTSIM/SOAP/issues/169 and https://github.com/SWIFTSIM/SOAP/issues/170)

None of these are road blocks to the JOSS paper, which looks great and I'm happy to recommend for publication.
