### 2025-07-03 81f096e
Update paper.bib
