+++ documentation/joss_paper/paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{eagle,
        author = {{Schaye}, Joop and {Crain}, Robert A. and {Bower}, Richard G. and {Furlong}, Michelle and {Schaller}, Matthieu and {Theuns}, Tom and {Dalla Vecchia}, Claudio and {Frenk}, Carlos S. and {McCarthy}, I.~G. and {Helly}, John C. and {Jenkins}, Adrian and {Rosas-Guevara}, Y.~M. and {White}, Simon D.~M. and {Baes}, Maarten and {Booth}, C.~M. and {Camps}, Peter and {Navarro}, Julio F. and {Qu}, Yan and {Rahmati}, Alireza and {Sawala}, Till and {Thomas}, Peter A. and {Trayford}, James},
         title = "{The EAGLE project: simulating the evolution and assembly of galaxies and their environments}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {methods: numerical, galaxies: evolution, galaxies: formation, cosmology: theory, Astrophysics - Astrophysics of Galaxies, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2015,
         month = jan,
@@ -20,7 +20,7 @@ archivePrefix = {arXiv},
 @ARTICLE{flamingo,
        author = {{Schaye}, Joop and {Kugel}, Roi and {Schaller}, Matthieu and {Helly}, John C. and {Braspenning}, Joey and {Elbers}, Willem and {McCarthy}, Ian G. and {van Daalen}, Marcel P. and {Vandenbroucke}, Bert and {Frenk}, Carlos S. and {Kwan}, Juliana and {Salcido}, Jaime and {Bah{\'e}}, Yannick M. and {Borrow}, Josh and {Chaikin}, Evgenii and {Hahn}, Oliver and {Hu{\v{s}}ko}, Filip and {Jenkins}, Adrian and {Lacey}, Cedric G. and {Nobels}, Folkert S.~J.},
         title = "{The FLAMINGO project: cosmological hydrodynamical simulations for large-scale structure and galaxy cluster surveys}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {methods: numerical, galaxies: clusters: general, galaxies: formation, cosmology: theory, large-scale structure of Universe, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2023,
         month = dec,
@@ -39,7 +39,7 @@ archivePrefix = {arXiv},
 @ARTICLE{flamingoCalibration,
        author = {{Kugel}, Roi and {Schaye}, Joop and {Schaller}, Matthieu and {Helly}, John C. and {Braspenning}, Joey and {Elbers}, Willem and {Frenk}, Carlos S. and {McCarthy}, Ian G. and {Kwan}, Juliana and {Salcido}, Jaime and {van Daalen}, Marcel P. and {Vandenbroucke}, Bert and {Bah{\'e}}, Yannick M. and {Borrow}, Josh and {Chaikin}, Evgenii and {Hu{\v{s}}ko}, Filip and {Jenkins}, Adrian and {Lacey}, Cedric G. and {Nobels}, Folkert S.~J. and {Vernon}, Ian},
         title = "{FLAMINGO: calibrating large cosmological hydrodynamical simulations with machine learning}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the Royal Astronomical Society},
      keywords = {methods: numerical, methods: statistical, galaxies: clusters: general, galaxies: formation, large-scale structure of Universe, cosmology: theory, Astrophysics - Cosmology and Nongalactic Astrophysics},
[…]
