### 2025-07-03 #8252 @ivastar
@robjmcgibbon I opened a [pull request](https://github.com/SWIFTSIM/SOAP/pull/171) with some editorial changes. Please review and merge as you see fit.

It would be great if you can also fold in the suggestions from @dnelson86 above about mentioning some other existing packages.

I would also request that you acknowledge any key dependancies (maybe in the acknowledgements) with the appropriate citations.

Let me know if you have questions.
