### 2025-07-11 a7a852c
Editorial changes (#171)
* Editorial changes.

* Update paper.bib

* A couple of more changes

* Update title capitalization
