+++ documentation/joss_paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'SOAP: A Python package for calculating the properties of galaxies and halos formed in cosmological simulations'
+title: 'SOAP: A Python Package for Calculating the Properties of Galaxies and Halos Formed in Cosmological Simulations'
 tags:
   - Python
   - astronomy
@@ -31,98 +31,98 @@ bibliography: paper.bib
 Cosmological simulations model the evolution of dark matter and baryons under 
 the influence of gravitational and hydrodynamic forces. Beginning at 
 high redshift, they capture the hierarchical formation of structures, where smaller 
-structures form first and later merge into larger ones. These simulations incorporate
-hydrodynamics to evolve the gas, and contain a number of subgrid prescriptions
-for modelling other important physical processes such as star formation. Once a 
-simulation has concluded, a halo-finding algorithm is used to identify bound 
-structures (subhalos) within the resulting particle distribution.
+structures form first and later merge into larger ones.These simulations incorporate
+hydrodynamics to evolve the gas and include a number of subgrid prescriptions
+for modelling important physical processes, such as star formation. Once the simulation has concluded, a halo-finding algorithm is used to identify bound
+structures (subhaloes) within the resulting particle distribution.
 
-Here we introduce SOAP (Spherical Overdensity & Aperture Processor), 
-a Python package designed to compute halo and galaxy properties from simulations that
-have been post-processed with a halo finder. SOAP takes a subhalo catalogue as 
-input and calculates a wide array of properties for each object. Its output 
+Here, we introduce `SOAP` (Spherical Overdensity & Aperture Processor), a Python package designed to compute halo and galaxy properties from simulations that
+have been post-processed with a halo finder. `SOAP` takes a subhalo catalogue as
+input and calculates a wide array of properties for each object. Its output
 is compatible with the swiftsimio package [@swiftsimio], enabling seamless unit
-handling. SOAP has already been employed to generate halo catalogues for the 
+handling. `SOAP` has already been used to generate halo catalogues for the
 FLAMINGO simulation suite [@flamingo; @flamingoCalibration], which includes the largest cosmological 
-hydrodynamical simulation run to date, and these have been used in more than
-20 publications so far[^1].
+hydrodynamic simulation to date. These catalogues have been used in more than
+20 publications to date [^1].
 
-SOAP is hosted on GitHub at https://github.com/SWIFTSIM/SOAP. We strongly encourage
-contributions to SOAP by opening issues and submitting pull requests.
+`SOAP` is hosted on GitHub. We strongly encourage
+contributions to `SOAP`, such as opening issues and submitting pull requests.
 
-# Statement of need
+# Statement of Need
 
-Modern galaxy simulations are often analysed by a large number of people. However,
-due to the enormous volume of data generated, it is often impractical for individuals
-to compute the specific properties they need by themselves. SOAP addresses this challenge
-by producing comprehensive catalogues with a wide range of properties that can be
-shared across the community. Given the substantial data volume, it is essential
-for the output to be processed in parallel. SOAP achieves this through the mpi4py library
-[@mpi4py], enabling it to scale efficiently across multiple compute nodes.
-Additionally, SOAP is designed to handle subvolumes of the simulation independently,
-allowing large simulations to be processed sequentially if required. This approach reduces the need
-for high-memory compute nodes, making it feasible to process
-simulation outputs without a large number of high-memory resources.
+Modern galaxy simulations are often analysed by a large number of researchers. However,
+due to the large volume of data generated, it is often impractical for individual users
+to compute the specific properties they require independently. `SOAP` addresses this challenge
+by producing comprehensive catalogues containing a wide range of properties that can be
+shared across the community. 
+
+Given the substantial volume of data, it is essential 
[…]
