### 2025-07-11 c0ca8e8
Set read_potential_energy default to false
