### 2025-07-11 65d718f
Add more references
