+++ documentation/joss_paper/paper.bib
@@ -196,3 +196,91 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@article{unyt,
+  doi = {10.21105/joss.00809},
+  url = {https://doi.org/10.21105/joss.00809},
+  year  = {2018},
+  month = {aug},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {28},
+  pages = {809},
+  author = {Nathan J. Goldbaum and John A. ZuHone and Matthew J. Turk and Kacper Kowalik and Anna L. Rosen},
+  title = {unyt: Handle,  manipulate,  and convert data with units in Python},
+  journal = {Journal of Open Source Software}
+}
+
+@Article{numpy,
+         title = {Array programming with {NumPy}},
+        author = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                 Travis E. Oliphant},
+          year = {2020},
+         month = sep,
+       journal = {Nature},
+        volume = {585},
+        number = {7825},
+         pages = {357--362},
+           doi = {10.1038/s41586-020-2649-2},
+     publisher = {Springer Science and Business Media {LLC}},
+           url = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+@ARTICLE{tangos,
+       author = {{Pontzen}, Andrew and {Tremmel}, Michael},
+        title = "{TANGOS: The Agile Numerical Galaxy Organization System}",
+      journal = {The Astrophysical Journal Supplement Series},
+     keywords = {methods: data analysis, methods: numerical, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2018,
+        month = aug,
+       volume = {237},
+       number = {2},
+          eid = {23},
+        pages = {23},
+          doi = {10.3847/1538-4365/aac832},
+archivePrefix = {arXiv},
+       eprint = {1803.00010},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2018ApJS..237...23P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{cosmic-profiles,
+       author = {{Dome}, Tibor},
+        title = "{CosmicProfiles: A Python package for radial profiling of finitely sampled dark matter halos and galaxies}",
+      journal = {Journal of Open Source Software},
+     keywords = {Cython, Python, cosmology, simulations, point clouds, astronomy},
+         year = 2023,
+        month = may,
+       volume = {8},
+       number = {85},
+          eid = {5008},
+        pages = {5008},
+          doi = {10.21105/joss.05008},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023JOSS....8.5008D},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
[…]
