### 2025-07-11 #8252 @robjmcgibbon
Thanks for the feedback @dnelson86 and @ivastar! Only got round to looking at it this morning since I've been on holiday.

I'm happy with the editorial changes that were suggested, so I've merged them. I  added a sentence with a couple of references to other packages, and added the key dependencies to the acknowledgements. Changes were in https://github.com/SWIFTSIM/SOAP/pull/124/commits/65d718fc6a600222c9e0a47f369b0f63747427b8

I've addressed one of the issues @dnelson86 raised which only required a very minor change. The other two will require a bit more time on my part, so I'd prefer to wait until I've caught up on other tasks from my holiday backlog before getting to them. However, if they need to be addressed before the paper can be published then I can prioritise them
