### 2025-07-12 bf21143
Read sorted HBT
