+++ SOAP/catalogue_readers/read_hbtplus.py
@@ -29,89 +29,112 @@ def read_hbtplus_groupnr(basename, read_potential_energies=False, registry=None)
     comm_rank = comm.Get_rank()
     comm_size = comm.Get_size()
 
-    # Find number of HBT output files
+    # Find number of HBT output files, and if we're dealing with sorted output
     if comm_rank == 0:
-        with h5py.File(hbt_filename(basename, 0), "r") as infile:
-            nr_files = int(infile["NumberOfFiles"][...])
+        if os.path.exists(hbt_filename(basename, 0)):
+            with h5py.File(hbt_filename(basename, 0), "r") as infile:
+                nr_files = int(infile["NumberOfFiles"][...])
+            sorted_file = False
+        elif os.path.exists(basename):
+            with h5py.File(basename, "r") as infile:
+                assert "Particles" in infile
+            nr_files = 1
+            sorted_file = True
+        else:
+            print(f"No HBT files found for basename {basename}")
+            comm.Abort()
     else:
         nr_files = None
+        sorted_file = None
     nr_files = comm.bcast(nr_files)
+    sorted_file = comm.bcast(sorted_file)
 
-    # Assign files to MPI ranks
-    files_per_rank = np.zeros(comm_size, dtype=int)
-    files_per_rank[:] = nr_files // comm_size
-    remainder = nr_files % comm_size
-    if remainder == 1:
-        files_per_rank[0] += 1
-    elif remainder > 1:
-        for i in range(remainder):
-            files_per_rank[int((comm_size - 1) * i / (remainder - 1))] += 1
-    assert np.sum(files_per_rank) == nr_files, f"{nr_files=}, {comm_size=}"
-    first_file_on_rank = np.cumsum(files_per_rank) - files_per_rank
+    # There are two different read routines. One is for the original HBTplus
+    # output. The second is for the "nice" output, where the subhalos are
+    # sorted by TrackId.
+    if not sorted_file:
+        # Assign files to MPI ranks
+        files_per_rank = np.zeros(comm_size, dtype=int)
+        files_per_rank[:] = nr_files // comm_size
+        remainder = nr_files % comm_size
+        if remainder == 1:
+            files_per_rank[0] += 1
+        elif remainder > 1:
+            for i in range(remainder):
+                files_per_rank[int((comm_size - 1) * i / (remainder - 1))] += 1
+        assert np.sum(files_per_rank) == nr_files, f"{nr_files=}, {comm_size=}"
+        first_file_on_rank = np.cumsum(files_per_rank) - files_per_rank
 
-    # Read in the halos from the HBT output:
-    # 'halos' will be an array of structs with the halo catalogue
-    # 'ids_bound' will be an array of particle IDs in halos, sorted by halo
-    # 'potential_energy' will be an array of the potential energy of each
-    #   particle, with the same order as 'ids_bound'
-    halos = []
-    ids_bound = []
-    if read_potential_energies:
-        potential_energies = []
+        halos = []
+        ids_bound = []
+        if read_potential_energies:
+            potential_energies = []
 
-    for file_nr in range(
-        first_file_on_rank[comm_rank],
-        first_file_on_rank[comm_rank] + files_per_rank[comm_rank],
-    ):
-        with h5py.File(hbt_filename(basename, file_nr), "r") as infile:
-            halos.append(infile["Subhalos"][...])
-            ids_bound.append(infile["SubhaloParticles"][...])
-            if read_potential_energies:
-                potential_energies.append(infile["PotentialEnergies"][...])
+        for file_nr in range(
+            first_file_on_rank[comm_rank],
+            first_file_on_rank[comm_rank] + files_per_rank[comm_rank],
+        ):
[…]
