### 2025-07-14 a4be752
Read sorted HBT (#172)
