### 2025-07-17 8cd45ad
Add node requirements
