### 2025-07-17 893b30a
Add script to recalculate COLIBRE FOF (#166)
* Add sbatch script for FOF radii

* Skip assert for ranks with no halos

* Nicer check

* More logging

* Add option to copy datasets

* Add node requirements
