### 2025-07-17 ed2c8ff
JOSS paper submission (#124)
* Record processing time for each halo

* Skip density check for satellites

* Use written chunks when restarting

* Store calculation time for each property

* Use lookup table for stellar ages

* Add links to footnotes

* Remove unneeded sort

* Squash sort bug

* Non iterative inertia tensors

* Catch io errors for chunk files

* Move filter to parameter file

* Remove VR FOFSubhalo halo type

* Remove GroupNr_all

* Descriptions for timing information

* Sort halos on chunk

* Report peak memory usage

* Copy values from smaller apertures

* Store filter metada
[…]
