+++ documentation/joss_paper/paper.md
@@ -1,0 +1,137 @@
+---
+title: 'SOAP: A Python Package for Calculating the Properties of Galaxies and Halos Formed in Cosmological Simulations'
+tags:
+  - Python
+  - astronomy
+authors:
+  - name: Robert McGibbon
+    affiliation: 1
+  - name: John C. Helly
+    affiliation: 2
+  - name: Joop Schaye 
+    affiliation: 1
+  - name: Matthieu Schaller
+    affiliation: "1, 3"
+  - name: Bert Vandenbroucke
+    affiliation: 1
+affiliations:
+ - index: 1
+   name: Leiden Observatory, Leiden University, PO Box 9513, 2300 RA Leiden, the Netherlands
+ - index: 2
+   name: Institute for Computational Cosmology, Department of Physics, University of Durham, South Road, Durham, DH1 3LE, UK
+ - index: 3
+   name: Lorentz Institute for Theoretical Physics, Leiden University, PO box 9506, 2300 RA Leiden, the Netherlands
+date: 24 January 2025
+bibliography: paper.bib
+
+---
+
+# Summary
+
+Cosmological simulations model the evolution of dark matter and baryons under 
+the influence of gravitational and hydrodynamic forces. Beginning at 
+high redshift, they capture the hierarchical formation of structures, where smaller 
+structures form first and later merge into larger ones.These simulations incorporate
+hydrodynamics to evolve the gas and include a number of subgrid prescriptions
+for modelling important physical processes, such as star formation. Once the simulation has concluded, a halo-finding algorithm is used to identify bound
+structures (subhaloes) within the resulting particle distribution.
+
+Here, we introduce `SOAP` (Spherical Overdensity & Aperture Processor), a Python package designed to compute halo and galaxy properties from simulations that
+have been post-processed with a halo finder. `SOAP` takes a subhalo catalogue as
+input and calculates a wide array of properties for each object. Its output
+is compatible with the swiftsimio package [@swiftsimio], enabling seamless unit
+handling. `SOAP` has already been used to generate halo catalogues for the
+FLAMINGO simulation suite [@flamingo; @flamingoCalibration], which includes the largest cosmological 
+hydrodynamic simulation to date. These catalogues have been used in more than
+20 publications to date [^1].
+
+`SOAP` is hosted on GitHub. We strongly encourage
+contributions to `SOAP`, such as opening issues and submitting pull requests.
+
+# Statement of Need
+
+Modern galaxy simulations are often analysed by a large number of researchers. However,
+due to the large volume of data generated, it is often impractical for individual users
+to compute the specific properties they require independently. `SOAP` addresses this challenge
+by producing comprehensive catalogues containing a wide range of properties that can be
+shared across the community.
+
+Given the substantial volume of data, it is essential for the output to be processed in parallel.
+`SOAP` achieves this using the mpi4py library
+[@mpi4py]. This enables `SOAP` to scale efficiently across multiple compute nodes.
+`SOAP` is also designed to handle subvolumes of the simulation independently, allowing 
+for large simulations to be processed sequentially if required. This approach reduces the need
+for high-memory computing nodes, making it possible to process
+simulation outputs without requiring a large number of high-memory resources.
+The ability to efficently process subhalos in parallel is a unique feature of `SOAP` when compared
+with other packages for computing galaxy properties [e.g. @cosmic-profiles; @tangos; @caesar].
+
+A large number of halo finders are used by the community
+to identify bound structures within simulation outputs. These employ a variety of methods 
+which can result in subhalo catalogues with significant differences [@hbt-herons].
+Therefore, it is important to be able to compare the results of various halo finders to 
+help quantify the uncertainty associated with structure finding.
+However, along with the different structure identification methods, the halo finder codes often 
[…]
