+++ documentation/joss_paper/paper.bib
@@ -1,0 +1,286 @@
+@ARTICLE{eagle,
+       author = {{Schaye}, Joop and {Crain}, Robert A. and {Bower}, Richard G. and {Furlong}, Michelle and {Schaller}, Matthieu and {Theuns}, Tom and {Dalla Vecchia}, Claudio and {Frenk}, Carlos S. and {McCarthy}, I.~G. and {Helly}, John C. and {Jenkins}, Adrian and {Rosas-Guevara}, Y.~M. and {White}, Simon D.~M. and {Baes}, Maarten and {Booth}, C.~M. and {Camps}, Peter and {Navarro}, Julio F. and {Qu}, Yan and {Rahmati}, Alireza and {Sawala}, Till and {Thomas}, Peter A. and {Trayford}, James},
+        title = "{The EAGLE project: simulating the evolution and assembly of galaxies and their environments}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+     keywords = {methods: numerical, galaxies: evolution, galaxies: formation, cosmology: theory, Astrophysics - Astrophysics of Galaxies, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2015,
+        month = jan,
+       volume = {446},
+       number = {1},
+        pages = {521-554},
+          doi = {10.1093/mnras/stu2058},
+archivePrefix = {arXiv},
+       eprint = {1407.7040},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015MNRAS.446..521S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@ARTICLE{flamingo,
+       author = {{Schaye}, Joop and {Kugel}, Roi and {Schaller}, Matthieu and {Helly}, John C. and {Braspenning}, Joey and {Elbers}, Willem and {McCarthy}, Ian G. and {van Daalen}, Marcel P. and {Vandenbroucke}, Bert and {Frenk}, Carlos S. and {Kwan}, Juliana and {Salcido}, Jaime and {Bah{\'e}}, Yannick M. and {Borrow}, Josh and {Chaikin}, Evgenii and {Hahn}, Oliver and {Hu{\v{s}}ko}, Filip and {Jenkins}, Adrian and {Lacey}, Cedric G. and {Nobels}, Folkert S.~J.},
+        title = "{The FLAMINGO project: cosmological hydrodynamical simulations for large-scale structure and galaxy cluster surveys}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+     keywords = {methods: numerical, galaxies: clusters: general, galaxies: formation, cosmology: theory, large-scale structure of Universe, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2023,
+        month = dec,
+       volume = {526},
+       number = {4},
+        pages = {4978-5020},
+          doi = {10.1093/mnras/stad2419},
+archivePrefix = {arXiv},
+       eprint = {2306.04024},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023MNRAS.526.4978S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@ARTICLE{flamingoCalibration,
[…]
