### 2025-07-17 #8252 @robjmcgibbon
Thanks @ivastar, I've merged the review branch into master and created a release. The DOI is https://zenodo.org/records/16034667, and I've checked the author list and license.
