### 2025-07-17 9cb8cf4
Update cold dense gas documentation
