+++ documentation/footnote_cold_dense.py
@@ -1,0 +1,2 @@
+\paragraph{$^{$FOOTNOTE_NUMBER$}$Cold dense gas particles}\label{footnote:$FOOTNOTE_NUMBER$} are particles with $T < 10^{$LOG_COLD_GAS_TEMP$}$K) and ($\frac{\rho}{m_{\mathrm{H}}} > 10^{$LOG_COLD_GAS_DENSITY$} \mathrm{cm}^{-3}$).
+
