+++ parameter_files/COLIBRE_THERMAL.yml
@@ -147,18 +147,10 @@ ApertureProperties:
     LinearMassWeightedOxygenOverHydrogenOfGas:
       snapshot: true
       snipshot: false
-    LinearMassWeightedDiffuseCarbonOverOxygenOfGas:
-      snapshot: true
-      snipshot: false
-    LinearMassWeightedNitrogenOverOxygenOfGas:
-      snapshot: true
-      snipshot: false
-    LinearMassWeightedCarbonOverOxygenOfGas:
-      snapshot: true
-      snipshot: false
-    LinearMassWeightedDiffuseNitrogenOverOxygenOfGas:
-      snapshot: true
-      snipshot: false
+    LinearMassWeightedDiffuseCarbonOverOxygenOfGas: false
+    LinearMassWeightedNitrogenOverOxygenOfGas: false
+    LinearMassWeightedCarbonOverOxygenOfGas: false
+    LinearMassWeightedDiffuseNitrogenOverOxygenOfGas: false
     LogarithmicMassWeightedDiffuseOxygenOverHydrogenOfAtomicGasHighLimit:
       snapshot: true
       snipshot: false
@@ -182,18 +174,10 @@ ApertureProperties:
     LogarithmicMassWeightedMagnesiumOverHydrogenOfStarsHighLimit: true
     LogarithmicMassWeightedIronOverHydrogenOfStarsLowLimit: true
     LogarithmicMassWeightedMagnesiumOverHydrogenOfStarsLowLimit: true
-    LogarithmicMassWeightedDiffuseNitrogenOverOxygenOfGasHighLimit:
-      snapshot: true
-      snipshot: false
-    LogarithmicMassWeightedDiffuseCarbonOverOxygenOfGasHighLimit:
-      snapshot: true
-      snipshot: false
-    LogarithmicMassWeightedDiffuseCarbonOverOxygenOfGasLowLimit:
-      snapshot: true
-      snipshot: false
-    LogarithmicMassWeightedDiffuseNitrogenOverOxygenOfGasLowLimit:
-      snapshot: true
-      snipshot: false
+    LogarithmicMassWeightedDiffuseNitrogenOverOxygenOfGasHighLimit: false
+    LogarithmicMassWeightedDiffuseCarbonOverOxygenOfGasHighLimit: false
+    LogarithmicMassWeightedDiffuseCarbonOverOxygenOfGasLowLimit: false
+    LogarithmicMassWeightedDiffuseNitrogenOverOxygenOfGasLowLimit: false
     LuminosityWeightedMeanStellarAge: true
     MassWeightedMeanStellarAge: true
     MolecularHydrogenMass: true
