### 2025-07-17 311a798
Add ratio property
