+++ SOAP/property_table.py
@@ -1156,6 +1156,22 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=0,
         ),
+        "DustSmallToLargeGrainRatioSFRWeighted": Property(
+            name="DustSmallToLargeGrainRatioSFRWeighted",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_mass",
+            description="The ratio of the dust mass in small grains to the mass in large grains, weighted by the SFR of the particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType0/Masses",
+                "PartType0/DustMassFractions",
+                "PartType0/StarFormationRates",
+            ],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
         "KineticEnergyGas": Property(
             name="KineticEnergyGas",
             shape=1,
