### 2025-07-18 c4dfbc3
Store masses instead
