+++ parameter_files/COLIBRE_HYBRID.yml
@@ -87,6 +87,9 @@ ApertureProperties:
     DustLargeGrainMassInColdDenseGas:
       snapshot: true
       snipshot: false
+    DustLargeGrainMassSFRWeighted:
+      snapshot: true
+      snipshot: false
     DustLargeGrainMassInMolecularGas:
       snapshot: true
       snipshot: false
@@ -111,7 +114,7 @@ ApertureProperties:
     DustSmallGrainMassInMolecularGas:
       snapshot: true
       snipshot: false
-    DustSmallToLargeGrainRatioSFRWeighted:
+    DustSmallGrainMassSFRWeighted:
       snapshot: true
       snipshot: false
     GasMass: true
