### 2025-07-18 7e0e2cf
Skip workflow for drafts
