+++ .github/workflows/main.yml
@@ -9,8 +9,6 @@ on:
     branches: [ master ]
   pull_request:
     branches: [ master ]
-    # Don't run for draft pull requests
-    types: [opened, synchronize, reopened, ready_for_review]
 
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
@@ -21,6 +19,9 @@ jobs:
   formatting:
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
+
+    # Don't run for draft PRs
+    if: github.event.pull_request.draft == false
 
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps:
