### 2025-07-22 #8252 @robjmcgibbon
Apologies for the missing ORCIDs @ivastar, I've added them to the paper (as can be seen in the pdf generated just above). There's a new DOI at https://zenodo.org/records/16321553, which has addressed the bullet points you listed above.
