+++ documentation/joss_paper/paper.bib
@@ -269,7 +269,7 @@ archivePrefix = {arXiv},
 }
 
 @misc{caesar,
- author = {Thompson, Robert},
+ author = {{Narayanan et al.}},
  title = {CAESAR},
  year = {2023},
  publisher = {GitHub},
