### 2025-07-23 #8252 @robjmcgibbon
> Yes, let's do that, maybe Narayanan et al?

Sure, done in https://github.com/SWIFTSIM/SOAP/commit/694c3a3ebbfedd83e306ce299f88e3cb604823aa
