### 2025-07-23 #8252 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
