### 2025-07-23 #8252 @ivastar
Sincere thanks to the reviewers @dnelson86, @stuartmcalpine, @nvanalfen for their detailed and expeditious work on this submission! Your work makes JOSS possible.

@robjmcgibbon, the Editor in Chief will take over from here for pre-publication checks.
