### 2025-07-23 #8252 @robjmcgibbon
> Sincere thanks to the reviewers @dnelson86, @stuartmcalpine, @nvanalfen for their detailed and expeditious work on this submission! Your work makes JOSS possible.

Yes, thank you to the reviewers for the clear and helpful feedback, it’s much appreciated! And many thanks to @ivastar for leading the review process!
