### 2025-07-25 81009a6
Add new options for fof radii script
