+++ misc/calculate_fof_radii.py
@@ -1,12 +1,12 @@
 #!/bin/env python
 
 """
-calculate_fof_positions.py
+calculate_fof_radii.py
 
-This script calculates the maximum and minimum particles positions for each FOF.
+This script calculates the maximum particle radii for each FOF.
 Usage:
 
-  mpirun -- python misc/calculate_fof_positions.py \
+  mpirun -- python misc/calculate_fof_radii.py \
           --snap-basename=SNAPSHOT \
           --fof-basename=FOF \
           --output-basename=OUTPUT
@@ -14,6 +14,9 @@ Usage:
 where SNAPSHOT is the basename of the snapshot files (the snapshot
 name without the .{file_nr}.hdf5 suffix), FOF is the basename of the
 fof catalogues, and OUTPUT is the basename of the output fof catalogues.
+
+For a full list of optional arguments run:
+  python misc/calculate_fof_radii.py -h
 """
 
 import argparse
@@ -57,21 +60,19 @@ parser.add_argument(
     "--output-basename",
     type=str,
     required=True,
-    help=("The basename for the output files"),
+    help="The basename for the output files",
 )
 parser.add_argument(
     "--null-fof-id",
     type=int,
     required=False,
     default=2147483647,
-    help=("The FOFGroupIDs of particles not in a FOF group"),
+    help="The FOFGroupIDs of particles not in a FOF group",
 )
 parser.add_argument(
     "--copy-datasets",
-    type=bool,
-    required=False,
-    default=False,
-    help=("Copy datasets from input FoF files (otherwise a link is created)"),
+    action='store_true',
+    help="Copy datasets from input FoF files (otherwise a link is created)",
 )
 parser.add_argument(
     "--n-test",
@@ -80,33 +81,66 @@ parser.add_argument(
     default=0,
     help="The number of FOFs to check. If -1 all objects will be checked",
 )
+parser.add_argument(
+    "--single-fof-file",
+    action='store_true',
+    help="If there is a single FoF file even though there are distributed snapshots",
+)
+parser.add_argument(
+    "--recalculate-sizes",
+    action='store_true',
+    help="Whether to calculate the number of particles in each FOF group",
+)
+parser.add_argument(
+    "--reset-ids",
+    action='store_true',
+    help=(
+        "Whether to set the group ids in the FOF catalgoues to 1...N ",
+        "(This is required for some runs as there was a bug where the incorrect",
+        "IDs were output)",
+    ),
+)
 
 args = parser.parse_args()
 snap_filename = args.snap_basename + ".{file_nr}.hdf5"
-fof_filename = args.fof_basename + ".{file_nr}.hdf5"
-output_filename = args.output_basename + ".{file_nr}.hdf5"
+if args.single_fof_file:
+    fof_filename = args.fof_basename + ".hdf5"
+    output_filename = args.output_basename + ".hdf5"
+else:
+    fof_filename = args.fof_basename + ".{file_nr}.hdf5"
+    output_filename = args.output_basename + ".{file_nr}.hdf5"
 os.makedirs(os.path.dirname(output_filename), exist_ok=True)
 
 if comm_rank == 0:
     start_time = time.time()
-    print(f"Running with {comm_size} ranks")
-    print(f"snap_filename: {snap_filename}")
-    print(f"fof_filename: {fof_filename}")
-    print(f"output_filename: {output_filename}")
-    print(f"null_fof_id: {args.null_fof_id}")
[…]
