### 2025-07-28 19511e3
Formatted and tests passing
