+++ SOAP/particle_selection/subhalo_properties.py
@@ -38,7 +38,9 @@ from SOAP.property_calculation.inertia_tensors import (
     get_inertia_tensor_mass_weighted,
     get_inertia_tensor_luminosity_weighted,
 )
-from SOAP.property_calculation.cylindrical_coordinates import calculate_cylindrical_velocities
+from SOAP.property_calculation.cylindrical_coordinates import (
+    calculate_cylindrical_velocities,
+)
 from SOAP.particle_filter.recently_heated_gas_filter import RecentlyHeatedGasFilter
 from SOAP.property_calculation.stellar_age_calculator import StellarAgeCalculator
 from SOAP.property_table import PropertyTable
@@ -1388,37 +1390,42 @@ class SubhaloParticleData:
     def star_cylindrical_velocities(self) -> unyt.unyt_array:
         """
         Calculate the velocities of the star particles in cyclindrical
-        coordinates, where the axes are centred on the halo centre,
+        coordinates, where the axes are centred on the stellar CoM,
         and the z axis is aligned with the stellar angular momentum.
         """
 
-        if self.Mstar == 0:
+        if self.Nstar < 2:
             return None
+
+        # Calculate the position of the stars relative to their CoM
+        pos = self.pos_star - (self.star_mass_fraction[:, None] * self.pos_star).sum(
+            axis=0
+        )
 
         # Calculate relative velocity of stars
         vrel = self.vel_star - self.vcom[None, :]
 
         # Get velocities in cylindrical coordinates
         return calculate_cylindrical_velocities(
-            self.pos_star,
+            pos,
             vrel,
             self.Lstar,
         )
 
     @lazy_property
-    def RotationalVelocityStars(self) -> unyt.unyt_array:
-        if self.Mstar == 0:
+    def StellarRotationalVelocity(self) -> unyt.unyt_array:
+        if self.Nstar < 2:
             return None
         v_cylindrical = self.star_cylindrical_velocities
         v_phi = v_cylindrical[:, 1]
         return (self.star_mass_fraction * v_phi).sum()
 
     @lazy_property
-    def CylindricalVelocityDispersionStars(self) -> unyt.unyt_array:
-        if self.Mstar == 0:
+    def StellarCylindricalVelocityDispersion(self) -> unyt.unyt_array:
+        if self.Nstar < 2:
             return None
         v_cylindrical = self.star_cylindrical_velocities
-        vel_disp = 0
+        vel_disp = 0 * v_cylindrical.units**2
         for i in range(3):
             mean = (self.star_mass_fraction * v_cylindrical[:, i]).sum()
             delta = self.star_mass_fraction * ((v_cylindrical[:, i] - mean) ** 2)
@@ -1627,7 +1634,10 @@ class SubhaloParticleData:
         if self.Mstar == 0:
             return None
         return get_inertia_tensor_mass_weighted(
-            self.mass_star, self.pos_star, 10 * self.HalfMassRadiusStar, max_iterations=1
+            self.mass_star,
+            self.pos_star,
+            10 * self.HalfMassRadiusStar,
+            max_iterations=1,
         )
 
     @lazy_property
@@ -2267,8 +2277,8 @@ class SubhaloProperties(HaloProperty):
             "Lgas",
             "Ldm",
             "Lstar",
-            "CylindricalVelocityDispersionStars",
-            "RotationalVelocityStars",
+            "StellarCylindricalVelocityDispersion",
+            "StellarRotationalVelocity",
             "kappa_corot_gas",
             "kappa_corot_star",
             "Lbaryons",
