### 2025-07-28 9ef71e0
Calculate rotational velocity (#174)
* Update tensors

* Skip workflow for drafts

* Attempt 2

* Cylindrical coords

* Formatted and tests passing
