+++ SOAP/property_table.py
@@ -1608,6 +1608,38 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=1,
         ),
+        "StellarRotationalVelocity": Property(
+            name="StellarRotationalVelocity",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_length/snap_time",
+            description="Mass weighted mean rotational velocity of the stars, in a cylindrical coordinate system where the axes are centred on the stellar CoM, and the z axis is aligned with the stellar angular momentum.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType4/Coordinates",
+                "PartType4/Masses",
+                "PartType4/Velocities",
+            ],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
+        "StellarCylindricalVelocityDispersion": Property(
+            name="StellarCylindricalVelocityDispersion",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_length/snap_time",
+            description="Velocity dispersion of the stars computed in a cylindrical coordinate system where the axes are centred on the stellar CoM, and the z axis is aligned with the stellar angular momentum.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType4/Coordinates",
+                "PartType4/Masses",
+                "PartType4/Velocities",
+            ],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
         "MaximumStellarBirthDensity": Property(
             name="MaximumStellarBirthDensity",
             shape=1,
