+++ SOAP/particle_selection/subhalo_properties.py
@@ -38,6 +38,9 @@ from SOAP.property_calculation.inertia_tensors import (
     get_inertia_tensor_mass_weighted,
     get_inertia_tensor_luminosity_weighted,
 )
+from SOAP.property_calculation.cylindrical_coordinates import (
+    calculate_cylindrical_velocities,
+)
 from SOAP.particle_filter.recently_heated_gas_filter import RecentlyHeatedGasFilter
 from SOAP.property_calculation.stellar_age_calculator import StellarAgeCalculator
 from SOAP.property_table import PropertyTable
@@ -1054,12 +1057,12 @@ class SubhaloParticleData:
         """
         Inertia tensor of the total mass distribution.
         Computed iteratively using an ellipsoid with volume equal to that of
-        a sphere with radius HalfMassRadiusTot. Only considers bound particles.
+        a sphere with radius 10 x HalfMassRadiusTot. Only considers bound particles.
         """
         if self.Mtot == 0:
             return None
         return get_inertia_tensor_mass_weighted(
-            self.mass, self.position, self.HalfMassRadiusTot
+            self.mass, self.position, 10 * self.HalfMassRadiusTot
         )
 
     @lazy_property
@@ -1067,38 +1070,38 @@ class SubhaloParticleData:
         """
         Reduced inertia tensor of the total mass distribution.
         Computed iteratively using an ellipsoid with volume equal to that of
-        a sphere with radius HalfMassRadiusTot. Only considers bound particles.
+        a sphere with radius 10 x HalfMassRadiusTot. Only considers bound particles.
         """
         if self.Mtot == 0:
             return None
         return get_inertia_tensor_mass_weighted(
-            self.mass, self.position, self.HalfMassRadiusTot, reduced=True
+            self.mass, self.position, 10 * self.HalfMassRadiusTot, reduced=True
         )
 
     @lazy_property
     def TotalInertiaTensorNoniterative(self) -> unyt.unyt_array:
         """
         Inertia tensor of the total mass distribution.
-        Computed using all bound particles within HalfMassRadiusTot.
+        Computed using all bound particles within 10 * HalfMassRadiusTot.
         """
         if self.Mtot == 0:
             return None
         return get_inertia_tensor_mass_weighted(
-            self.mass, self.position, self.HalfMassRadiusTot, max_iterations=1
+            self.mass, self.position, 10 * self.HalfMassRadiusTot, max_iterations=1
         )
 
     @lazy_property
     def TotalInertiaTensorReducedNoniterative(self) -> unyt.unyt_array:
         """
         Reduced inertia tensor of the total mass distribution.
-        Computed using all bound particles within HalfMassRadiusTot.
+        Computed using all bound particles within 10 * HalfMassRadiusTot.
         """
         if self.Mtot == 0:
             return None
         return get_inertia_tensor_mass_weighted(
             self.mass,
             self.position,
-            self.HalfMassRadiusTot,
+            10 * self.HalfMassRadiusTot,
             reduced=True,
             max_iterations=1,
         )
@@ -1185,12 +1188,12 @@ class SubhaloParticleData:
         """
         Inertia tensor of the gas mass distribution.
         Computed iteratively using an ellipsoid with volume equal to that of
-        a sphere with radius HalfMassRadiusGas. Only considers bound particles.
+        a sphere with radius 10 x HalfMassRadiusGas. Only considers bound particles.
         """
         if self.Mgas == 0:
             return None
         return get_inertia_tensor_mass_weighted(
-            self.mass_gas, self.pos_gas, self.HalfMassRadiusGas
+            self.mass_gas, self.pos_gas, 10 * self.HalfMassRadiusGas
         )
 
[…]
