### 2025-07-28 9b1d886
Cold dense gas documentation (#173)
* Update cold dense gas documentation

* Add ratio property

* Units

* Store masses instead
