+++ SOAP/property_table.py
@@ -283,6 +283,14 @@ class PropertyTable:
             "SOAP/DescendantIndex",
             "SOAP/ProgenitorIndex",
         ],
+        "footnote_cold_dense.py": [
+            "DustGraphiteMassInColdDenseGas",
+            "DustLargeGrainMassInColdDenseGas",
+            "DustSilicatesMassInColdDenseGas",
+            "DustSmallGrainMassInColdDenseGas",
+            "GasMassInColdDenseGas",
+            "GasMassInColdDenseDiffuseMetals",
+        ],
     }
 
     # dictionary with human-friendly descriptions of the various lossy
@@ -1035,6 +1043,22 @@ class PropertyTable:
             output_physical=True,
             a_scale_exponent=0,
         ),
+        "DustLargeGrainMassSFRWeighted": Property(
+            name="DustLargeGrainMassSFRWeighted",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_mass",
+            description="The dust mass in large grains, weighted by the SFR of the particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType0/Masses",
+                "PartType0/DustMassFractions",
+                "PartType0/StarFormationRates",
+            ],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
         "DustSilicatesMass": Property(
             name="DustSilicatesMass",
             shape=1,
@@ -1144,6 +1168,22 @@ class PropertyTable:
                 "PartType0/DustMassFractions",
                 "PartType0/Densities",
                 "PartType0/Temperatures",
+            ],
+            output_physical=True,
+            a_scale_exponent=0,
+        ),
+        "DustSmallGrainMassSFRWeighted": Property(
+            name="DustSmallGrainMassSFRWeighted",
+            shape=1,
+            dtype=np.float32,
+            unit="snap_mass",
+            description="The dust mass in small grains, weighted by the SFR of the particles.",
+            lossy_compression_filter="FMantissa9",
+            dmo_property=False,
+            particle_properties=[
+                "PartType0/Masses",
+                "PartType0/DustMassFractions",
+                "PartType0/StarFormationRates",
             ],
             output_physical=True,
             a_scale_exponent=0,
@@ -5004,6 +5044,19 @@ Name & Shape & Type & Units & SH & ES & IS & EP & SO & Category & Compression\\\
                 with open(f"documentation/{fnote}", "r") as ifile:
                     fnstr = ifile.read()
                 fnstr = fnstr.replace("$FOOTNOTE_NUMBER$", f"{i+1}")
+                # Substitute in values from parameter file
+                if "$LOG_COLD_GAS_TEMP$" in fnstr:
+                    params = self.parameters.get_cold_dense_params()
+                    assert params["initialised"]
+                    T = f'{np.log10(params["maximum_temperature_K"]):.2g}'
+                    fnstr = fnstr.replace("$LOG_COLD_GAS_TEMP$", T)
+                if "$LOG_COLD_GAS_DENSITY$" in fnstr:
+                    params = self.parameters.get_cold_dense_params()
+                    assert params["initialised"]
+                    rho = (
+                        f'{np.log10(params["minimum_hydrogen_number_density_cm3"]):.2g}'
+                    )
+                    fnstr = fnstr.replace("$LOG_COLD_GAS_DENSITY$", rho)
                 ofile.write(f"{fnstr}\n\n")
 
         # Particle limits for each filter
@@ -5079,9 +5132,11 @@ Group name (HDF5) & Group name (swiftsimio) & Inclusive? & Filter \\\\
                     aperture_name = f"{int(multiplier)}x{prop}"
[…]
