### 2025-07-28 0f770f4
Add note on new BoundSubhalo tensor radius
