+++ documentation/footnote_tensor.tex
@@ -15,3 +15,5 @@ The reduced inertia tensor is calculated as
 where $r_k$ is the radial distance of the particle.
 
 We do not calculate the inertia tensor if there are less than 20 particles within the initial sphere.
+
+For when calculating the inertia tensor for a bound subhalo we use a sphere with a radius equal to 10 times the half mass radius of the particles being considered.
