### 2025-07-28 c392287
Add footenote on averaged SFR
