### 2025-07-28 800dd7d
Add footnote for averaged SFR (#177)
* Footnotes are tex files..

* Add footenote on averaged SFR
