### 2025-07-28 d7bc22d
Initial editorial tweaks for JOSS publication
