### 2025-07-29 1ab5fe9
Initial editorial tweaks for JOSS publication (#178)
