+++ documentation/joss_paper/paper.bib
@@ -173,6 +173,47 @@ author = {Lisandro Dalcín and Rodrigo Paz and Mario Storti},
 keywords = {Message passing, MPI, High level languages, Parallel Python}
 }
 
+@article{mpi4py_2,
+title = {MPI for Python: Performance improvements and MPI-2 extensions},
+journal = {Journal of Parallel and Distributed Computing},
+volume = {68},
+number = {5},
+pages = {655-662},
+year = {2008},
+issn = {0743-7315},
+doi = {https://doi.org/10.1016/j.jpdc.2007.09.005},
+url = {https://www.sciencedirect.com/science/article/pii/S0743731507001712},
+author = {Lisandro Dalcín and Rodrigo Paz and Mario Storti and Jorge D’Elía},
+keywords = {Message passing, MPI, High-level languages, Parallel Python},
+}
+
+@article{mpi4py_3,
+title = {Parallel distributed computing using Python},
+journal = {Advances in Water Resources},
+volume = {34},
+number = {9},
+pages = {1124-1139},
+year = {2011},
+note = {New Computational Methods and Software Tools},
+issn = {0309-1708},
+doi = {https://doi.org/10.1016/j.advwatres.2011.04.013},
+url = {https://www.sciencedirect.com/science/article/pii/S0309170811000777},
+author = {Lisandro Dalcín and Rodrigo Paz and Pablo Kler and Alejandro Cosimo},
+keywords = {Python, MPI, PETSc},
+}
+
+@ARTICLE{mpi4py_4,
+  author={Dalcín, Lisandro and Fang, Yao-Lung L.},
+  journal={Computing in Science & Engineering},
+  title={mpi4py: Status Update After 12 Years of Development},
+  year={2021},
+  volume={23},
+  number={4},
+  pages={47-54},
+  keywords={Python;Graphics processing units;High performance computing;Semantics;Programming;Task analysis},
+  doi={10.1109/MCSE.2021.3083216}}
+
+
 @ARTICLE{gadget,
        author = {{Springel}, Volker},
         title = "{The cosmological simulation code GADGET-2}",
