### 2025-07-29 f947b0d
Fixes for additional mpi4py references in JOSS paper (#179)
