### 2025-07-29 #8252 @warrickball
Thanks once again to @dnelson86, @stuartmcalpine & @nvanalfen for reviewing and @ivastar for editing this submission! JOSS simply wouldn't be possible without its community of volunteers.
